% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_prevalence.R
\name{act_prevalance}
\alias{act_prevalance}
\title{Activations prevalence.}
\usage{
act_prevalance(act_list)
}
\arguments{
\item{act_list}{List of activations from \code{\link{id_activations}}. All
should have the same sessions, fields, and brainstructures.}
}
\value{
A list containing the prevalances of activation, as a proportion of
the results from \code{act_list}.
}
\description{
Activations prevalence.
}
