% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{aic_Param}
\alias{aic_Param}
\title{aic}
\arguments{
\item{aic}{(For prewhitening) Use the Akaike information criterion (AIC) to
select AR model orders between \code{0} and \code{ar_order}? Default:
\code{FALSE}.}
}
\description{
aic
}
