% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAIC.R
\name{AICc}
\alias{AICc}
\title{Corrected AIC}
\usage{
AICc(y, demean = FALSE, order.max = 10)
}
\arguments{
\item{y}{The autocorrelated data}

\item{demean}{Demean \code{y}? Default: \code{FALSE}.}

\item{order.max}{The model order limit. Default: \code{10}.}
}
\value{
The cAIC
}
\description{
Computes corrected AIC (AICc).
}
\keyword{internal}
