% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{ar_smooth_Param}
\alias{ar_smooth_Param}
\title{ar_smooth}
\arguments{
\item{ar_smooth}{(For prewhitening) The FWHM parameter for spatially
smoothing the coefficient estimates for the AR model to use for
prewhitening. Recall that
\eqn{\sigma = \frac{FWHM}{2*sqrt(2*log(2)}}. Set to \code{0} to not smooth
the estimates. Default: \code{5}.}
}
\description{
ar_smooth
}
