% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{areColors}
\alias{areColors}
\title{Check whether each element of vector x is a valid color representation}
\usage{
areColors(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
A logical vector indicating which of the elements of x are valid color representations
}
\description{
Check whether each element of vector x is a valid color representation
}
