% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R, R/util_RED.R
\name{asplit}
\alias{asplit}
\title{Array split}
\usage{
asplit(X, margin)

asplit(X, margin)
}
\arguments{
\item{X}{An array of dimension greater than 1}

\item{margin}{The margin at which the array should be split}
}
\value{
A list of subarrays split along the \code{margin} dimension

A list of subarrays split along the \code{margin} dimension
}
\description{
Generalizes the \code{split} function to work with arrays of arbitrary dimension greater than 2.

Generalizes the \code{split} function to work with arrays of arbitrary dimension greater than 2.
}
