% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{avg_betas_over_sessions_Param}
\alias{avg_betas_over_sessions_Param}
\title{avg_betas_over_sessions}
\arguments{
\item{avg_betas_over_sessions}{Average estimates for betas over multiple
sessions? Default: \code{FALSE}.}
}
\description{
avg_betas_over_sessions
}
