% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{avg_sessions_Param}
\alias{avg_sessions_Param}
\title{avg_sessions}
\arguments{
\item{avg_sessions}{Average task estimates (betas) over multiple sessions?
Default: \code{TRUE}.}
}
\description{
avg_sessions
}
