% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{Bayes_Param}
\alias{Bayes_Param}
\title{Bayes}
\arguments{
\item{Bayes}{If TRUE, will fit a spatial Bayesian GLM in addition to the
classical GLM. Classical GLM results are always returned.}
}
\description{
Bayes
}
