% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM.R
\name{BayesGLM}
\alias{BayesGLM}
\title{}
\usage{
BayesGLM(
  data,
  beta_names = NULL,
  vertices = NULL,
  faces = NULL,
  mesh = NULL,
  mask = NULL,
  scale_BOLD = TRUE,
  scale_design = TRUE,
  num.threads = 4,
  return_INLA_result = TRUE,
  outfile = NULL,
  verbose = FALSE,
  contrasts = NULL,
  avg_sessions = TRUE,
  trim_INLA = TRUE
)
}
\arguments{
\item{data}{}

\item{beta_names}{}

\item{vertices}{}

\item{faces}{}

\item{mesh}{}

\item{mask}{}

\item{scale_BOLD}{}

\item{scale_design}{}

\item{num.threads}{}

\item{return_INLA_result}{}

\item{outfile}{}

\item{verbose}{}

\item{contrasts}{}

\item{avg_sessions}{}

\item{trim_INLA}{\verb{}\verb{}}
}
\value{

}
\description{

}
\section{}{
NA
}

