% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM.R
\name{BayesGLM_2d}
\alias{BayesGLM_2d}
\title{Spatial Bayesian GLM for fMRI task activation on slice volumetric data}
\usage{
BayesGLM_2d(
  BOLD,
  design = NULL,
  onsets = NULL,
  TR = NULL,
  nuisance = NULL,
  nuisance_include = c("drift", "dHRF"),
  binary_mask = NULL,
  scale_BOLD = TRUE,
  scale_design = TRUE,
  num.threads = 4,
  GLM_method = "both",
  session_names = NULL,
  return_INLA_result = TRUE,
  write_dir = NULL,
  outfile = NULL,
  verbose = FALSE,
  contrasts = NULL
)
}
\arguments{
\item{BOLD}{A list of sessions, each with a three-dimensional array in which
the first two dimensions correspond to the size of the fMRI slice in space
and the last dimension corresponds to time}

\item{design}{A TxK task design matrix (or list of such matrices, for
multiple-session modeling) with column names representing tasks. Each
column represents the expected BOLD response due to each task, a
convolution of the hemodynamic response function (HRF) and the task
stimulus.  Must be provided if and only if onsets=NULL.  Note that the
scale of the regressors will affect the scale and interpretation of the
beta coefficients, so imposing a proper scale (e.g., set maximum to 1) is
recommended.}

\item{onsets}{A matrix of onsets (first column) and durations (second column)
for each task in SECONDS, organized as a list where each element of the
list corresponds to one task. Names of list should be task names. (Or for
multi-session modeling, a list of such lists.)  Must be provided if and
only if design=NULL.}

\item{TR}{The temporal resolution of the data in seconds.  Must be provided
if onsets provided.}

\item{nuisance}{(Optional) A TxJ matrix of nuisance signals (or list of such
matrices, for multiple-session modeling).}

\item{nuisance_include}{(Optional) Additional nuisance covariates to include.
Default is 'drift' (linear and quadratic drift terms) and 'dHRF' (temporal
derivative of each column of design matrix).}

\item{binary_mask}{(optional) a binary brain slice image used to mask
the BOLD data and make a more efficient network mesh for the
neighborhood definitions}

\item{scale_BOLD}{If TRUE (default), scale timeseries data so estimates
represent percent signal change.  Else, center but do not scale.}

\item{scale_design}{If TRUE (default), scale design matrix so maximum value
is equal to 1.  Else, do not scale.}

\item{num.threads}{Maximum number of threads the inla-program will use for
model estimation}

\item{GLM_method}{Either 'Bayesian' for spatial Bayesian GLM only,
'classical' for the classical GLM only, or 'both' to return both classical
and Bayesian estimates of task activation.}

\item{session_names}{(Optional) A vector of names corresponding to each
session.}

\item{return_INLA_result}{If TRUE, object returned will include the INLA
model object (can be large).  Default is FALSE.  Required for running
\code{id_activations} on \code{BayesGLM} model object (but not for running
\code{BayesGLM_joint} to get posterior quantities of group means or
contrasts).}

\item{write_dir}{(Optional) The path to a directory where results will be
saved. Defaults to the current working directory.}

\item{outfile}{(Optional) File name (without extension) of output file for
BayesGLM result to use in Bayesian group modeling.}

\item{verbose}{Logical indicating if INLA should run in a verbose mode
(default FALSE).}

\item{contrasts}{A list of contrast vectors to be passed to
\code{\link{inla}}.}
}
\value{
An object of class BayesGLM, a list containing ...
}
\description{
Spatial Bayesian GLM for fMRI task activation on slice volumetric data
}
