% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{BayesGLM_check_design}
\alias{BayesGLM_check_design}
\title{Check design matrix or matrices}
\usage{
BayesGLM_check_design(design, design_type, nS, nT)
}
\arguments{
\item{design, design_type, nS, nT}{See \code{"BayesGLM"}}
}
\value{
The design matrix as an \code{nS}-length list.
}
\description{
Check design matrix or matrices
}
\keyword{internal}
