% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{BayesGLM_check_nuisance}
\alias{BayesGLM_check_nuisance}
\title{Check nuisance matrix or matrices}
\usage{
BayesGLM_check_nuisance(design, design_type, nS, nT)
}
\arguments{
\item{nuisance, nS, nT}{See \code{"BayesGLM"}}
}
\value{
The nuisance matrix as an \code{nS}-length list.
}
\description{
Check nuisance matrix or matrices
}
\keyword{internal}
