% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_cifti_format_results.R
\name{BayesGLM_cifti_format_results}
\alias{BayesGLM_cifti_format_results}
\title{Format BayesGLM results into \code{"xifti"} objects}
\usage{
BayesGLM_cifti_format_results(
  BayesGLM_results,
  do,
  spatial,
  submeta,
  session_names,
  field_names,
  method = c("classical", "Bayesian")
)
}
\arguments{
\item{BayesGLM_results}{The list of BayesGLM results}

\item{do, spatial, submeta, session_names, field_names}{See \code{BayesGLM_cifti}.}

\item{method}{\code{"classical"} or \code{"Bayesian"}}
}
\value{
The list of \code{"xifti"} objects
}
\description{
Format beta estimates of the BayesGLM results into list of \code{"xifti"}
objects for each session.
}
\keyword{internal}
