% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{BayesGLM_cifti_make_DCT}
\alias{BayesGLM_cifti_make_DCT}
\title{Make DCT bases}
\usage{
BayesGLM_cifti_make_DCT(DCT, hpf, nT, TR, verbose)
}
\arguments{
\item{DCT, hpf, nT, TR, verbose}{See \code{BayesGLM_cifti}}
}
\value{
The matrix of DCT bases, or \code{NULL} if none
}
\description{
Make DCT bases
}
\keyword{internal}
