% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_format_cifti.R
\name{BayesGLM_format_cifti}
\alias{BayesGLM_format_cifti}
\title{Format BayesGLM_fun results into \code{"xifti"} objects}
\usage{
BayesGLM_format_cifti(
  BGLMs,
  do,
  spatial,
  submeta,
  session_names,
  field_names,
  method = c("classical", "Bayesian")
)
}
\arguments{
\item{BGLMs}{The list of \code{"BGLM"} objects, from \code{BayesGLM_fun}.}

\item{do, spatial, submeta, session_names, field_names}{See \code{BayesGLM}.}

\item{method}{\code{"classical"} or \code{"Bayesian"}}
}
\value{
The list of \code{"xifti"} objects
}
\description{
Format beta estimates of the BayesGLM_fun results into list of \code{"xifti"}
objects for each session.
}
\keyword{internal}
