% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_check_data_inputs.R
\name{BayesGLM_format_scrub}
\alias{BayesGLM_format_scrub}
\title{Format scrub}
\usage{
BayesGLM_format_scrub(scrub, nS_expect = NULL, nT_expect = NULL)
}
\arguments{
\item{scrub}{The \code{scrub} argument input. Will be formatted to a
\code{nS}-length list.}

\item{nS_expect}{The expected number of sessions, if known.}

\item{nT_expect}{The expected number of timepoints, if known. For
multi-session data this is a session-length vector.}
}
\value{
\code{scrub}
}
\description{
Format scrub for \code{BayesGLM}, \code{fit_bayesglm},
\code{multiGLM}, and \code{multiGLM_fun}.
}
\keyword{internal}
