% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_check_data_inputs.R
\name{BayesGLM_is_valid_one_design}
\alias{BayesGLM_is_valid_one_design}
\title{Is a valid design?}
\usage{
BayesGLM_is_valid_one_design(design)
}
\arguments{
\item{design}{The design matrix/array}
}
\description{
Is a valid design?
}
\keyword{internal}
