% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_check_data_inputs.R
\name{BayesGLM_is_valid_one_nuisance}
\alias{BayesGLM_is_valid_one_nuisance}
\title{Is a valid nuisance?}
\usage{
BayesGLM_is_valid_one_nuisance(nuisance)
}
\arguments{
\item{nuisance}{The nuisance matrix}
}
\description{
Is a valid nuisance?
}
\keyword{internal}
