% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_check_data_inputs.R
\name{BayesGLM_is_valid_one_scrub}
\alias{BayesGLM_is_valid_one_scrub}
\title{Is a valid scrub?}
\usage{
BayesGLM_is_valid_one_scrub(scrub)
}
\arguments{
\item{scrub}{The scrub matrix}
}
\description{
Is a valid scrub?
}
\keyword{internal}
