% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_group.R
\name{BayesGLM_joint.one_model}
\alias{BayesGLM_joint.one_model}
\title{Computes Bayesian GLM posterior means, quantiles and excursions sets (areas of activation) for summaries across sessions based on a single multi-session model}
\usage{
BayesGLM_joint.one_model(
  result,
  contrasts = NULL,
  quantiles = NULL,
  excursion_type,
  gamma = 0,
  alpha = 0.05,
  no_cores = NULL,
  nsamp_theta = 50,
  nsamp_beta = 100
)
}
\arguments{
\item{result}{An object of class BayesGLM containing the results of a multi-session analysis. Must contain INLA_result!  (See \code{help(BayesGLM)}.)}

\item{contrasts}{A list of vectors, each length M*K, specifying the contrast(s) of interest across sessions, where M is the number of sessions analyzed and K is the number of tasks.  See Details for more information.}

\item{quantiles}{Vector of posterior quantiles to return in addition to the posterior mean}

\item{excursion_type}{The type of excursion function for the contrast (">", "<", "!="), or a vector thereof (each element corresponding to one contrast).  If none provided, no inference performed.}

\item{gamma}{Activation threshold for the excursion set, or a vector thereof (each element corresponding to one contrast).}

\item{alpha}{Significance level for activation for the excursion set}

\item{no_cores}{The number of cores to use for sampling in parallel. If NULL, do not run in parallel.}

\item{nsamp_theta}{Number of theta values to sample from posterior}

\item{nsamp_beta}{Number of beta vectors to sample conditional on each theta value sampled}
}
\value{
A list containing...
}
\description{
Computes Bayesian GLM posterior means, quantiles and excursions sets (areas of activation) for summaries across sessions based on a single multi-session model
}
\details{
Performs posterior inference on summaries or contrasts across the sessions in a multi-session Bayesian GLM.  Let M be the number of sessions analyzed in the
multi-session Bayesian GLM, and let K be the number of tasks.  Computes the posterior mean of each contrast and produces excursions sets representing areas of
activation (above a specified activation threshold gamma) for each contrast map based on the joint (across locations) posterior distribution.

Each contrast vector specifies an across-session summary of interest.  For example, the contrast vector \code{rep(c(1/M, rep(0, M-1)), K)} represents the across-session average
for the first task; \code{c( rep(c(1/M1, rep(0, M1-1)), K), rep(c(-1/M2, rep(0, M2-1)), K))} represents the difference between the first M1 sessions and the remaining M2 sessions (M1+M2=M) for the first task;
\code{rep( c(1/M,-1/M,rep(0, K-2)) ,M)} represents the difference between the first two tasks, averaged over all sessions.
}
\note{
This function requires the \code{INLA} package, which is not a CRAN package. See \url{http://www.r-inla.org/download} for easy installation instructions.
}
