% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{BayesGLM_make_DCT}
\alias{BayesGLM_make_DCT}
\title{Make DCT bases}
\usage{
BayesGLM_make_DCT(hpf, nT, TR)
}
\arguments{
\item{hpf, nT, TR}{See \code{BayesGLM}}
}
\value{
The matrix of DCT bases, or \code{NULL} if none
}
\description{
Make DCT bases
}
\keyword{internal}
