% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_session_names.R
\name{BayesGLM_session_names}
\alias{BayesGLM_session_names}
\title{Get \code{session_names} for GLM}
\usage{
BayesGLM_session_names(nS, session_names, BOLD_names, design_names)
}
\description{
Get \code{session_names} for GLM
}
\keyword{internal}
