% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM.R
\name{BayesGLM_slice}
\alias{BayesGLM_slice}
\title{Spatial Bayesian GLM for fMRI task activation on 2d slice volumetric data}
\usage{
BayesGLM_slice(
  BOLD,
  design = NULL,
  onsets = NULL,
  TR = NULL,
  nuisance = NULL,
  nuisance_include = c("drift", "dHRF"),
  binary_mask = NULL,
  scale_BOLD = TRUE,
  scale_design = TRUE,
  num.threads = 4,
  GLM_method = "both",
  session_names = NULL,
  return_INLA_result = TRUE,
  write_dir = NULL,
  outfile = NULL,
  verbose = FALSE,
  contrasts = NULL
)
}
\arguments{
\item{BOLD}{A list of sessions, each with a three-dimensional array in which
the first two dimensions correspond to the size of the fMRI slice in space
and the last dimension corresponds to time}

\item{design, onsets, TR}{All or none must be provided.

\code{design} is a \eqn{T x K} task design matrix (or list of such
matrices, for multiple-session modeling) with column names representing
tasks. Each column represents the expected BOLD response due to each task,
a convolution of the hemodynamic response function (HRF) and the task
stimulus. Note that the scale of the regressors will affect the scale and
interpretation of the beta coefficients, so imposing a proper scale (e.g.,
set maximum to 1) is recommended.

\code{onsets} is a matrix of onsets (first column) and durations (second column)
for each task in seconds, organized as a list where each element of the
list corresponds to one task. Names of list should be task names. (Or for
multi-session modeling, a list of such lists.)

\code{TR} is the temporal resolution of the data in seconds.}

\item{nuisance}{(Optional) A TxJ matrix of nuisance signals (or list of such
matrices, for multiple-session modeling).}

\item{nuisance_include}{(Optional) Additional nuisance covariates to include.
Default is 'drift' (linear and quadratic drift terms) and 'dHRF' (temporal
derivative of each column of design matrix).}

\item{binary_mask}{(optional) a binary brain slice image used to mask
the BOLD data and make a more efficient network mesh for the
neighborhood definitions}

\item{scale_BOLD}{Scale timeseries data so estimates represent percent signal
change? Default: \code{TRUE}. If \code{FALSE}, the data will just be
centered.}

\item{scale_design}{Scale the design matrix by dividing each column by its
maximum and then subtracting the mean? Default: \code{TRUE}.}

\item{num.threads}{The maximum number of threads to use in the inla-program
for model estimation. Default: \code{4}.}

\item{GLM_method}{Either 'Bayesian' for spatial Bayesian GLM only,
'classical' for the classical GLM only, or 'both' to return both classical
and Bayesian estimates of task activation.}

\item{session_names}{(Optional) A vector of names corresponding to each
session.}

\item{return_INLA_result}{Return the INLA model object? (It can be large.)
Default: \code{TRUE}. Required for running \code{\link{id_activations}}
after, but not for running BayesGLM_joint after to get
posterior quantities of group means or contrasts.}

\item{write_dir}{(Optional) The path to a directory where results will be
saved. Defaults to the current working directory.}

\item{outfile}{(Optional) File name (without extension) of output file for
BayesGLM result to use in Bayesian group modeling.}

\item{verbose}{Should INLA be run in verbose mode? Default: \code{FALSE}.}

\item{contrasts}{List of contrast vectors to be passed to \code{\link{inla}}.}
}
\value{
An object of class BayesGLM, a list containing ...
}
\description{
Spatial Bayesian GLM for fMRI task activation on 2d slice volumetric data
}
