% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM.R
\name{BayesGLM_surface}
\alias{BayesGLM_surface}
\title{}
\usage{
BayesGLM_surface(
  data,
  vertices = NULL,
  faces = NULL,
  mesh = NULL,
  mask = NULL,
  scale_BOLD = TRUE,
  scale_design = TRUE,
  num.threads = 4,
  return_INLA_result = TRUE,
  outfile = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{}

\item{vertices}{}

\item{faces}{}

\item{mesh}{\verb{}\verb{}\verb{}}

\item{mask}{}

\item{scale_BOLD}{}

\item{num.threads}{}

\item{return_INLA_result}{}

\item{outfile}{}

\item{verbose}{}
}
\value{

}
\description{

}
\note{

}
