% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_vol3D.R, R/BayesGLM_vol3D_RED.R
\name{BayesGLM_vol3D}
\alias{BayesGLM_vol3D}
\title{Applies spatial Bayesian GLM to task fMRI data for 3D subcortical volumes}
\usage{
BayesGLM_vol3D(
  data,
  locations,
  labels,
  groups_df,
  scale = TRUE,
  return_INLA_result = FALSE,
  outfile = NULL,
  GLM = TRUE,
  num.threads = 6,
  verbose = FALSE
)

BayesGLM_vol3D(
  data,
  locations,
  labels,
  groups_df,
  scale = TRUE,
  return_INLA_result = FALSE,
  outfile = NULL,
  GLM = TRUE,
  num.threads = 6,
  verbose = FALSE
)
}
\arguments{
\item{data}{A list of sessions, where each session is a list with elements
BOLD, design and nuisance.  See \code{?create.session} and \code{?is.session} for more details.
List element names represent session names.}

\item{locations}{Vx3 matrix of x,y,z coordinates of each voxel}

\item{labels}{Vector of length V of region labels}

\item{groups_df}{Data frame indicating the name and model group of each region.  See Details.}

\item{scale}{If TRUE, scale timeseries data so estimates represent percent signal change.  If FALSE, just center the data and design to exclude the baseline field.}

\item{return_INLA_result}{If TRUE, object returned will include the INLA model object (can be large).  Default is TRUE. Required for running \code{id_activations} on \code{BayesGLM} model object.}

\item{outfile}{File name where results will be written (for use by \code{BayesGLM_grp}).}

\item{GLM}{If TRUE, classical GLM estimates will also be returned}

\item{num.threads}{Maximum number of threads the inla-program will use for model estimation}

\item{verbose}{Boolean indicating if the inla-program should run in a verbose mode (default FALSE).}
}
\value{
A list containing...

A list containing...
}
\description{
Applies spatial Bayesian GLM to task fMRI data for 3D subcortical volumes

Applies spatial Bayesian GLM to task fMRI data for 3D subcortical volumes
}
\details{
The subcortical data is separated into regions, whose sizes range from approximately 100 voxels to approximately 9000 voxels.  Smaller regions are grouped together to improve model fit.
The \code{groups_df} argument specifies which regions are grouped together.  This argument should be a data frame with R rows (the number of regions) and three columns: label, region, and group.
The label column is the numerical identifier of each region; the region column contains the region names, and the group column contains the model group assignments (e.g. 1,2,3). Regions to be excluded
from analysis are indicated by NA in the group assignment.

The subcortical data is separated into regions, whose sizes range from approximately 100 voxels to approximately 9000 voxels.  Smaller regions are grouped together to improve model fit.
The \code{groups_df} argument specifies which regions are grouped together.  This argument should be a data frame with R rows (the number of regions) and three columns: label, region, and group.
The label column is the numerical identifier of each region; the region column contains the region names, and the group column contains the model group assignments (e.g. 1,2,3). Regions to be excluded
from analysis are indicated by NA in the group assignment.
}
\note{
This function requires the \code{INLA} package, which is not a CRAN package. See \url{http://www.r-inla.org/download} for easy installation instructions.

This function requires the \code{INLA} package, which is not a CRAN package. See \url{http://www.r-inla.org/download} for easy installation instructions.
}
\examples{
\dontrun{}
\dontrun{}
}
