% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_functions.R
\name{BayesGLMEM}
\alias{BayesGLMEM}
\title{Expectation Maximization for the Bayesian General Linear Model}
\usage{
BayesGLMEM(
  data,
  vertices = NULL,
  faces = NULL,
  mesh = NULL,
  mask = NULL,
  scale_BOLD = TRUE,
  scale_design = TRUE,
  num.threads = 4,
  outfile = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{A list of sessions, where each session is a list with elements
BOLD, design and nuisance. See \code{?create.session} and \code{?is.session}
for more details.
List element names represent session names.}

\item{vertices}{A \eqn{V x 3} matrix, where each row contains the Euclidean
coordinates at which a given vertex in the mesh is located. \eqn{V} is the
number of vertices in the mesh}

\item{faces}{An \eqn{F x 3} matrix, where each row contains the vertex
indices for a given triangular face in the mesh. \eqn{F} is the number of
faces in the mesh.}

\item{mesh}{An \code{"inla.mesh"} object (see \code{\link{make_mesh}} for
surface data).}

\item{mask}{(Optional) A length \eqn{V} logical vector indicating if each
vertex is to be included.}

\item{scale_BOLD}{Scale timeseries data so estimates represent percent signal
change? Default: \code{TRUE}. If \code{FALSE}, the data will just be
centered.}

\item{scale_design}{Scale the design matrix by dividing each column by its
maximum and then subtracting the mean? Default: \code{TRUE}.}

\item{num.threads}{Not used yet.}

\item{outfile}{File name where results will be written (for use by
\code{BayesGLM2}).}

\item{verbose}{Should occasional updates be printed? Default: \code{TRUE}.}
}
\value{
A list containing...
}
\description{
Applies spatial Bayesian GLM to task fMRI data
}
\section{INLA Requirement}{

This function requires the \code{INLA} package, which is not a CRAN package.
See \url{http://www.r-inla.org/download} for easy installation instructions.
}

