% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_functions.R
\name{BayesGLMEM_cifti}
\alias{BayesGLMEM_cifti}
\title{BayesGLM Expectation Maximization for CIFTI}
\usage{
BayesGLMEM_cifti(
  cifti_fname,
  surfL_fname = NULL,
  surfR_fname = NULL,
  brainstructures = c("left", "right"),
  wb_path = NULL,
  design = NULL,
  onsets = NULL,
  TR = NULL,
  nuisance = NULL,
  nuisance_include = c("drift", "dHRF"),
  scale_BOLD = TRUE,
  scale_design = TRUE,
  EM_method = "joint",
  use_SQUAREM = TRUE,
  pct_change_limit = 1,
  num.threads = 1,
  GLM_method = "both",
  session_names = NULL,
  resamp_res = 10000,
  verbose = FALSE,
  outfile = NULL
)
}
\arguments{
\item{cifti_fname}{File path (or vector thereof, for multiple-session modeling) of CIFTI-format fMRI timeseries data (*.dtseries.nii).}

\item{surfL_fname}{File path of GIFTI-format left cortical surface (*.surf.gii). Must be provided if brainstructures includes "left" and GLM_method is "Bayesian" or "both".}

\item{surfR_fname}{File path of GIFTI-format right cortical surface (*.surf.gii). Must be provided if brainstructures includes "right" and GLM_method is "Bayesian" or "both".}

\item{brainstructures}{Character vector indicating which brain structure(s)
to obtain: \code{"left"} (left cortical surface) and/or \code{"right"} (right
cortical surface). Default: \code{c("left","right")} (entire cortical surface).
Note that the subcortical models have not yet been implemented.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable.
If not provided, should be set with
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}

\item{design, onsets, TR}{Either provide \code{design}, or provide both \code{onsets} and \code{TR}.

\code{design} is a \eqn{T x K} task design matrix (or list of such
matrices, for multiple-session modeling) with column names representing
tasks. Each column represents the expected BOLD response due to each task,
a convolution of the hemodynamic response function (HRF) and the task
stimulus. Note that the scale of the regressors will affect the scale and
interpretation of the beta coefficients, so imposing a proper scale (e.g.,
set maximum to 1) is recommended.

\code{onsets} is a matrix of onsets (first column) and durations (second column)
for each task in seconds, organized as a list where each element of the
list corresponds to one task. Names of list should be task names. (Or for
multi-session modeling, a list of such lists.)

\code{TR} is the temporal resolution of the data in seconds.}

\item{nuisance}{(Optional) A TxJ matrix of nuisance signals (or list of such matrices, for multiple-session modeling).}

\item{nuisance_include}{(Optional) Additional nuisance covariates to include.  Default is 'drift' (linear and quadratic drift terms) and 'dHRF' (temporal derivative of each column of design matrix).}

\item{scale_BOLD}{Scale timeseries data so estimates represent percent signal
change? Default: \code{TRUE}. If \code{FALSE}, the data will just be
centered.}

\item{scale_design}{Scale the design matrix by dividing each column by its
maximum and then subtracting the mean? Default: \code{TRUE}.}

\item{EM_method}{Either "joint" or "separate" for choosing whether covariates
should share hyperparameter values.}

\item{use_SQUAREM}{(logical) Should the SQUAREM package be used to speed up
convergence?}

\item{pct_change_limit}{The numeric threshold for the percent
(NOT probability) change below which the EM algorithm will stop updates
(default = 1).}

\item{num.threads}{(optional) allows users to specify the number of threads used
to work in parallel across the different tasks}

\item{GLM_method}{Either 'Bayesian' for spatial Bayesian GLM EM only, 'classical' for the classical GLM only, or 'both' to return both classical and Bayesian estimates of task activation.}

\item{session_names}{(Optional) A vector of names corresponding to each
session.}

\item{resamp_res}{The number of vertices to which each cortical surface should be resampled, or NULL if no resampling is to be performed. For computational feasibility, a value of 10000 or lower is recommended.}

\item{verbose}{Should INLA be run in verbose mode? Default: \code{FALSE}.}

\item{outfile}{(Optional) File name (without extension) of output file for
\code{"BayesGLM"} result to use in Bayesian group modeling.
\code{"_left.rds"} or \code{"_right.rds"} will be appended for the left
cortex and right cortex results, respectively. Default: \code{NULL}
(do not save the results to any file).}
}
\value{
An object of class \code{"BayesGLM"}, a list containing...
}
\description{
Performs spatial Bayesian GLM expectation maximization on the cortical
surface for fMRI task activation
}
\section{Connectome Workbench Requirement}{

This function uses a system wrapper for the 'wb_command' executable. The
user must first download and install the Connectome Workbench, available
from https://www.humanconnectome.org/software/get-connectome-workbench .
The \code{wb_path} argument is the full file path to the Connectome
Workbench folder. (The full file path to the 'wb_cmd' executable also
works.)
}

