% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_activations.R
\name{BH_FDR}
\alias{BH_FDR}
\title{Significance using Benjamini-Hochsberg False Discovery Rate}
\usage{
BH_FDR(p, FDR = 0.05)
}
\arguments{
\item{p}{A vector of p-values}

\item{FDR}{The false discovery rate}
}
\value{
A 0-1 vector that classifies the p-value vector as "significant" or not.
}
\description{
Significance using Benjamini-Hochsberg False Discovery Rate
}
