% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_template-data.R
\docType{data}
\name{binary_template}
\alias{binary_template}
\title{Binary Template of Single Axial Slice of Resampled MNI152 Brain Image}
\format{
An object of class \code{"matrix" "array"}
}
\usage{
data(binary_template)
}
\description{
Data come from the Montreal Neurological Institute, as supplied with FSL
}
\references{
M. Jenkinson, C.F. Beckmann, T.E. Behrens, M.W. Woolrich,
S.M. Smith. FSL. NeuroImage, 62:782-90, 2012

M. Jenkinson and S.M. Smith. A global optimization method for robust affine registration of brain images. Medical Image Analysis, 5(2):143-156, 2001.

M. Jenkinson, P.R. Bannister, J.M. Brady, and S.M. Smith. Improved optimization for the robust and accurate linear registration and motion correction of brain images. NeuroImage, 17(2):825-841, 2002.
}
\keyword{datasets}
