% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs_BayesGLM.R
\name{BOLD_Param_BayesGLM}
\alias{BOLD_Param_BayesGLM}
\title{BOLD}
\arguments{
\item{BOLD}{fMRI timeseries data in CIFTI format ("*.dtseries.nii").
For single-session analysis this can be a file path to a CIFTI file or a
\code{"xifti"} object from the \code{ciftiTools} package. For multi-session
analysis this can be a vector of file paths or a list of \code{"xifti"}
objects.

If \code{BOLD} is a \code{"xifti"} object(s), the surfaces, if any, will be
used for the spatial model. However, if \code{surfL} and \code{surfR} are
provided, they will override any surfaces in \code{BOLD}.}
}
\description{
BOLD
}
