% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary.R
\name{boundary_layers}
\alias{boundary_layers}
\title{}
\usage{
boundary_layers(faces, mask, boundary_width = 10)
}
\arguments{
\item{faces}{}

\item{mask}{}

\item{boundary_width}{}
}
\value{

}
\description{
\verb{}
}
