% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{boundary_mask}
\alias{boundary_mask}
\title{}
\usage{
boundary_mask(faces, mask, boundary_width)
}
\arguments{
\item{faces}{}

\item{mask}{}

\item{boundary_width}{}
}
\value{

}
\description{
\verb{}
}
\keyword{internal}
