% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs_BayesGLM.R
\name{brainstructures_Param_BayesGLM}
\alias{brainstructures_Param_BayesGLM}
\title{brainstructures}
\arguments{
\item{brainstructures}{Character vector indicating which brain structure(s)
of \code{BOLD} to analyze: \code{"left"} cortex; \code{"right"} cortex;
and/or \code{"subcortical"} structures. Or \code{"all"} to model all three.
Default: \code{c("left","right")} (cortex only).}
}
\description{
brainstructures
}
