% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs_BayesGLM_cifti.R
\name{brainstructures_param_BayesGLM_cifti}
\alias{brainstructures_param_BayesGLM_cifti}
\title{brainstructures}
\arguments{
\item{brainstructures}{Character vector indicating which brain structure(s)
of \code{BOLD} to analyze: \code{"left"} cortex; \code{"right"} cortex;
and/or \code{"subcortical"} structures. Default: \code{c("left","right")}
(cortex only).}
}
\description{
brainstructures
}
