% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{buffer_Param}
\alias{buffer_Param}
\title{buffer}
\arguments{
\item{buffer}{For volumetric model. The number of extra voxel layers around
the bounding box. Set to \code{NULL} for no buffer. (We recommend not
changing \code{buffer} unless you know what you're doing. Instead, to reduce
the number of boundary voxels, adjust \code{nbhd_order}).}
}
\description{
buffer
}
