% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cbind2}
\alias{cbind2}
\title{\code{cbind} if first argument might be \code{NULL}}
\usage{
cbind2(mat_or_NULL, to_add)
}
\arguments{
\item{mat_or_NULL}{\code{NULL} or a numeric matrix}

\item{to_add}{A numeric matrix with the same number of rows as \code{mat_or_NULL}}
}
\value{
\code{cbind(mat_or_NULL, to_add)}, or just \code{to_add} if the first argument is NULL.
}
\description{
\code{cbind}, but return the second argument if the first is \code{NULL}
}
\keyword{internal}
