% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_HRFs.R
\name{cderiv}
\alias{cderiv}
\title{Central derivative}
\usage{
cderiv(x)
}
\arguments{
\item{x}{A numeric matrix, or a vector which will be converted to a
single-column matrix.}
}
\value{
\code{x} with the derivative taken for each column.
}
\description{
Take the central derivative of numeric vectors by averaging the forward and
backward differences.
}
