% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{check_BayesGLM}
\alias{check_BayesGLM}
\title{Check arguments and packages for BayesGLM-related functions}
\usage{
check_BayesGLM(require_PARDISO = TRUE)
}
\arguments{
\item{require_PARDISO}{Is PARDISO required? Default: \code{TRUE}.}
}
\value{
\code{NULL}, invisibly
}
\description{
Check arguments and packages for \code{BayesGLM} and \code{BayesGLM_cifti}.
}
\keyword{internal}
