% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{check_INLA}
\alias{check_INLA}
\title{Check INLA and PARDISO}
\usage{
check_INLA(require_PARDISO = TRUE)
}
\arguments{
\item{require_PARDISO}{Is PARDISO required? Default: \code{FALSE}.}
}
\value{
\code{NULL}, invisibly
}
\description{
Check INLA and PARDISO
}
\keyword{internal}
