% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM2.R
\name{cholQsample}
\alias{cholQsample}
\title{Sample from the multivariate normal distribution with Cholesky(Q)}
\usage{
cholQsample(n, mu, cholQ)
}
\arguments{
\item{n}{number of samples}

\item{mu}{mean vector}

\item{cholQ}{Cholesky decomposition of the precision (found via \code{Matrix::Cholesky(Q)})}
}
\value{
an n x p matrix of samples from the MVN distribution
}
\description{
Sample from the multivariate normal distribution with Cholesky(Q)
}
\keyword{internal}
