% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicalGLM.R
\name{classicalGLM}
\alias{classicalGLM}
\title{Fit classical massive univariate GLM to task fMRI data}
\usage{
classicalGLM(data, scale_BOLD = TRUE, scale_design = TRUE)
}
\arguments{
\item{data}{A list of sessions, where each session is a list with elements
BOLD, design and nuisance.  See \code{?create.session} and \code{?is.session} for more details.
List element names represent session names.}

\item{scale_BOLD}{If TRUE, scale timeseries data so estimates represent
percent signal change.  If FALSE, just center the data and design to
exclude the baseline field.}

\item{scale_design}{If TRUE, scale the design matrix by dividing each column
by its maximum value, and then subtracting the new column mean.}
}
\value{
A list of classical GLM task activation estimates, where each element represents a session.
}
\description{
Fit classical massive univariate GLM to task fMRI data
}
