% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicalGLM2.R
\name{classicalGLM2}
\alias{classicalGLM2}
\title{Group-level Classical GLM}
\usage{
classicalGLM2(
  results,
  brainstructure = NULL,
  session_name = NULL,
  gamma = 0,
  correction = c("FWER", "FDR"),
  alpha = 0.05
)
}
\arguments{
\item{results}{Either (1) a list of length M of objects of class
BayesGLM_cifti with classical results, or (2) a character vector of length M
of file names output from the BayesGLM_cifti function. M is the number of
subjects.}

\item{brainstructure}{Character string indicating which brain structure(s)
to obtain: \code{"cortexL"} (left cortical surface) or \code{"cortexR"} (right
cortical surface) or \code{"subcortical"}. Default: \code{c("cortexL","cortexR")}
(entire cortical surface).}

\item{session_name}{(character) The name of the session that should be
examined. If \code{NULL} (default), the average across all sessions is used.}

\item{gamma}{(For inference only) List of vectors of activation thresholds
for the excursion set (each element corresponding to one contrast).
Remember that if a contrast is not specified, the average is found.}

\item{correction}{Type of multiple comparisons correction, 'FDR'
(Benjamini Hochberg) or 'FWER' (Bonferroni correction)}

\item{alpha}{Significance level (e.g. 0.05)}
}
\value{
a list containing...
}
\description{
Group-level Classical GLM
}
