% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicalGLM.R
\name{classicalGLM_pw}
\alias{classicalGLM_pw}
\title{Fit classical massive univariate GLM to task fMRI data after prewhitening}
\usage{
classicalGLM_pw(data, scale_BOLD = TRUE, scale_design = TRUE)
}
\arguments{
\item{data}{A list of sessions, where each session is a list with elements
BOLD, design and nuisance.  See \code{?create.session} and \code{?is.session} for more details.
List element names represent session names.}

\item{scale_BOLD}{Scale timeseries data so estimates represent percent signal
change? Default: \code{TRUE}. If \code{FALSE}, the data will just be
centered.}

\item{scale_design}{Scale the design matrix by dividing each column by its
maximum and then subtracting the mean? Default: \code{TRUE}.}
}
\value{
A list of classical GLM task activation estimates, where each element represents a session.
}
\description{
Fit classical massive univariate GLM to task fMRI data after prewhitening
}
