% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{combine_sessions_Param}
\alias{combine_sessions_Param}
\title{combine_sessions}
\arguments{
\item{combine_sessions}{Average task estimates (betas) over multiple sessions?
Default: \code{TRUE}.}
}
\description{
combine_sessions
}
