% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_templateICA.R
\name{compute_mu_s}
\alias{compute_mu_s}
\title{}
\usage{
compute_mu_s(y_vec, D, Dinv_s0, R_inv, theta, P, C_diag)
}
\arguments{
\item{y_vec}{}

\item{D}{}

\item{Dinv_s0}{}

\item{R_inv}{}

\item{theta}{}

\item{P}{}

\item{C_diag}{}
}
\value{

}
\description{

}
