% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_vertex_areas.R
\name{compute_vertex_areas}
\alias{compute_vertex_areas}
\title{Compute surface areas of each vertex in a triangular mesh.}
\usage{
compute_vertex_areas(mesh)
}
\arguments{
\item{mesh}{object of type mesh}
}
\value{
Vector of areas
}
\description{
Compute surface areas of each vertex in a triangular mesh.
}
\examples{
\dontrun{}
}
