% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spde_surf.R
\name{create_spde_surf}
\alias{create_spde_surf}
\title{Create SPDE for 2D surface data}
\usage{
create_spde_surf(mesh)
}
\arguments{
\item{mesh}{A list with elements \code{n} (the number of vertices),
\code{loc} (the n x 3 matrix of vertices), \code{graph$tv} (the matrix of
faces with 3 columns), and \code{idx$loc} (a vector of the indices ,
usually 1:n)}
}
\value{
SPDE object representing triangular mesh structure on data locations
}
\description{
Create SPDE for 2D surface data
}
\keyword{internal}
