% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCT_FFT_detrending.R
\name{dct_bases}
\alias{dct_bases}
\title{Generate cosine bases for the DCT}
\usage{
dct_bases(T_, n)
}
\arguments{
\item{T_}{Length of timeseries}

\item{n}{Number of cosine bases}
}
\value{
Matrix with cosine bases along columns
}
\description{
Generate cosine bases for the DCT
}
\keyword{internal}
