% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs_BayesGLM.R
\name{design_Param_BayesGLM}
\alias{design_Param_BayesGLM}
\title{design}
\arguments{
\item{design}{A numeric matrix or \code{data.frame}, or a
\code{"BayesfMRI_design"} object from \code{make_design}. Can also
be an array where the third dimension is the same length as the number of
data locations, to model each location with its own design.}
}
\description{
design
}
