% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs_BayesGLM_cifti.R
\name{detrending_param_BayesGLM_cifti}
\alias{detrending_param_BayesGLM_cifti}
\title{detrending}
\arguments{
\item{TR}{Temporal resolution of the data, in seconds.}

\item{hpf, DCT}{Add DCT bases to \code{nuisance} to apply a temporal high-pass
filter to the data, for detrending? Only one of these should be provided:
the filter frequency \code{hpf}, or the number of DCT bases \code{DCT}.
If \code{hpf} is provided, \code{TR} must be provided too and the
corresponding number of DCT bases will be calculated.

Default: \code{DCT=4}. For typical \code{TR} and length of \code{BOLD}, four
DCT bases amounts to a lower frequency cutoff than the approximately .01 Hz
used in most studies. We selected this default to err on the side of
retaining more low-frequency information, but recommend setting these
arguments to values most appropriate for the analysis at hand.

Using at least two DCT bases is as sufficient for detrending as using linear
and quadratic drift terms in the design matrix. So if DCT detrending is
being used used, there is no need to add linear and quadratic drift terms to
\code{nuisance}.}
}
\description{
detrending
}
