% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{dgCMatrix_cols_to_zero}
\alias{dgCMatrix_cols_to_zero}
\title{Set column values to zero for sparse matrix}
\usage{
dgCMatrix_cols_to_zero(mat, cols)
}
\arguments{
\item{mat}{dgCMatrix}

\item{cols}{The column indices to set to zero}
}
\value{
The modified sparse matrix
}
\description{
Set column values to zero for sparse matrix
}
\keyword{internal}
