% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticICA.cifti.R
\name{diagnosticICA.cifti}
\alias{diagnosticICA.cifti}
\title{Run diagnostic ICA based on CIFTI-format BOLD data and CIFTI-based template}
\usage{
diagnosticICA.cifti(
  cifti_fname,
  templates,
  brainstructures = c("left", "right"),
  scale = TRUE,
  maxQ = NULL,
  maxiter = 100,
  epsilon = 0.001,
  verbose = TRUE,
  write_dir = NULL
)
}
\arguments{
\item{cifti_fname}{File path of CIFTI-format timeseries data (ending in .dtseries.nii).}

\item{templates}{LIST or VECTOR of templates, each the result of call to \code{estimate_template.cifti()} (one template for each group).
Either a list of objects of class \code{template.cifti} OR a vector of file path prefixes (the part before "_mean.dscalar.nii"
and "_var.dscalar.nii") of cifti files written out by \code{estimate_template.cifti()}.}

\item{brainstructures}{Character vector indicating which brain structure(s)
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right
cortical surface) and/or \code{"subcortical"} (subcortical and cerebellar
gray matter). Can also be \code{"all"} (obtain all three brain structures).
Default: \code{c("left","right")} (cortical surface only).  For spatial
modeling (performed if \code{surfL_fname} and/or \code{surfR_fname} provided), subcortical
locations will be ignored.}

\item{scale}{Logical indicating whether BOLD data should be scaled by the spatial
standard deviation before model fitting. If done when estimating templates, should be done here too.}

\item{maxQ}{Maximum number of ICs (template+nuisance) to identify (L <= maxQ <= T)}

\item{maxiter}{Maximum number of EM iterations}

\item{epsilon}{Smallest proportion change between iterations (e.g. .01)}

\item{verbose}{If TRUE, display progress of algorithm}

\item{write_dir}{Where should any output files be written? \code{NULL} (default) will write them to the current working directory.}
}
\value{
A list containing the subject IC estimates (class 'xifti'), the subject IC variance estimates (class 'xifti'), and the result of the model call to \code{diagnosticICA} (class 'dICA')
}
\description{
Run diagnostic ICA based on CIFTI-format BOLD data and CIFTI-based template
}
