% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticICA.R
\name{diagnosticICA}
\alias{diagnosticICA}
\title{}
\usage{
diagnosticICA(
  template_mean,
  template_var,
  BOLD,
  scale = TRUE,
  maxQ = NULL,
  maxiter = 100,
  epsilon = 0.001,
  verbose = TRUE
)
}
\arguments{
\item{template_mean}{}

\item{template_var}{}

\item{BOLD}{}

\item{scale}{}

\item{maxQ}{}

\item{maxiter}{}

\item{epsilon}{}

\item{verbose}{}
}
\value{

}
\description{

}
