% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.getSqrtInvCpp}
\alias{.getSqrtInvCpp}
\title{Get the prewhitening matrix for a single data location}
\usage{
.getSqrtInvCpp(AR_coeffs, nTime, avg_var)
}
\arguments{
\item{AR_coeffs}{a length-p vector where p is the AR order}

\item{nTime}{(integer) the length of the time series that is being prewhitened}

\item{avg_var}{a scalar value of the residual variances of the AR model}
}
\description{
Get the prewhitening matrix for a single data location
}
