% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.initialKP}
\alias{.initialKP}
\title{Find the initial values of kappa2 and phi}
\usage{
.initialKP(theta, spde, w, n_sess, tol, verbose)
}
\arguments{
\item{theta}{a vector of length two containing the range and scale parameters
kappa2 and phi, in that order}

\item{spde}{a list containing the sparse matrix elements Cmat, Gmat, and GtCinvG}

\item{w}{the beta_hat estimates for a single task}

\item{n_sess}{the number of sessions}

\item{tol}{the stopping rule tolerance}

\item{verbose}{(logical) Should intermediate output be displayed?}
}
\description{
Find the initial values of kappa2 and phi
}
