% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_utils.R
\name{ELL}
\alias{ELL}
\title{Expected log-likelihood function}
\usage{
ELL(Q, sigma2, model_data, Psi, mu, Sigma, A)
}
\arguments{
\item{Q}{precision matrix}

\item{sigma2}{noise variance}

\item{model_data}{list with X and y}

\item{Psi}{data locations to mesh locations matrix}

\item{mu}{posterior mean of w}

\item{Sigma}{posterior covariance of w}

\item{A}{crossprod(X\%*\%Psi)}
}
\value{
scalar expected log-likelihood
}
\description{
Expected log-likelihood function
}
\keyword{internal}
