% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_diagnosticICA.R
\name{EM_diagnosticICA}
\alias{EM_diagnosticICA}
\title{}
\usage{
EM_diagnosticICA(
  template_mean,
  template_var,
  BOLD,
  theta0,
  C_diag,
  maxiter = 100,
  epsilon = 0.001,
  verbose = TRUE
)
}
\arguments{
\item{template_mean}{}

\item{template_var}{}

\item{BOLD}{}

\item{theta0}{}

\item{C_diag}{}

\item{maxiter}{}

\item{epsilon}{}

\item{verbose}{}
}
\value{

}
\description{

}
