% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{EM_Param}
\alias{EM_Param}
\title{EM}
\arguments{
\item{EM}{(logical) Should the EM implementation of the Bayesian GLM be used?
Default: \code{FALSE}. This method is still in development.}
}
\description{
EM
}
