% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_templateICA.R
\name{EM_templateICA}
\alias{EM_templateICA}
\alias{EM_templateICA.spatial}
\alias{EM_templateICA.independent}
\title{}
\usage{
EM_templateICA.spatial(
  template_mean,
  template_var,
  mesh,
  BOLD,
  theta0,
  C_diag,
  common_smoothness = TRUE,
  maxiter = 100,
  epsilon = 0.001,
  verbose = FALSE,
  dim_reduce_flag
)

EM_templateICA.independent(
  template_mean,
  template_var,
  BOLD,
  theta0,
  C_diag,
  maxiter = 100,
  epsilon = 0.001,
  verbose
)
}
\arguments{
\item{template_mean}{}

\item{template_var}{}

\item{mesh}{\verb{}}

\item{BOLD}{}

\item{theta0}{}

\item{C_diag}{}

\item{common_smoothness}{}

\item{maxiter}{}

\item{epsilon}{}

\item{verbose}{}

\item{dim_reduce_flag}{}
}
\value{

}
\description{

}
\details{

}
