% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.R
\name{estimate_template.cifti}
\alias{estimate_template.cifti}
\title{}
\usage{
estimate_template.cifti(
  cifti_fnames,
  cifti_fnames2 = NULL,
  GICA_fname,
  inds = NULL,
  scale = TRUE,
  brainstructures = c("left", "right"),
  verbose = TRUE
)
}
\arguments{
\item{cifti_fnames}{}

\item{cifti_fnames2}{}

\item{GICA_fname}{}

\item{inds}{}

\item{scale}{}

\item{brainstructures}{}

\item{verbose}{}
}
\value{

}
\description{

}
