% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursions.R
\name{excursions.inla.no.spurious}
\alias{excursions.inla.no.spurious}
\title{}
\usage{
excursions.inla.no.spurious(
  result.inla,
  stack,
  name = NULL,
  tag = NULL,
  u,
  alpha = 1,
  type,
  n.iter = 10000,
  F.limit,
  u.link = FALSE,
  method,
  ind = NULL,
  max.size,
  verbose = 0,
  max.threads = 0,
  seed = NULL,
  mesh,
  area.limit,
  use.continuous = FALSE,
  plot.progress = FALSE
)
}
\arguments{
\item{result.inla}{}

\item{stack}{}

\item{name}{}

\item{tag}{}

\item{u}{}

\item{alpha}{}

\item{type}{}

\item{n.iter}{}

\item{F.limit}{}

\item{u.link}{}

\item{method}{}

\item{ind}{}

\item{max.size}{}

\item{verbose}{}

\item{max.threads}{}

\item{seed}{}

\item{mesh}{}

\item{area.limit}{}

\item{use.continuous}{}

\item{plot.progress}{}
}
\value{

}
\description{

}
