% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_estimates.R
\name{extract_estimates}
\alias{extract_estimates}
\title{Extract Estimates of Activation}
\usage{
extract_estimates(object, session_names, stat = "mean")
}
\arguments{
\item{object}{An object of class ‘"inla"’, a result of a call to inla}

\item{session_names}{Vector of fMRI session names}

\item{stat}{A string representing the posterior summary statistic to be returned}
}
\value{
Estimates from inla model
}
\description{
Obtains the posterior mean or other summary statistic for each latent field
}
\examples{
\dontrun{}
}
