% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM2.R
\name{F.logwt}
\alias{F.logwt}
\title{}
\usage{
F.logwt(theta, spde, mu_theta, Q_theta, M)
}
\arguments{
\item{theta}{}

\item{spde}{}

\item{mu_theta}{}

\item{Q_theta}{}

\item{M}{}
}
\value{

}
\description{

}
\section{}{
NA
}

\keyword{internal}
