% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCT_FFT_detrending.R
\name{fft_detrend}
\alias{fft_detrend}
\title{FFT detrending}
\usage{
fft_detrend(X, N)
}
\arguments{
\item{X}{\eqn{T \times V} numeric matrix. Each column is a voxel or vertex
time series.}

\item{N}{Number of FFT entries to remove from each end of the vector}
}
\value{
Detrended \code{X}
}
\description{
FFT detrending
}
\keyword{internal}
