% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursions.R
\name{find.connected}
\alias{find.connected}
\title{Find connected indices on mesh}
\usage{
find.connected(ind, mesh)
}
\arguments{
\item{ind}{indices}

\item{mesh}{The inla.mesh}
}
\value{
list of connected components in \code{ind}
}
\description{
Find connected indices on mesh
}
\keyword{internal}
