% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursions.R
\name{find.smallest.activation}
\alias{find.smallest.activation}
\title{Find smallest activation}
\usage{
find.smallest.activation(res, mesh, area.limit, factor, area.el)
}
\arguments{
\item{res}{Undocumented}

\item{mesh}{Undocumented}

\item{area.limit}{Undocumented}

\item{factor}{Undocumented}

\item{area.el}{Undocumented}
}
\description{
Find smallest activation
}
