% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_design.R
\name{format_EV}
\alias{format_EV}
\title{Is this a valid entry in \code{EVs}?}
\usage{
format_EV(EV)
}
\arguments{
\item{EV}{The putative EVs matrix or data frame}
}
\value{
Length-one logical vector.
}
\description{
Is this valid data for a single task's EVs? Expects a data.frame or
numeric matrix with two numeric columns, EVs and durations, and at least
one row. Or, the value \code{NA} for an empty task.
}
\keyword{internal}
