% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spde_vol3D.R
\name{galerkin_db}
\alias{galerkin_db}
\title{Create FEM matrices}
\usage{
galerkin_db(FV, P)
}
\arguments{
\item{FV}{Matrix of faces in triangularization}

\item{P}{Matrix of vertex locations in triangularization}
}
\value{
A list of matrices C and G appearing in sparse SPDE precision
}
\description{
Create FEM matrices
}
