% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{get_nV}
\alias{get_nV}
\title{Get number of locations for various masks}
\usage{
get_nV(spatial)
}
\arguments{
\item{spatial}{\code{spatial}}
}
\value{
A list of two: \code{T} for the total number of locations, and
\code{D} for the number of data locations. If \code{spatial} is provided for
voxel data, there is also \code{DB} for the number of data locations plus
the number of boundary locations.
}
\description{
Get number of locations for various masks: total, model, data.
}
\keyword{internal}
