% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{get_posterior_densities}
\alias{get_posterior_densities}
\title{Extracts posterior density estimates for hyperparameters}
\usage{
get_posterior_densities(object, spde, beta_names = NULL)
}
\arguments{
\item{object}{An object of class \code{"inla"}, a result of a call to
\code{inla()}}

\item{spde}{The model used for the latent fields in the \code{inla()} call,
an object of class \code{"inla.spde"}}

\item{beta_names}{(Optional) Descriptive names of model regressors (tasks).}
}
\value{
Long-form data frame containing posterior densities for the
hyperparameters associated with each latent field
}
\description{
Extracts posterior density estimates for hyperparameters
}
\section{INLA Requirement}{

This function requires the \code{INLA} package, which is not a CRAN package.
See \url{http://www.r-inla.org/download} for easy installation instructions.
}

\keyword{internal}
