% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bayesglm_utils.R
\name{get_posterior_densities2}
\alias{get_posterior_densities2}
\title{Extracts posterior density estimates for hyperparameters for volumetric SPDE}
\usage{
get_posterior_densities2(INLA_model_obj, field_names)
}
\arguments{
\item{INLA_model_obj}{An object of class \code{"inla"}, a result of a call to
\code{inla()}}

\item{field_names}{Descriptive names of model regressors (fields).}
}
\value{
Long-form data frame containing posterior densities for the
hyperparameters associated with each latent field
}
\description{
Extracts posterior density estimates for hyperparameters for volumetric SPDE
}
\section{INLA Requirement}{

This function requires the \code{INLA} package, which is not a CRAN package.
See \url{https://www.r-inla.org/download-install} for easy installation instructions.
}

\keyword{internal}
