% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM.R
\name{get_posterior_densities_vol3D}
\alias{get_posterior_densities_vol3D}
\title{}
\usage{
get_posterior_densities_vol3D(object, spde)
}
\arguments{
\item{object}{}

\item{spde}{}
}
\value{

}
\description{

}
\note{

}
