% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spde_vol3D.R
\name{get_spacing}
\alias{get_spacing}
\title{Determine grid spacing}
\usage{
get_spacing(locations)
}
\arguments{
\item{locations}{Vector of vertex locations in one dimension}
}
\value{
Value of minimum spacing between two locations
}
\description{
Determine grid spacing
}
