% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prewhitening.R
\name{getInvCovAR}
\alias{getInvCovAR}
\title{Compute inverse covariance matrix for AR process (up to a constant scaling factor)}
\usage{
getInvCovAR(ar, ntime)
}
\arguments{
\item{ar}{vector of p AR parameters}

\item{ntime}{number of time points in timeseries}
}
\value{
inverse covariance matrix for AR process (up to a constant scaling factor)
}
\description{
Compute inverse covariance matrix for AR process (up to a constant scaling factor)
}
