% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prewhitening.R
\name{getSqrtInv}
\alias{getSqrtInv}
\title{Compute square root of a symmetric matrix (such as inverse covariance mat) using SVD}
\usage{
getSqrtInv(Inv)
}
\arguments{
\item{Inv}{Symmetric matrix}
}
\value{
Matrix square root of \code{Inv}
}
\description{
Compute square root of a symmetric matrix (such as inverse covariance mat) using SVD
}
\keyword{internal}
