% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_classical.R
\name{GLM_classical}
\alias{GLM_classical}
\title{Classical GLM}
\usage{
GLM_classical(
  BOLD,
  design,
  nK2,
  nV_D,
  field_names,
  design_type,
  vcols_ss,
  nT_ss,
  nD,
  var_resid,
  sqrtInv_all,
  do_pw,
  compute_SE = TRUE
)
}
\arguments{
\item{BOLD, design, nK2, nV_D, field_names, design_type}{See \code{BayesGLM}.}

\item{vcols_ss, nT_ss, nD, var_resid, sqrtInv_all, do_pw}{See \code{BayesGLM}.}

\item{compute_SE}{compute SE?}
}
\value{
A list of results
}
\description{
Classical GLM for BayesGLM
}
\keyword{internal}
