% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_compare.R
\name{GLM_compare}
\alias{GLM_compare}
\title{Classical GLM for multiple models}
\usage{
GLM_compare(
  data,
  spatial,
  spatial_type,
  session_names,
  field_names,
  design_type,
  valid_cols,
  nT,
  nD,
  var_resid,
  sqrtInv_all
)
}
\arguments{
\item{data, spatial, spatial_type, session_names, field_names, design_type}{See \code{BayesGLM}.}

\item{valid_cols, nT, nD, var_resid, sqrtInv_all}{See \code{BayesGLM}.}
}
\value{
A list of results
}
\description{
Classical GLM for multiple models
}
\keyword{internal}
