% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_multi.R
\name{GLM_multi}
\alias{GLM_multi}
\title{GLM multi}
\usage{
GLM_multi(y, X, X2)
}
\arguments{
\item{y, X, X2}{BOLD, design, nuisance}

\item{spatial}{Spatial info}
}
\value{
Results for GLM multi
}
\description{
GLM multi
}
\keyword{internal}
