% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_utils.R
\name{GLMEM_chol_init}
\alias{GLMEM_chol_init}
\title{Initialize Cholesky decomposition for computational purposes}
\usage{
GLMEM_chol_init(theta, spde, K, A)
}
\arguments{
\item{theta}{vector of Bayesian GLM hyperparameters}

\item{spde}{spde object}

\item{K}{number of task covariates}

\item{A}{A <- spam::crossprod(model_data$X\%*\%Psi)}
}
\value{
The Cholesky decomposition of Sig_inv, used for improving computation
time and decreasing memory use.
}
\description{
Initialize Cholesky decomposition for computational purposes
}
