% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_utils.R
\name{GLMEM_fixptjoint}
\alias{GLMEM_fixptjoint}
\title{Fixed point function for the joint BayesGLMEM update algorithm}
\usage{
GLMEM_fixptjoint(theta, spde, model_data, Psi, K, A)
}
\arguments{
\item{theta}{a vector containing kappa2, phi, and sigma2, in that order}

\item{spde}{the spde object}

\item{model_data}{the model_data object containing \code{y} and \code{X}}

\item{Psi}{a conversion matrix (N by V) (or N by n)}

\item{K}{number of covariates}

\item{A}{The value for Matrix::crossprod(X\%*\%Psi) (saves time on computation)}
}
\value{
a vector with the same length as \code{theta}, the EM updates
}
\description{
Fixed point function for the joint BayesGLMEM update algorithm
}
\keyword{internal}
