% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{Calculate gradient}
\usage{
gradient(x)
}
\arguments{
\item{x}{A vector or matrix. If a matrix, compute gradient of each column of x.}
}
\value{
Gradient of x
}
\description{
Calculate gradient
}
