% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_regression.R
\name{hat_matrix}
\alias{hat_matrix}
\title{Hat matrix}
\usage{
hat_matrix(design)
}
\arguments{
\item{design}{The \eqn{T \times Q} design matrix}
}
\value{
The \eqn{T \times T} hat matrix
}
\description{
Get the hat matrix from a design matrix
}
\keyword{internal}
