% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs_BayesGLM.R
\name{hpf_Param_BayesGLM}
\alias{hpf_Param_BayesGLM}
\title{hpf}
\arguments{
\item{hpf}{Add DCT bases to \code{nuisance} to apply a temporal high-pass
filter to the data, for detrending? \code{hpf} is the filter frequency.
Use \code{NULL} to skip detrending. Detrending is strongly recommended for
fMRI data, to help reduce the autocorrelation in the residuals, so
\code{NULL} will induce a warning. Use \code{"already"} to disable the
warning while skipping highpass filtering.

Using at least two DCT bases is as sufficient for detrending as using linear
and quadratic drift terms in the nuisance matrix. So if DCT detrending is
being used here, there is no need to add linear and quadratic drift terms to
\code{nuisance}.}
}
\description{
hpf
}
