% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_activations.R
\name{id_activations.2means}
\alias{id_activations.2means}
\title{}
\usage{
id_activations.2means(
  model_obj,
  field_name = NULL,
  type = "point",
  n_sample = NULL
)
}
\arguments{
\item{model_obj}{\verb{}}

\item{field_name}{}

\item{type}{}

\item{n_sample}{}
}
\value{
\verb{}\verb{}\verb{}
}
\description{

}
