% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_activations.R
\name{id_activations.classical}
\alias{id_activations.classical}
\title{}
\usage{
id_activations.classical(
  model_obj,
  field_inds = NULL,
  session_name = NULL,
  alpha = 0.05,
  threshold = 0,
  correction = c("FWER", "FDR", "none"),
  mesh = NULL
)
}
\arguments{
\item{model_obj}{}

\item{field_inds}{}

\item{session_name}{}

\item{alpha}{}

\item{threshold}{}

\item{correction}{}

\item{mesh}{}
}
\value{

}
\description{

}
