% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_activations.R
\name{id_activations.em}
\alias{id_activations.em}
\title{Identify activations using joint posterior probabilities with EM results}
\usage{
id_activations.em(
  model_obj,
  tasks = NULL,
  sessions = NULL,
  alpha = 0.05,
  threshold
)
}
\arguments{
\item{model_obj}{An object of class \code{"BayesGLM"}, a result of a call
to \code{BayesGLMEM}.}

\item{tasks}{Name of latent field or vector of names on which to identify activations.  By default, analyze all tasks.}

\item{sessions}{(character) The name of the session that should be examined.
If \code{NULL} (default), the first session is used.}

\item{alpha}{Significance level (e.g. 0.05)}

\item{threshold}{Activation threshold (e.g. 1 for 1 percent signal change if scale=TRUE in model estimation)}
}
\value{
A list with two elements: \code{active}, which gives a matrix of zeros
and ones of the same dimension as \code{model_obj$task_estimates${sessions}},
and \code{excur_result}, an object of class \code{"excurobj"} (see \code{\link{excursions.inla}} for
more information).
}
\description{
Identifies areas of activation given an activation threshold and significance
level using joint posterior probabilities
}
\details{
For a given latent field, identifies locations that exceed a certain activation
threshold (e.g. 1 percent signal change) at a given significance level, based on the joint
posterior distribution of the latent field.
}
\keyword{internal}
