% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_activations.R
\name{id_activations.posterior}
\alias{id_activations.posterior}
\title{}
\usage{
id_activations.posterior(
  model_obj,
  field_name = NULL,
  threshold,
  alpha = 0.05,
  area.limit = NULL
)
}
\arguments{
\item{model_obj}{}

\item{field_name}{}

\item{threshold}{}

\item{alpha}{}

\item{area.limit}{}
}
\value{
\verb{}\verb{}\verb{}\verb{}
}
\description{

}
\details{

}
