% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_activations.R
\name{id_activations}
\alias{id_activations}
\title{Identify areas of activation}
\usage{
id_activations(
  model_obj,
  field_name = NULL,
  threshold = NULL,
  alpha = NULL,
  area.limit = NULL
)
}
\arguments{
\item{model_obj}{An object of class ‘"BayesGLM"’, a result of a call to BayesGLM}

\item{field_name}{Name of latent field, or vector of names, on which to identify activations}

\item{threshold}{For method='posterior' only: activation threshold (e.g. 0.01 for 1 percent signal change)}

\item{alpha}{For method='posterior' only: Significance level (e.g. 0.05)}

\item{area.limit}{For method='posterior' only: Below this value, activations will be considered spurious.  If NULL (default), no limit.}
}
\value{
A nested list, where the first layer separates by session, and the
second layer is another list of two elements: \code{active}, which gives a
matrix of zeros and ones of the same dimension as
\code{model_obj$beta_estimates${session_name}}, and \code{excur_result}, an
object of class \code{"excurobj"} (see \code{\link{excursions.inla}} for more information)
}
\description{
For each latent field, identifies locations that are activated
}
