% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_activations.R
\name{id_activations_cifti}
\alias{id_activations_cifti}
\title{}
\usage{
id_activations_cifti(
  model_obj,
  field_names = NULL,
  session_name = NULL,
  alpha = 0.05,
  method = c("Bayesian", "classical"),
  threshold = NULL,
  area.limit = NULL,
  correction = c("FWER", "FDR", "none"),
  excur_method = c("EB", "QC"),
  verbose = TRUE
)
}
\arguments{
\item{model_obj}{}

\item{field_names}{}

\item{session_name}{}

\item{alpha}{}

\item{method}{}

\item{threshold}{}

\item{area.limit}{}

\item{correction}{}

\item{excur_method}{}

\item{verbose}{}
}
\description{

}
