% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_utils.R
\name{init_objfn}
\alias{init_objfn}
\title{Objective function for the initialization of kappa2 and phi}
\usage{
init_objfn(theta, spde, beta_hat)
}
\arguments{
\item{theta}{a vector c(kappa2,phi)}

\item{spde}{an spde object}

\item{beta_hat}{vector}
}
\value{
scalar
}
\description{
Objective function for the initialization of kappa2 and phi
}
\keyword{internal}
