% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initialKP}
\alias{initialKP}
\title{Find the initial values of kappa2 and phi}
\usage{
initialKP(theta, spde, w, n_sess, tol, verbose)
}
\arguments{
\item{spde}{a list containing the sparse matrix elements Cmat, Gmat, and GtCinvG}

\item{w}{the beta_hat estimates for a single task}

\item{n_sess}{the number of sessions}

\item{tol}{the stopping rule tolerance}

\item{kappa2}{a scalar scale parameter}

\item{phi}{a scalar range parameter}
}
\description{
Find the initial values of kappa2 and phi
}
