% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{INLA_Description}
\alias{INLA_Description}
\title{INLA}
\description{
INLA
}
\section{INLA Requirement}{

This function requires the \code{INLA} package, which is not a CRAN package.
See \url{http://www.r-inla.org/download} for easy installation instructions.
}

