% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{INLA_Latent_Fields_Limit_Description}
\alias{INLA_Latent_Fields_Limit_Description}
\title{INLA Latent Fields}
\description{
INLA Latent Fields
}
\section{INLA Latent Fields Limit}{

INLA computation times increase greatly when the number of columns in the
design matrix exceeds five: when there are more than five tasks, or more
than three tasks each with a temporal derivative modeled as a field. In
cases like the latter, we recommend modeling the temporal derivatives as
nuisance signals using the option \code{dHRF_as="nuisance"}, rather than
modeling the temporal derivatives as fields.
}

