% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_mask.R
\name{intersect_mask}
\alias{intersect_mask}
\title{Intersection mask for BayesGLM or activations result}
\usage{
intersect_mask(x)
}
\arguments{
\item{x}{The list of \code{"fit_bglm"}, \code{"BGLM"}, or \code{"act_BGLM"} objects.}
}
\value{
The intersections masks
}
\description{
Intersection mask for BayesGLM or activations result
}
\keyword{internal}
