% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.BfMRI.sess.R
\name{is.a_session}
\alias{is.a_session}
\title{Validate an individual session in a \code{"BfMRI.sess"} object.}
\usage{
is.a_session(x)
}
\arguments{
\item{x}{The putative entry in a \code{"BfMRI.sess"} object.}
}
\value{
Logical. Is \code{x} a valid entry in a \code{"BfMRI.sess"} object?
}
\description{
Check if object is valid for a list entry in \code{"BfMRI.sess"}.
}
\details{
A valid entry in a \code{"BfMRI.sess"} object is a list with the following
named fields:
\itemize{
\item{"BOLD"}{\eqn{T \times V} matrix of BOLD responses, rows are time points, columns are voxels}
\item{"design"}{\eqn{T \times K} matrix containing the K task regressors}
\item{"nuisance"}{Optional. \eqn{T \times L} matrix containing the L nuisance regressors}
}
}
\keyword{internal}
