% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.BfMRI.sess.R
\name{is.BfMRI.sess}
\alias{is.BfMRI.sess}
\title{Validate a \code{"BfMRI.sess"} object.}
\usage{
is.BfMRI.sess(x)
}
\arguments{
\item{x}{The putative \code{"BfMRI.sess"} object.}
}
\value{
Logical. Is \code{x} a valid \code{"BfMRI.sess"} object?
}
\description{
Check if object is valid for a \code{"BfMRI.sess"} object.
}
\details{
A valid entry in a \code{"BfMRI.sess"} object is a list of lists, each with
the following named fields:
\itemize{
\item{"BOLD"}{\eqn{T \times V} matrix of BOLD responses, rows are time points, columns are voxels}
\item{"design"}{\eqn{T \times K} matrix containing the K task regressors}
\item{"nuisance"}{Optional. \eqn{T \times L} matrix containing the L nuisance regressors}
}
In addition, all sessions must have the same number of data locations and tasks.
}
