% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is.matrix.or.df}
\alias{is.matrix.or.df}
\title{Is a matrix or data.frame?}
\usage{
\method{is.matrix}{or.df}(q)
}
\arguments{
\item{q}{The object}
}
\value{
Length-one logical.
}
\description{
Is this a matrix or data.frame?
}
\keyword{internal}
