% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.session.R
\name{is.session_pw}
\alias{is.session_pw}
\title{Is session? (Prewhitened data)}
\usage{
is.session_pw(sess)
}
\arguments{
\item{sess}{A list representing a task fMRI session (see Details)}
}
\value{
True if all checks pass, or an (hopefully useful) error message
}
\description{
Verify that the \code{"session"} object is valid.
}
\details{
A valid "session" object is a list with the following named fields
\itemize{
\item BOLD: \eqn{T x V} matrix of BOLD responses, rows are time points, columns are voxels
\item design: \eqn{T x K} matrix containing the K task regressors
\item nuisance (optional): \eqn{T x L} matrix containing the L nuisance regressors
}
}
