% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_matrix_or_df}
\alias{is_matrix_or_df}
\title{Is a matrix or data.frame?}
\usage{
is_matrix_or_df(x)
}
\arguments{
\item{x}{The object}
}
\value{
Length-one logical.
}
\description{
Is this a matrix or data.frame?
}
\keyword{internal}
