% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_onsets}
\alias{is_onsets}
\title{Is this a valid \code{onsets}?}
\usage{
is_onsets(x)
}
\arguments{
\item{x}{The putative onsets matrix or data frame}
}
\value{
Length-one logical vector.
}
\description{
Is this valid data for event onsets? Expects a data.frame or numeric matrix
with two numeric columns, onsets and durations, and at least one row.
Raise an error if invalid; do nothing if valid.
}
\keyword{internal}
