% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_utils.R
\name{kappa_init_fn}
\alias{kappa_init_fn}
\title{Function to optimize over kappa2}
\usage{
kappa_init_fn(kappa2, phi, spde, beta_hat)
}
\arguments{
\item{kappa2}{scalar}

\item{phi}{scalar}

\item{spde}{an spde object}

\item{beta_hat}{vector}
}
\value{
a scalar
}
\description{
Function to optimize over kappa2
}
\keyword{internal}
