% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_templateICA.R
\name{LL2_kappa}
\alias{LL2_kappa}
\title{}
\usage{
LL2_kappa(kappa, mesh, OplusW, u, v, C1 = 1/(4 * pi), Q = NULL)
}
\arguments{
\item{kappa}{}

\item{mesh}{\verb{}}

\item{OplusW}{}

\item{u}{}

\item{v}{}

\item{C1}{}

\item{Q}{}
}
\value{

}
\description{

}
\details{

}
