% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_templateICA.R
\name{LL_SQUAREM}
\alias{LL_SQUAREM}
\title{}
\usage{
LL_SQUAREM(
  theta_vec,
  template_mean,
  template_var,
  mesh,
  BOLD,
  C_diag,
  s0_vec,
  D,
  Dinv_s0,
  common_smoothness,
  verbose,
  dim_reduce_flag
)
}
\arguments{
\item{theta_vec}{}

\item{template_mean}{}

\item{template_var}{}

\item{mesh}{}

\item{BOLD}{}

\item{C_diag}{}

\item{s0_vec}{}

\item{D}{}

\item{Dinv_s0}{}

\item{common_smoothness}{}

\item{verbose}{}

\item{dim_reduce_flag}{}
}
\value{

}
\description{

}
