% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bayesglm_utils.R
\name{log_kappa_tau}
\alias{log_kappa_tau}
\title{Make \code{log_kappa} and \code{log_tau}}
\usage{
log_kappa_tau(spatial, hyperpriors, verbose)
}
\arguments{
\item{spatial, hyperpriors, verbose}{See \code{fit_bayesglm}}
}
\description{
Make \code{log_kappa} and \code{log_tau}
}
\keyword{internal}
