% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logDetQ}
\alias{logDetQ}
\title{Find the log of the determinant of Q}
\usage{
logDetQ(kappa2, in_list, n_sess)
}
\arguments{
\item{kappa2}{a scalar}

\item{in_list}{a list with elements Cmat, Gmat, and GtCinvG}

\item{n_sess}{the integer number of sessions}
}
\description{
Find the log of the determinant of Q
}
