% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_2d_mesh.R
\name{make_2d_mesh}
\alias{make_2d_mesh}
\title{Make a 2-D Mesh}
\usage{
make_2d_mesh(mask)
}
\arguments{
\item{mask}{A binary (0/1) matrix that should be used to make a mesh for two-dimensional data}
}
\value{
A \code{inla.mesh} object that can be used to run the \code{BayesGLM_surface} function with two-dimensional data
}
\description{
Make a 2-D Mesh
}
\examples{
data(binary_template)
mesh_2d <- make_2d_mesh(binary_template)
}
