% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_A_mat.R
\name{make_A_mat}
\alias{make_A_mat}
\title{Make A matrix}
\usage{
make_A_mat(spatial)
}
\arguments{
\item{spatial}{See \code{BayesGLM}}
}
\value{
The A matrix
}
\description{
Make A matrix
}
\keyword{internal}
