% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_list.R
\name{make_data_list}
\alias{make_data_list}
\title{Make data list}
\usage{
make_data_list(y, X, betas, repls)
}
\arguments{
\item{y}{Vectorized BOLD data (all voxels, sessions, etc.)}

\item{X}{List of sparse design matrices from each session, each created using organize_data}

\item{betas}{List of bbeta objects from organize_replicates}

\item{repls}{List of repl objects from organize_replicates}
}
\value{
List
}
\description{
Make data list
}
\examples{
\dontrun{}
}
