% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_design.R
\name{make_design}
\alias{make_design}
\title{Make design for BayesGLM}
\usage{
make_design(
  design = NULL,
  design_multiple = NULL,
  onsets = NULL,
  nTime = NULL,
  TR = NULL,
  scale_design = TRUE
)
}
\arguments{
\item{design}{\eqn{T \times K} task design matrix. Each column
represents the expected BOLD response for a field, a convolution of
a hemodynamic response function (HRF) (or its derivative) with a task stimulus.
Field names must be given by the column names.

For multi-session modeling, this argument should be a list of such matrices,
and column names must match across sessions. Session names must be given by
the list names.

If any field is missing from any session, include a column of zeros in its
place.

Tasks convolved with an HRF derivative may be included in
\code{design} to model them spatially; alternatively, provide these in
\code{BayesGLM(_cifti)} as columns in \code{nuisance} to treat as nuisance
signals. Note that INLA computation times increase if the design matrix has
more than five columns.}

\item{design_multiple}{Same as \code{design}, but rather than a matrix, use
a \eqn{T \times K \times D} array of \eqn{D} different design matrices for
model comparison.}

\item{onsets}{A list where each element represents a
task, and the value of each element is a matrix of onsets (first column) and
durations (second column) for each stimuli (each row) of the corresponding
task. The units of both columns is seconds. Task names must be given by
the list names. \code{nTime} and \code{TR} are required if \code{onsets} is
provided.

For multi-session modeling, this argument should be a list of such lists,
and the task names must match across sessions. Session names must be given by
the list names.

If any task is missing from any session, include a list element equal to
\code{NA} in its place.}

\item{nTime, TR}{These are used if and only if \code{onsets} is provided.

\code{nTime} is the number of timepoints (volumes) in the task fMRI data.
This can either be a single positive integer, or a vector of positive
integers if there are multiple sessions each with different numbers of
timepoints. There is no default value, so \code{nTime} is required if
\code{onsets} is provided.

\code{TR} is the temporal resolution of the data, in seconds. There is no
default value, so \code{TR} is required if \code{onsets} is provided.}

\item{scale_design}{Scale the design matrix by dividing each column by its
maximum and then subtracting the mean? Default: \code{TRUE}. If
\code{FALSE}, the design matrix is centered but not scaled.}
}
\value{
A list of two: the design matrix/arrays or list of them; and, extra
results if \code{onsets} was provided.
}
\description{
Make the design matrix/array, or matrices/arrays, for \code{\link{BayesGLM}}
or \code{\link{BayesGLM_cifti}}, from information about the tasks.
}
\details{
Provide either (1) \code{design}, (2) \code{design_multiple}, or (3) all of
\code{onsets}, \code{nTime}, \code{TR}, and \code{dHRF}.
}
