% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_design.R
\name{make_design_from_onsets}
\alias{make_design_from_onsets}
\title{Make design from onsets}
\usage{
make_design_from_onsets(onsets, nTime, TR)
}
\arguments{
\item{onsets, nTime, TR}{See \code{\link{make_design}}.}
}
\value{
A list of two: the list of design matrices; and, extra results.
}
\description{
Helper function to \code{\link{make_design}} for constructing a design matrix
from \code{onsets}, \code{nTime}, and \code{TR}.
}
\keyword{internal}
