% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{make_formula}
\alias{make_formula}
\title{Make Formula}
\usage{
make_formula(beta_names, repl_names, hyper_initial = NULL)
}
\arguments{
\item{beta_names}{char vector of the names of each bbeta object in the environment}

\item{repl_names}{char vector of the names of each replicate object in the environment}

\item{hyper_initial}{Optional vector of initial values for hyperparameters of each latent field OR a list with each element corresponding to one column of the X matrix}
}
\value{
A formula representing the Bayesian GLM to be passed to \code{inla()}
}
\description{
Make Formula
}
\keyword{internal}
