% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_HRFs.R
\name{make_HRFs}
\alias{make_HRFs}
\title{}
\usage{
make_HRFs(onsets, TR, duration, downsample = 100)
}
\arguments{
\item{onsets}{}

\item{TR}{}

\item{duration}{}

\item{downsample}{}
}
\value{

}
\description{

}
