% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{make_mask}
\alias{make_mask}
\title{Create a mask based on vertices that are invalid}
\usage{
make_mask(data)
}
\arguments{
\item{data}{A list of sessions, where each session is a list with elements
BOLD, design and nuisance. See \code{?create.session} and \code{?is.session}
for more details.}
}
\value{
A logical vector indicating valid vertices, or NULL if there were no invalid vertices
}
\description{
Create a mask based on vertices that are invalid
}
