% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mesh.R, R/make_mesh2.R
\name{make_mesh}
\alias{make_mesh}
\title{Make Mesh}
\usage{
make_mesh(faces, verts, mask = NULL)

make_mesh(faces, verts, mask = NULL)
}
\arguments{
\item{faces}{Matrix of faces}

\item{verts}{Matrix of vertices}

\item{mask}{Mask to be applied to mesh. If none given, full mesh is used.}

\item{mask}{}
}
\value{
Triangular mesh from matrices and vertices
}
\description{
Make triangular mesh from faces and vertices. If mask is given, then mesh is masked.
}
