% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_templateICA.R
\name{make_Pmat}
\alias{make_Pmat}
\title{Creates permutation matrix P to regroup elements of s by locations instead of by ICs}
\usage{
make_Pmat(Q, V)
}
\arguments{
\item{Q}{The number of template ICs}

\item{V}{The number of spatial locations}
}
\value{
P Permutation matrix size QVxQV
}
\description{
Creates permutation matrix P to regroup elements of s by locations instead of by ICs
}
\details{
If s=(s1,...,sQ) is grouped by ICs 1,...Q, then Ps=(s(1),...,s(V)) is grouped by locations 1,...,V
}
