% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_utils.R
\name{make_Q}
\alias{make_Q}
\title{Make the full SPDE precision based on theta, the spde, and the number of sessions}
\usage{
make_Q(theta, spde, n_sess)
}
\arguments{
\item{theta}{the hyperparameter theta vector of length K * 2 + 1, where the
first K elements are the kappas, the next K elements are the phis, and the
last element (unused here) corresponds to sigma2}

\item{spde}{a list containing three spde elements: Cmat, Gmat, and GtCinvG}

\item{n_sess}{The integer number of sessions}
}
\value{
An SPDE prior matrix
}
\description{
Make the full SPDE precision based on theta, the spde, and the number of sessions
}
\keyword{internal}
