% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bayesglm_utils.R
\name{make_replicates}
\alias{make_replicates}
\title{Make replicates}
\usage{
make_replicates(nSess, field_names, spatial)
}
\arguments{
\item{nSess}{The number of sessions sharing hyperparameters (can be different fields)}

\item{field_names}{Vector of names for each field}

\item{spatial}{Spatial info}
}
\value{
replicates vector and betas for sessions
}
\description{
beta and repl vectors are of length \eqn{nMesh \times nSess \times n_field}.
The ith repl vector is an indicator vector for the cells corresponding to the ith column of x.
The ith beta vector contains data indices (e.g. 1,...,V) in the cells corresponding to the ith column of x.
}
\keyword{internal}
