% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_est_resid_var_pw.R
\name{make_sqrtInv_all}
\alias{make_sqrtInv_all}
\title{Make \code{sqrtInv_all}}
\usage{
make_sqrtInv_all(nT, nV, do_pw, n_threads, ar_order, AR_coefs_avg, var_avg)
}
\arguments{
\item{nT, nV, do_pw, n_threads, ar_order, AR_coefs_avg, var_avg}{See \code{\link{GLM_est_resid_var_pw}}.}
}
\value{
\code{sqrtInv_all}
}
\description{
Make \code{sqrtInv_all} for prewhitening
}
\keyword{internal}
