% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_templateICA_mesh.R
\name{make_templateICA_mesh}
\alias{make_templateICA_mesh}
\title{}
\usage{
make_templateICA_mesh(mask = NULL, vertices = NULL, faces = NULL, locs = NULL)
}
\arguments{
\item{mask}{}

\item{vertices}{}

\item{faces}{}

\item{locs}{}
}
\value{

}
\description{

}
\note{

}
