% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary.R
\name{mask_with_boundary}
\alias{mask_with_boundary}
\title{}
\usage{
mask_with_boundary(
  vertices,
  faces,
  mask,
  width1 = 4,
  k1 = 2,
  width2 = 6,
  k2 = 3
)
}
\arguments{
\item{faces}{}

\item{mask}{}

\item{width1, width2}{}

\item{k1, k2}{}
}
\value{

}
\description{

}
\details{

}
