% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{match_input}
\alias{match_input}
\title{}
\usage{
match_input(
  user,
  expected,
  fail_action = c("stop", "warning", "message", "nothing"),
  user_value_label = NULL
)
}
\arguments{
\item{user}{}

\item{expected}{}

\item{fail_action}{}

\item{user_value_label}{}
}
\value{

}
\description{

}
\keyword{internal}
