% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{max_threads_Param}
\alias{max_threads_Param}
\title{max_threads}
\arguments{
\item{max_threads}{The maximum number of threads to use in the inla-program
for model estimation. \code{0} (default) will use the maximum number of
threads allowed by the system.}
}
\description{
max_threads
}
