% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{mean_var_Tol_Param}
\alias{mean_var_Tol_Param}
\title{mean and variance tolerance}
\arguments{
\item{meanTol, varTol}{Tolerance for mean and variance of each data location.
Locations which do not meet these thresholds are masked out of the analysis.
Default: \code{1e-6} for both.}
}
\description{
mean and variance tolerance
}
