% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{melt_mat2}
\alias{melt_mat2}
\title{Melt list of matrices}
\usage{
melt_mat2(X_list)
}
\arguments{
\item{X_list}{list of matrices}
}
\value{
A data.frame
}
\description{
See \code{melt_mat}
}
\keyword{internal}
