% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{mesh_Param_either}
\alias{mesh_Param_either}
\title{mesh: either}
\arguments{
\item{mesh}{An \code{"inla.mesh"} object (see \code{\link{make_mesh}} for
surface data) or \code{"BayesfMRI.spde"} object (see
\code{\link{create_spde_vol3D}} for subcortical data).}
}
\description{
mesh: either
}
