% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{mesh_Param_inla}
\alias{mesh_Param_inla}
\title{mesh: INLA only}
\arguments{
\item{mesh}{An \code{"inla.mesh"} object (see \code{\link{make_mesh}} for
surface data).}
}
\description{
mesh: INLA only
}
