% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mesh.R
\name{mesh_to_surf}
\alias{mesh_to_surf}
\title{Convert mesh to surface.}
\usage{
mesh_to_surf(mesh, hemisphere = NULL)
}
\arguments{
\item{mesh}{The \code{"BfMRI.mesh"} object}

\item{hemisphere}{The hemisphere (\code{"left"} or \code{"right"}), if known.
Default: \code{NULL} (unknown).}
}
\value{
The \code{"surf"} object.
}
\description{
Convert \code{"BfMRI.mesh"} object to \code{"surf"} object.
}
\keyword{internal}
