% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiGLM.R
\name{multiGLM}
\alias{multiGLM}
\title{multiGLM}
\usage{
multiGLM(
  BOLD,
  design,
  nuisance = NULL,
  scale_BOLD = c("auto", "mean", "sd", "none"),
  verbose = 1,
  meanTol = 1e-06,
  varTol = 1e-06
)
}
\arguments{
\item{BOLD, design, nuisance}{Session-length list of numeric matrices/arrays,
each with volumes along the first dimension.}

\item{scale_BOLD}{Controls scaling the BOLD response at each location.
\describe{
\item{"auto":}{   (default) Use \code{"mean"} scaling, except if
demeaned data is detected (any location's mean < 1), use \code{"sd"}
scaling.}
\item{"mean":}{   Scale the data to percent local signal change.}
\item{"sd":}{   Scale the data by local standard deviation.}
\item{"none":}{   Center the data but do not scale it.}
}}

\item{verbose}{\code{1} (default) to print occasional updates during model
computation; \code{2} for occasional updates as well as running INLA in
verbose mode (if \code{Bayes}), or \code{0} for no printed updates.}

\item{meanTol, varTol}{Tolerance for mean, variance and SNR of each data location.
Locations which do not meet these thresholds are masked out of the analysis.
Default: \code{1e-6} for mean and variance, \code{50} for SNR.}

\item{spatial}{Gives the spatial information. For surface data, a list of two:
\describe{
\item{surf}{A list of two: \code{vertices} \eqn{V \times 3} numeric matrix of vertex locations in XYZ coordinate space, and \code{faces}, \eqn{F \times 3} matrix of positive integers defining the triangular faces.}
\item{mask}{Mask of locations with valid data.}
}
For voxel data, a list of six:
\describe{
\item{label}{3D array of labeled locations to include in the model.}
\item{trans_mat}{Projection matrix to convert voxel indices to XYZ position. Can be \code{NULL}.}
\item{trans_units}{XYZ units. Can be \code{NULL}.}
\item{nbhd_order}{See documentation for \code{\link{BayesGLM_cifti}}.}
\item{buffer}{See documentation for \code{\link{BayesGLM_cifti}}.}
}}
}
\value{
A \code{"CompareGLM"} object: a list with elements
\describe{
\item{field_estimates}{The estimated coefficients for the Bayesian model.}
\item{mask}{A mask of \code{mesh_orig} indicating the locations inside \code{mesh}.}
\item{design}{The design matrix, after centering and scaling, but before any nuisance regression or prewhitening.}
\item{field_names}{The names of the fields.}
\item{session_names}{The names of the sessions.}
\item{hyperpar_posteriors}{Hyperparameter posterior densities.}
\item{theta_estimates}{Theta estimates from the Bayesian model.}
\item{posterior_Sig_inv}{For joint group modeling.}
\item{mu_theta}{For joint group modeling.}
\item{Q_theta}{For joint group modeling.}
\item{y}{For joint group modeling: The BOLD data after any centering, scaling, nuisance regression, or prewhitening.}
\item{X}{For joint group modeling: The design matrix after any centering, scaling, nuisance regression, or prewhitening.}
\item{prewhiten_info}{Vectors of values across locations: \code{phi} (AR coefficients averaged across sessions), \code{sigma_sq} (residual variance averaged across sessions), and AIC (the maximum across sessions).}
\item{call}{match.call() for this function call.}
}
}
\description{
Performs classical GLM for task fMRI activation, comparing multiple designs
}
\section{INLA Requirement}{

This function requires the \code{INLA} package, which is not a CRAN package.
See \url{https://www.r-inla.org/download-install} for easy installation instructions.
}

