% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{n_threads_Param}
\alias{n_threads_Param}
\title{n_threads}
\arguments{
\item{n_threads}{The maximum number of threads to use for parallel
computations: prewhitening parameter estimation, and the inla-program model
estimation. Default: \code{4}. Note that parallel prewhitening requires the
\code{parallel} package.}
}
\description{
n_threads
}
