% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{nbhd_order_Param}
\alias{nbhd_order_Param}
\title{nbhd_order}
\arguments{
\item{nbhd_order}{For volumetric model. What order neighborhood around data
locations to keep? \code{0} for no neighbors, \code{1} for 1st-order
neighbors, \code{2} for 1st- and 2nd-order neighbors, etc. Smaller values
will provide greater computational efficiency at the cost of higher variance
around the edge of the data.}
}
\description{
nbhd_order
}
