% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_utils.R
\name{neg_kappa_fn}
\alias{neg_kappa_fn}
\title{The negative of the objective function for kappa}
\usage{
neg_kappa_fn(kappa2, spde, phi, Sigma, mu)
}
\arguments{
\item{kappa2}{scalar}

\item{spde}{an spde object}

\item{phi}{scalar}

\item{Sigma}{dgCMatrix}

\item{mu}{dgeMatrix}
}
\value{
a scalar
}
\description{
The negative of the objective function for kappa
}
\keyword{internal}
