% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_utils.R
\name{neg_kappa_fn2}
\alias{neg_kappa_fn2}
\title{The negative of the objective function for kappa - Sig_inv}
\usage{
neg_kappa_fn2(kappa2, spde, phi, P, mu, Vh)
}
\arguments{
\item{kappa2}{scalar}

\item{spde}{an spde object}

\item{phi}{scalar}

\item{P}{Matrix of dimension nk by N_s found by \code{inla.qsolve(Sig_inv,Vh)}}

\item{mu}{dgeMatrix}
}
\value{
a scalar
}
\description{
The negative of the objective function for kappa - Sig_inv
}
\keyword{internal}
