% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_utils.R
\name{neg_kappa_fn4}
\alias{neg_kappa_fn4}
\title{Streamlined negative objective function for kappa2 using precompiled values}
\usage{
neg_kappa_fn4(kappa2, spde, a_star, b_star, n_sess)
}
\arguments{
\item{kappa2}{scalar}

\item{spde}{SPDE prior matrix}

\item{a_star}{precomputed coefficient (scalar)}

\item{b_star}{precomputed coefficient (scalar)}

\item{n_sess}{number of sessions (scalar)}
}
\value{
scalar output of the negative objective function
}
\description{
Streamlined negative objective function for kappa2 using precompiled values
}
\keyword{internal}
