% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_regress.R
\name{nuisance_regress}
\alias{nuisance_regress}
\title{Regresses nuisance parameters from response or predictor variables}
\usage{
nuisance_regress(Y, Z)
}
\arguments{
\item{Y}{a T by M matrix representing the response or predictor variables from which to regress nuisance.}

\item{Z}{a T by J matrix where J is the number of nuisance variables. It contains of nuisance regressors.}
}
\value{
A T by M matrix representing the residuals of Y after regressing out Z
}
\description{
Regresses nuisance parameters from response or predictor variables
}
