% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_data.R
\name{organize_data}
\alias{organize_data}
\title{}
\usage{
organize_data(y, X, transpose = TRUE)
}
\arguments{
\item{y}{}

\item{X}{}

\item{transpose}{}
}
\value{
\verb{}\verb{}
}
\description{

}
\details{
\verb{}
\verb{}
\verb{}
}
