% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_data.R
\name{organize_data_pw}
\alias{organize_data_pw}
\title{}
\usage{
organize_data_pw(y, X, transpose = TRUE)
}
\arguments{
\item{y}{}

\item{X}{}

\item{transpose}{}
}
\value{

}
\description{

}
\details{

}
