% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.act_BGLM}
\alias{plot.act_BGLM}
\title{S3 method: use \code{\link[ciftiTools]{view_xifti}} to plot a \code{"act_BGLM"} object}
\usage{
\method{plot}{act_BGLM}(x, idx = NULL, title = NULL, session = NULL, ...)
}
\arguments{
\item{x}{An object of class "act_BGLM"}

\item{idx}{Which field should be plotted? Give the numeric indices or the
names. \code{NULL} (default) will show all fields. This argument overrides
the \code{idx} argument to \code{\link[ciftiTools]{view_xifti}}.}

\item{title}{If NULL, the field names associated with idx will be used.}

\item{session}{Which session should be plotted? \code{NULL} (default) will
use the first.}

\item{...}{Additional arguments to \code{\link[ciftiTools]{view_xifti}}}
}
\value{
Result of the call to \code{ciftiTools::view_cifti_surface}.
}
\description{
S3 method: use \code{\link[ciftiTools]{view_xifti}} to plot a \code{"act_BGLM"} object
}
