% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.BayesGLM}
\alias{plot.BayesGLM}
\title{S3 method: use \code{plot_BayesGLM_slice} to plot a \code{"BayesGLM"} object}
\usage{
\method{plot}{BayesGLM}(x, ...)
}
\arguments{
\item{x}{An object of class "BayesGLM"}

\item{...}{Additional arguments to \code{\link{plot_BayesGLM_slice}}}
}
\description{
S3 method: use \code{plot_BayesGLM_slice} to plot a \code{"BayesGLM"} object
}
