% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.BayesGLM2_cifti}
\alias{plot.BayesGLM2_cifti}
\title{S3 method: use \code{\link[ciftiTools]{view_xifti_surface}} to plot a \code{"BayesGLM2_cifti"} object}
\usage{
\method{plot}{BayesGLM2_cifti}(x, idx = NULL, what = c("contrasts", "activations"), ...)
}
\arguments{
\item{x}{An object of class "BayesGLM2_cifti"}

\item{idx}{Which contrast should be plotted? Give the numeric index.
\code{NULL} (default) will show all contrasts. This argument overrides
the \code{idx} argument to \code{\link[ciftiTools]{view_xifti_surface}}.}

\item{what}{Estimates of the \code{"contrasts"} (default), or their
thresholded \code{"activations"}.}

\item{...}{Additional arguments to \code{\link[ciftiTools]{view_xifti_surface}}}
}
\value{
Result of the call to \code{ciftiTools::view_cifti_surface}.
}
\description{
S3 method: use \code{\link[ciftiTools]{view_xifti_surface}} to plot a \code{"BayesGLM2_cifti"} object
}
