% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BayesGLM_2D.R
\name{plot.BayesGLM_2d}
\alias{plot.BayesGLM_2d}
\title{Plot results from a BayesfMRI object for 2D Analyses}
\usage{
\method{plot}{BayesGLM_2d}(BayesGLM_object, mask, session_name = NULL, zlim = NULL)
}
\arguments{
\item{BayesGLM_object}{An object of class "BayesfMRI"}

\item{mask}{The mask used to convert from cortical surface to volumetric data}

\item{session_name}{The name of the session to plot the results from (defaults to the first session)}

\item{zlim}{The color limits for plotting the coefficient values. Defaults to the minimum and maximum of the point estimates}
}
\value{
A ggplot2 object
}
\description{
Plot results from a BayesfMRI object for 2D Analyses
}
