% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_design.R
\name{plot.BfMRI_design}
\alias{plot.BfMRI_design}
\title{S3 method: use \code{\link[ciftiTools]{view_xifti}} to plot a \code{"BGLM"} object}
\usage{
\method{plot}{BfMRI_design}(x, ...)
}
\arguments{
\item{x}{An object of class "BfMRI_design".}

\item{...}{Additional arguments to \code{\link{plot_design}}.}
}
\value{
Result of the call to \code{\link{plot_design}}
}
\description{
S3 method: use \code{\link[ciftiTools]{view_xifti}} to plot a \code{"BGLM"} object
}
