% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.prev_BayesGLM_cifti}
\alias{plot.prev_BayesGLM_cifti}
\title{S3 method: use \code{\link[ciftiTools]{view_xifti}} to plot a \code{"prev_BayesGLM_cifti"} object}
\usage{
\method{plot}{prev_BayesGLM_cifti}(
  x,
  idx = NULL,
  session = NULL,
  drop_zeros = NULL,
  colors = "plasma",
  zlim = c(round(1/x$n_results - 0.005, 2), 1),
  ...
)
}
\arguments{
\item{x}{An object of class "prev_BayesGLM_cifti"}

\item{idx}{Which task should be plotted? Give the numeric indices or the
names. \code{NULL} (default) will show all tasks. This argument overrides
the \code{idx} argument to \code{\link[ciftiTools]{view_xifti}}.}

\item{session}{Which session should be plotted? \code{NULL} (default) will
use the first.}

\item{drop_zeros}{Color locations without any activation across all results
(zero prevalence) the same color as the medial wall? Default: \code{NULL} to
drop the zeros if only one \code{idx} is being plotted.}

\item{colors, zlim}{See \code{\link[ciftiTools]{view_xifti}}. Here, the defaults
are overrided to use the Viridis \code{"plasma"} color scale between
\code{1/nA} and 1, where \code{nA} is the number of results in \code{x}.}

\item{...}{Additional arguments to \code{\link[ciftiTools]{view_xifti}}}
}
\value{
Result of the call to \code{ciftiTools::view_cifti_surface}.
}
\description{
S3 method: use \code{\link[ciftiTools]{view_xifti}} to plot a \code{"prev_BayesGLM_cifti"} object
}
