% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_BayesGLM_slice}
\alias{plot_BayesGLM_slice}
\title{}
\usage{
plot_BayesGLM_slice(BayesGLM_object, session_name = NULL, zlim = NULL)
}
\arguments{
\item{BayesGLM_object}{}

\item{session_name}{}

\item{zlim}{}
}
\value{

}
\description{

}
