% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_slice}
\alias{plot_slice}
\title{Plot a 2D image using ggplot}
\usage{
plot_slice(X, color_palette = NULL, zlim = NULL)
}
\arguments{
\item{X}{A matrix or list of matrices to be plotted}

\item{color_palette}{A color palette as a data frame with two columns. The
first column should contain colors (in HEX, character, or other format),
and the second column should contain locations for color breaks on the unit
interval describing the locations for color switching. This will default to
the output from \code{ROY_BIG_BL} palette from the \code{ciftiTools}
package.}

\item{zlim}{A vector of length 2 describing the endpoints of the color
palette.}
}
\value{
A ggplot graphical object.
}
\description{
Plot a 2D image using ggplot
}
