% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_utils.R
\name{prep_kappa2_optim}
\alias{prep_kappa2_optim}
\title{Find values for coefficients used in objective function for kappa2}
\usage{
prep_kappa2_optim(spde, mu, phi, P, vh)
}
\arguments{
\item{spde}{an spde object}

\item{mu}{the mean}

\item{phi}{scale parameter}

\item{P}{Matrix of dimension nk by N_s found by \code{solve(Sig_inv,Vh)}}

\item{vh}{random matrix of -1 and 1 of dim \code{dim(P)}}
}
\value{
a list with two elements, which are precomputed coefficients for the
optimization function
}
\description{
Find values for coefficients used in objective function for kappa2
}
\keyword{internal}
