% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prewhitening.R
\name{prewhiten.v}
\alias{prewhiten.v}
\title{}
\usage{
prewhiten.v(AR_coeffs, ntime, AR_var = 1)
}
\arguments{
\item{AR_coeffs}{}

\item{ntime}{}

\item{AR_var}{}
}
\value{

}
\description{

}
\examples{
\dontrun{
x <- arima.sim(list(ar = c(.3,.2,.1)), n = 100)
acf(x)
ar_model <- ar.yw(x,order.max = 6, aic = F)
pw_mat <- prewhiten.v(AR_coeffs = ar_model$ar,
                      ntime = length(x),
                      AR_var = ar_model$var.pred)
y <- as.vector(pw_mat \%*\% as.matrix(x))
acf(y)
plot(x, type = 'l')
lines(y, col = 'red')
}
}
