% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prewhitening.R
\name{prewhiten_prep}
\alias{prewhiten_prep}
\title{Estimate residual autocorrelation for prewhitening}
\usage{
prewhiten_prep(
  data,
  mask = NULL,
  scale_BOLD = TRUE,
  scale_design = TRUE,
  ar_order = 6
)
}
\arguments{
\item{data}{List of sessions (see \code{is.session} and \code{is.session_pw})}

\item{mask}{(Optional) A length \eqn{V} logical vector indicating if each
vertex is to be included.}

\item{scale_BOLD}{(logical) Should the BOLD response be scaled? (Default is TRUE)}

\item{scale_design}{(logical) Should the design matrix be scaled? (Default is TRUE)}

\item{ar_order}{Order of the AR used to prewhiten the data at each location}
}
\value{
The prewhitened data (in a list), the smoothed, averaged AR
coefficient estimates used in the prewhitening, the smoothed, average
residual variance after prewhitening, and the value given for \code{ar_order}.
}
\description{
Estimate residual autocorrelation for prewhitening
}
