% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prewhitening.R
\name{prewhiten_smooth}
\alias{prewhiten_smooth}
\title{}
\usage{
prewhiten_smooth(prewhiten_result, ar_smooth = NULL, cifti_data, hemisphere)
}
\arguments{
\item{prewhiten_result}{}

\item{ar_smooth}{}

\item{cifti_data}{}

\item{hemisphere}{}
}
\description{

}
