% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prewhitening.R
\name{pw_estimate}
\alias{pw_estimate}
\title{Estimate residual autocorrelation for prewhitening}
\usage{
pw_estimate(resids, ar_order)
}
\arguments{
\item{ar_order}{Order of the AR used to prewhiten the data at each location}

\item{data}{Estimated residuals}
}
\value{
Estimaed AR coefficients and residual variance at every vertex
}
\description{
Estimate residual autocorrelation for prewhitening
}
