% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary.R
\name{radial_order}
\alias{radial_order}
\title{Order Vertices on Circular Manifold}
\usage{
radial_order(verts)
}
\value{
Index ordering of \code{verts}
}
\description{
Order vertices on circular manifold by radians (after 2D CMDS projection).
}
