% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_from_design.R
\name{remove_from_design}
\alias{remove_from_design}
\title{Remove from design}
\usage{
remove_from_design(design, rm_name = NULL, rm_pattern = ".*_dHRF|.*_ddHRF")
}
\arguments{
\item{design}{A \code{"BayesfMRI_design"} object from \code{\link{make_design}}.}

\item{rm_name, rm_pattern}{The fields to remove. Either provide the names to
\code{rm_name} or the patterns to match with \code{\link[base]{grepl}} to
\code{rm_pattern}.

By default, \code{rm_pattern} specifies the removal of fields ending in
"_dHRF" or "_ddHRF". If \code{design} was constructed from \code{onsets},
then these are the HRF derivatives.}
}
\value{
A list of two: the modified \code{"BayesfMRI_design"} object, and
the removed columns.
}
\description{
Remove field from result of \code{\link{make_design}}.
}
