% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs_BayesGLM.R
\name{resamp_res_Param_BayesGLM}
\alias{resamp_res_Param_BayesGLM}
\title{resamp_res}
\arguments{
\item{resamp_res}{For cortex spatial model. The number of vertices to which
each cortical surface should be resampled, or \code{NULL} to not resample.

For computational feasibility, a value of \code{10000} (default) or lower is
recommended for Bayesian spatial modeling. If \code{Bayes=FALSE},
\code{resamp_res} can be set to \code{NULL} for full-resolution classical
modeling.}
}
\description{
resamp_res
}
