% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_mask.R
\name{retro_mask_BGLM}
\alias{retro_mask_BGLM}
\title{Retroactively mask locations from BayesGLM result.}
\usage{
retro_mask_BGLM(x, mask)
}
\arguments{
\item{x}{The BayesGLM result}

\item{mask}{The mask to be applied to \code{x} (on top of any masks already
applied to it.)}
}
\value{
The masked result
}
\description{
Work in progress.
}
\keyword{internal}
