% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_mask.R
\name{retro_mask_fit_bglm}
\alias{retro_mask_fit_bglm}
\title{Retroactively mask locations from fit_bglm result.}
\usage{
retro_mask_fit_bglm(x, mask)
}
\arguments{
\item{x}{The \code{"fit_bglm"} result}

\item{mask}{The mask to be applied to \code{x}. It's relative to the full
mesh/voxel array, not \code{x}.}
}
\value{
The masked result
}
\description{
Retroactively mask locations from fit_bglm result.
}
\keyword{internal}
