% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{return_INLA_Param}
\alias{return_INLA_Param}
\title{return_INLA}
\arguments{
\item{return_INLA}{Return the INLA model object? (It can be large.) Use
\code{"trimmed"} (default) to return only the more relevant results,
\code{"minimal"} to return just enough for \code{\link{id_activations}}, or
\code{"full"} to return the full output of \code{inla}.}
}
\description{
return_INLA
}
