% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{return_INLA_result_Param_FALSE}
\alias{return_INLA_result_Param_FALSE}
\title{return_INLA_result: FALSE}
\arguments{
\item{return_INLA_result}{Return the INLA model object? (It can be large.)
Default: \code{FALSE}. Required for running \code{id_activations}
after, but not for running BayesGLM_joint after to get
posterior quantities of group means or contrasts.}
}
\description{
return_INLA_result: FALSE
}
