% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_BOLD.R
\name{scale_BOLD}
\alias{scale_BOLD}
\title{Scale the BOLD timeseries}
\usage{
scale_BOLD(BOLD, scale = c("mean", "sd", "none"), v_means = NULL)
}
\arguments{
\item{BOLD}{fMRI data as a locations by time (\eqn{V \times T}) numeric
matrix.}

\item{scale}{Option for scaling the BOLD response.}

\item{v_means}{Original means of the BOLD data. ONLY provide if data has already been centered.

\if{html}{\out{<div class="sourceCode">}}\preformatted{\code{"auto"} (default) will use \code{"mean"} scaling except if demeaned
}\if{html}{\out{</div>}}

data is detected (if any mean is less than one), in which case \code{"sd"}
scaling will be used instead.

\if{html}{\out{<div class="sourceCode">}}\preformatted{\code{"mean"} scaling will scale the data to percent local signal change.

\code{"sd"} scaling will scale the data by local standard deviation.

\code{"none"} will only center the data, not scale it.
}\if{html}{\out{</div>}}}
}
\value{
Scale to units of percent local signal change and centers
}
\description{
Scale the BOLD timeseries
}
