% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{scale_design_Param}
\alias{scale_design_Param}
\title{scale_design}
\arguments{
\item{scale_design}{Scale the design matrix by dividing each column by its
maximum and then subtracting the mean? Default: \code{TRUE}. If
\code{FALSE}, the design matrix is centered but not scaled.}
}
\description{
scale_design
}
