% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_timeseries.R
\name{scale_timeseries}
\alias{scale_timeseries}
\title{}
\usage{
scale_timeseries(BOLD, scale = TRUE, transpose = TRUE)
}
\arguments{
\item{BOLD}{}

\item{scale}{}

\item{transpose}{}
}
\value{

}
\description{

}
