% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs_BayesGLM.R
\name{scrub_Param_BayesGLM}
\alias{scrub_Param_BayesGLM}
\title{scrub}
\arguments{
\item{scrub}{(Optional) A \eqn{T \times N_{scrub}} matrix of spike regressors
(one 1 value at the timepoint to scrub, and 0 for all other values), or a
logical vector indicating the timepoints to scrub (\code{TRUE} to scrub, and
\code{FALSE} to keep). For multi-session data, a session-length list of
such matrices or logical vectors.

The spike regressors will be included in the nuisance
regression, and afterwards the timepoints indicated in \code{scrub} will be
removed from the BOLD data and design matrix.}
}
\description{
scrub
}
