% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{session_names_Param}
\alias{session_names_Param}
\title{session_names}
\arguments{
\item{session_names}{(Optional, and only relevant for multi-session modeling)
Names of each session. Default: \code{NULL}. Will overwrite the names of the
\code{BOLD} data, if both are provided.}
}
\description{
session_names
}
