% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_slice_data.R
\name{simulate_slice_data}
\alias{simulate_slice_data}
\title{}
\usage{
simulate_slice_data(
  num_sessions = 1,
  num_tasks = 2,
  active_centers = NULL,
  active_size = NULL,
  beta_weights = NULL,
  vary_active = T,
  num_time = 200,
  binary_template = NULL
)
}
\arguments{
\item{num_sessions}{}

\item{num_tasks}{}

\item{active_centers}{}

\item{active_size}{}

\item{beta_weights}{}

\item{vary_active}{}

\item{num_time}{}

\item{binary_template}{}
}
\value{

}
\description{

}
