% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_and_PW.R
\name{sparse_and_PW}
\alias{sparse_and_PW}
\title{Organize data for Bayesian GLM}
\usage{
sparse_and_PW(
  BOLD,
  design,
  spatial,
  spatial_type,
  field_names,
  design_type,
  valid_cols,
  nT,
  sqrtInv_all
)
}
\arguments{
\item{BOLD, design, spatial, spatial_type}{See \code{BayesGLM0}.}

\item{field_names, design_type}{See \code{BayesGLM0}.}

\item{valid_cols, nT, sqrtInv_all}{See \code{BayesGLM0}.}
}
\value{
A list containing fields \code{y} and \code{A} (see Details)
}
\description{
Transforms the usual TxV BOLD data matrix Y into vector form, and
the usual TxK design matrix X into big sparse matrix form for use in
Bayesian GLM.
}
\details{
The Bayesian GLM requires \code{y} (a vector of length TV containing the BOLD data)
and \code{X_k} (a sparse TVxV matrix corresponding to the kth field regressor) for each field k.
The design matrices are combined as \code{A=cbind(X_1,...,X_K)}.

The Bayesian GLM requires \code{y} (a vector of length TV containing the BOLD data)
and \code{X_k} (a sparse TVxV matrix corresponding to the kth field regressor) for each field k.
The design matrices are combined as \code{A=cbind(X_1,...,X_K)}.
}
\keyword{internal}
