% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPDE_from_mesh.R
\name{SPDE_from_mesh}
\alias{SPDE_from_mesh}
\title{SPDE from mesh model}
\usage{
SPDE_from_mesh(spatial)
}
\arguments{
\item{spatial}{See \code{BayesGLM}.}
}
\value{
List
}
\description{
SPDE from mesh model
}
\keyword{internal}
