% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPDE_from_mesh.R
\name{SPDE_from_surf}
\alias{SPDE_from_surf}
\title{SPDE from mesh model}
\usage{
SPDE_from_surf(spatial, logkappa = NULL, logtau = NULL)
}
\arguments{
\item{spatial}{See \code{BayesGLM}.}

\item{logkappa, logtau}{vector of min, max and initial value for prior on log kappa and log tau. Min and max are extreme quantiles, not hard constraints.}
}
\value{
List
}
\description{
SPDE from mesh model
}
\keyword{internal}
