% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPDE_from_voxel.R
\name{SPDE_from_voxel}
\alias{SPDE_from_voxel}
\title{SPDE from voxel model}
\usage{
SPDE_from_voxel(spatial, qc_mask = NULL, logkappa = NULL, logtau = NULL)
}
\arguments{
\item{spatial}{See \code{BayesGLM}.}

\item{qc_mask}{QC mask.}

\item{logkappa, logtau}{vector of min, max and initial value for prior on log kappa and log tau. Min and max are extreme quantiles, not hard constraints.}
}
\value{
List: \code{spde} and \code{spatial}.
}
\description{
SPDE from voxel model
}
\keyword{internal}
