% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_utils.R
\name{spde_Q_phi}
\alias{spde_Q_phi}
\title{Calculate the SPDE covariance}
\usage{
spde_Q_phi(kappa2, phi, spde)
}
\arguments{
\item{kappa2}{A scalar}

\item{phi}{A scalar}

\item{spde}{An object containing the information about the
mesh structure for the SPDE prior}
}
\value{
The SPDE prior matrix
}
\description{
Calculate the SPDE covariance
}
\keyword{internal}
