% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_sessions.R
\name{summarize_sessions}
\alias{summarize_sessions}
\title{Computes Bayesian GLM posterior means, quantiles and excursions sets (areas of activation) for summaries or contrasts across sessions within a single multi-session model}
\usage{
summarize_sessions(
  result,
  contrasts = NULL,
  quantiles = NULL,
  excursion_type,
  gamma = 0,
  alpha = 0.05,
  nsamp_theta = 50,
  nsamp_beta = 100
)
}
\arguments{
\item{result}{An object of class BayesGLM containing the results of a multi-session analysis. Must contain INLA_result!  (See \code{help(BayesGLM)}.)}

\item{contrasts}{(Optional) A list of vectors, each length J*K, specifying the contrast(s) of interest across sessions, where J is the number of sessions analyzed and K is the number of tasks.  See Details for more information. Default is to compute the average for each task across sessions.}

\item{quantiles}{(Optional) Vector of posterior quantiles to return in addition to the posterior mean}

\item{excursion_type}{(For inference only) The type of excursion function for the contrast (">", "<", "!="), or a vector thereof (each element corresponding to one contrast).  If none provided, no inference performed.}

\item{gamma}{(For inference only) Activation threshold for the excursion set, or a vector thereof (each element corresponding to one contrast).}

\item{alpha}{(For inference only) Significance level for activation for the excursion set}

\item{nsamp_theta}{Number of theta values to sample from posterior. Default is 50.}

\item{nsamp_beta}{Number of beta vectors to sample conditional on each theta value sampled. Default is 100.}
}
\value{
A list containing...
}
\description{
Computes Bayesian GLM posterior means, quantiles and excursions sets (areas of activation) for summaries or contrasts across sessions within a single multi-session model
}
\details{
Performs posterior inference on summaries or contrasts across the sessions in a multi-session Bayesian GLM.  Let J be the number of sessions analyzed in the
multi-session Bayesian GLM, and let K be the number of tasks.  Computes the posterior mean of each contrast and produces excursions sets representing areas of
activation (above a specified activation threshold gamma) for each contrast map based on the joint (across locations) posterior distribution.

Each contrast vector specifies an across-session summary of interest. The default is one contrast vector for each task representing the across-session average.
Examples of contrast vectors include:
\code{rep(c(1/J, rep(0, J-1)), K)} represents the across-session average for the first task
\code{rep( c(1/J,-1/J,rep(0, K-2)) ,J)} represents the difference between the first two tasks, averaged over all sessions
}
\note{
This function requires the \code{INLA} package, which is not a CRAN package. See \url{http://www.r-inla.org/download} for easy installation instructions.
}
