% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_activations_utils.R
\name{summary.act_BayesGLM}
\alias{summary.act_BayesGLM}
\alias{print.summary.act_BayesGLM}
\alias{print.act_BayesGLM}
\title{Summarize a \code{"act_BayesGLM"} object}
\usage{
\method{summary}{act_BayesGLM}(object, ...)

\method{print}{summary.act_BayesGLM}(x, ...)

\method{print}{act_BayesGLM}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"act_BayesGLM"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"summary.act_BayesGLM"}.}
}
\value{
A \code{"summary.act_BayesGLM"} object, a list summarizing the
properties of \code{object}.

\code{NULL}, invisibly.

\code{NULL}, invisibly.
}
\description{
Summary method for class \code{"act_BayesGLM"}
}
