% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_activations.methods.R
\name{summary.act_BGLM}
\alias{summary.act_BGLM}
\alias{print.summary.act_BGLM}
\alias{print.act_BGLM}
\title{Summarize a \code{"act_BGLM"} object}
\usage{
\method{summary}{act_BGLM}(object, ...)

\method{print}{summary.act_BGLM}(x, ...)

\method{print}{act_BGLM}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"act_BGLM"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"summary.act_BGLM"}.}
}
\value{
A \code{"summary.act_BGLM"} object, a list summarizing the
properties of \code{object}.

\code{NULL}, invisibly.

\code{NULL}, invisibly.
}
\description{
Summary method for class \code{"act_BGLM"}
}
