% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.methods.R
\name{summary.act_BGLM0}
\alias{summary.act_BGLM0}
\alias{print.summary.act_BGLM0}
\alias{print.act_BGLM0}
\title{Summarize a \code{"act_BGLM0"} object}
\usage{
\method{summary}{act_BGLM0}(object, ...)

\method{print}{summary.act_BGLM0}(x, ...)

\method{print}{act_BGLM0}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"act_BGLM0"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"summary.act_BGLM0"}.}
}
\value{
A \code{"summary.act_BGLM0"} object, a list summarizing the
properties of \code{object}.

\code{NULL}, invisibly.

\code{NULL}, invisibly.
}
\description{
Summary method for class \code{"act_BGLM0"}
}
