% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.methods.R
\name{summary.act_fit_bglm}
\alias{summary.act_fit_bglm}
\alias{print.summary.act_fit_bglm}
\alias{print.act_fit_bglm}
\title{Summarize a \code{"act_fit_bglm"} object}
\usage{
\method{summary}{act_fit_bglm}(object, ...)

\method{print}{summary.act_fit_bglm}(x, ...)

\method{print}{act_fit_bglm}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"act_fit_bglm"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"summary.act_fit_bglm"}.}
}
\value{
A \code{"summary.act_fit_bglm"} object, a list summarizing the
properties of \code{object}.

\code{NULL}, invisibly.

\code{NULL}, invisibly.
}
\description{
Summary method for class \code{"act_fit_bglm"}
}
