% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{summary.BayesGLM}
\alias{summary.BayesGLM}
\alias{print.summary.BayesGLM}
\alias{print.BayesGLM}
\title{Summarise BayesGLM objects}
\usage{
\method{summary}{BayesGLM}(object, ...)

\method{print}{summary.BayesGLM}(x, ...)

\method{print}{BayesGLM}(x, ...)
}
\arguments{
\item{object}{an object of class "BayesGLM"}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class "summary.BayesGLM"}
}
\description{
Summary method for class "BayesGLM"
}
