% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM2_utils.R
\name{summary.BayesGLM2}
\alias{summary.BayesGLM2}
\alias{print.summary.BayesGLM2}
\alias{print.BayesGLM2}
\title{Summarize a \code{"BayesGLM2"} object}
\usage{
\method{summary}{BayesGLM2}(object, ...)

\method{print}{summary.BayesGLM2}(x, ...)

\method{print}{BayesGLM2}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"BayesGLM2"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"summary.BayesGLM2"}.}
}
\value{
A \code{"summary.BayesGLM2"} object, a list summarizing the
properties of \code{object}.

\code{NULL}, invisibly.

\code{NULL}, invisibly.
}
\description{
Summary method for class \code{"BayesGLM2"}
}
