% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{summary.BayesGLM_cifti}
\alias{summary.BayesGLM_cifti}
\alias{print.summary.BayesGLM_cifti}
\alias{print.BayesGLM_cifti}
\title{Summarize a \code{"BayesGLM_cifti"} object}
\usage{
\method{summary}{BayesGLM_cifti}(object, ...)

\method{print}{summary.BayesGLM_cifti}(x, ...)

\method{print}{BayesGLM_cifti}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"BayesGLM_cifti"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"summary.BayesGLM_cifti"}.}
}
\value{
A \code{"summary.BayesGLM_cifti"} object, a list summarizing the
properties of \code{object}.

\code{NULL}, invisibly.

\code{NULL}, invisibly.
}
\description{
Summary method for class \code{"BayesGLM_cifti"}
}
