% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{summary.BayesGLMfMRIobj}
\alias{summary.BayesGLMfMRIobj}
\alias{print.summary.BayesGLMfMRIobj}
\alias{print.BayesGLMfMRIobj}
\title{Summarise BayesGLMfMRI objects}
\usage{
\method{summary}{BayesGLMfMRIobj}(object, ...)

\method{print}{summary.BayesGLMfMRIobj}(x, ...)

\method{print}{BayesGLMfMRIobj}(x, ...)
}
\arguments{
\item{object}{an object of class "BayesGLMfMRIobj"}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class "summary.BayesGLMfMRIobj"}
}
\description{
Summary method for class "BayesGLMfMRIobj"
}
