% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{summary.BayesGLMobj}
\alias{summary.BayesGLMobj}
\alias{print.summary.BayesGLMobj}
\alias{print.BayesGLMobj}
\title{Summarise BayesGLM objects}
\usage{
\method{summary}{BayesGLMobj}(object, ...)

\method{print}{summary.BayesGLMobj}(x, ...)

\method{print}{BayesGLMobj}(x, ...)
}
\arguments{
\item{object}{an object of class "BayesGLMobj"}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class "summary.BayesGLMobj"}
}
\description{
Summary method for class "BayesGLMobj"
}
