% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_design.methods.R
\name{summary.BfMRI_design}
\alias{summary.BfMRI_design}
\alias{print.summary.BfMRI_design}
\alias{print.BfMRI_design}
\title{Summarize a \code{"BfMRI_design"} object}
\usage{
\method{summary}{BfMRI_design}(object, ...)

\method{print}{summary.BfMRI_design}(x, ...)

\method{print}{BfMRI_design}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"BfMRI_design"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"summary.BfMRI_design"}.}
}
\value{
A \code{"summary.BfMRI_design"} object, a list summarizing the
properties of \code{object}.

\code{NULL}, invisibly.

\code{NULL}, invisibly.
}
\description{
Summary method for class \code{"BfMRI_design"}
}
