% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGLM.methods.R
\name{summary.BGLM}
\alias{summary.BGLM}
\alias{print.summary.BGLM}
\alias{print.BGLM}
\title{Summarize a \code{"BGLM"} object}
\usage{
\method{summary}{BGLM}(object, ...)

\method{print}{summary.BGLM}(x, ...)

\method{print}{BGLM}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"BGLM"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"summary.BGLM"}.}
}
\value{
A \code{"summary.BGLM"} object, a list summarizing the
properties of \code{object}.

\code{NULL}, invisibly.

\code{NULL}, invisibly.
}
\description{
Summary method for class \code{"BGLM"}
}
