% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGLM0.methods.R
\name{summary.BGLM0}
\alias{summary.BGLM0}
\alias{print.summary.BGLM0}
\alias{print.BGLM0}
\title{Summarize a \code{"BGLM0"} object}
\usage{
\method{summary}{BGLM0}(object, ...)

\method{print}{summary.BGLM0}(x, ...)

\method{print}{BGLM0}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"BGLM0"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"summary.BGLM0"}.}
}
\value{
A \code{"summary.BGLM0"} object, a list summarizing the properties
of \code{object}.

\code{NULL}, invisibly.

\code{NULL}, invisibly.
}
\description{
Summary method for class \code{"BGLM0"}
}
