% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGLM2.methods.R
\name{summary.BGLM02}
\alias{summary.BGLM02}
\alias{print.summary.BGLM02}
\alias{print.BGLM02}
\title{Summarize a \code{"BGLM02"} object}
\usage{
\method{summary}{BGLM02}(object, ...)

\method{print}{summary.BGLM02}(x, ...)

\method{print}{BGLM02}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"BGLM02"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"summary.BGLM02"}.}
}
\value{
A \code{"summary.BGLM02"} object, a list summarizing the
properties of \code{object}.

\code{NULL}, invisibly.

\code{NULL}, invisibly.
}
\description{
Summary method for class \code{"BGLM02"}
}
