% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGLM2.methods.R
\name{summary.BGLM2}
\alias{summary.BGLM2}
\alias{print.summary.BGLM2}
\alias{print.BGLM2}
\title{Summarize a \code{"BGLM2"} object}
\usage{
\method{summary}{BGLM2}(object, ...)

\method{print}{summary.BGLM2}(x, ...)

\method{print}{BGLM2}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"BGLM2"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"summary.BGLM2"}.}
}
\value{
A \code{"summary.BGLM2"} object, a list summarizing the
properties of \code{object}.

\code{NULL}, invisibly.

\code{NULL}, invisibly.
}
\description{
Summary method for class \code{"BGLM2"}
}
