% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGLM0.methods.R
\name{summary.fit_bglm}
\alias{summary.fit_bglm}
\alias{print.summary.fit_bglm}
\alias{print.fit_bglm}
\title{Summarize a \code{"fit_bglm"} object}
\usage{
\method{summary}{fit_bglm}(object, ...)

\method{print}{summary.fit_bglm}(x, ...)

\method{print}{fit_bglm}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"fit_bglm"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"summary.fit_bglm"}.}
}
\value{
A \code{"summary.fit_bglm"} object, a list summarizing the properties
of \code{object}.

\code{NULL}, invisibly.

\code{NULL}, invisibly.
}
\description{
Summary method for class \code{"fit_bglm"}
}
