% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.BayesGLM2.R
\name{summary.fit_bglm2}
\alias{summary.fit_bglm2}
\alias{print.summary.fit_bglm2}
\alias{print.fit_bglm2}
\title{Summarize a \code{"fit_bglm2"} object}
\usage{
\method{summary}{fit_bglm2}(object, ...)

\method{print}{summary.fit_bglm2}(x, ...)

\method{print}{fit_bglm2}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"fit_bglm2"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"summary.fit_bglm2"}.}
}
\value{
A \code{"summary.fit_bglm2"} object, a list summarizing the
properties of \code{object}.

\code{NULL}, invisibly.

\code{NULL}, invisibly.
}
\description{
Summary method for class \code{"fit_bglm2"}
}
