% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.methods.R, R/prevalence_utils.R
\name{summary.prev_BGLM}
\alias{summary.prev_BGLM}
\alias{print.summary.prev_BGLM}
\alias{print.prev_BGLM}
\title{Summarize a \code{"prev_BGLM"} object}
\usage{
\method{summary}{prev_BGLM}(object, ...)

\method{print}{summary.prev_BGLM}(x, ...)

\method{print}{prev_BGLM}(x, ...)

\method{summary}{prev_BGLM}(object, ...)

\method{print}{summary.prev_BGLM}(x, ...)

\method{print}{prev_BGLM}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"prev_BGLM"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"summary.prev_BGLM"}.}
}
\value{
A \code{"summary.prev_BGLM"} object, a list summarizing the
properties of \code{object}.

\code{NULL}, invisibly.

\code{NULL}, invisibly.

A \code{"summary.prev_BGLM"} object, a list summarizing the
properties of \code{object}.

\code{NULL}, invisibly.

\code{NULL}, invisibly.
}
\description{
Summary method for class \code{"prev_BGLM"}

Summary method for class \code{"prev_BGLM"}
}
