% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_prevalence.methods.R,
%   R/act_prevalence_utils.R
\name{summary.prev_BGLM0}
\alias{summary.prev_BGLM0}
\alias{print.summary.prev_BGLM0}
\alias{print.prev_BGLM0}
\title{Summarize a \code{"prev_BGLM0"} object}
\usage{
\method{summary}{prev_BGLM0}(object, ...)

\method{print}{summary.prev_BGLM0}(x, ...)

\method{print}{prev_BGLM0}(x, ...)

\method{summary}{prev_BGLM0}(object, ...)

\method{print}{summary.prev_BGLM0}(x, ...)

\method{print}{prev_BGLM0}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"prev_BGLM0"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"summary.prev_BGLM0"}.}
}
\value{
A \code{"summary.prev_BGLM0"} object, a list summarizing the
properties of \code{object}.

\code{NULL}, invisibly.

\code{NULL}, invisibly.

A \code{"summary.prev_BGLM0"} object, a list summarizing the
properties of \code{object}.

\code{NULL}, invisibly.

\code{NULL}, invisibly.
}
\description{
Summary method for class \code{"prev_BGLM0"}

Summary method for class \code{"prev_BGLM0"}
}
