% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.methods.R, R/prevalence_utils.R
\name{summary.prev_fit_bglm}
\alias{summary.prev_fit_bglm}
\alias{print.summary.prev_fit_bglm}
\alias{print.prev_fit_bglm}
\title{Summarize a \code{"prev_fit_bglm"} object}
\usage{
\method{summary}{prev_fit_bglm}(object, ...)

\method{print}{summary.prev_fit_bglm}(x, ...)

\method{print}{prev_fit_bglm}(x, ...)

\method{summary}{prev_fit_bglm}(object, ...)

\method{print}{summary.prev_fit_bglm}(x, ...)

\method{print}{prev_fit_bglm}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"prev_fit_bglm"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"summary.prev_fit_bglm"}.}
}
\value{
A \code{"summary.prev_fit_bglm"} object, a list summarizing the
properties of \code{object}.

\code{NULL}, invisibly.

\code{NULL}, invisibly.

A \code{"summary.prev_fit_bglm"} object, a list summarizing the
properties of \code{object}.

\code{NULL}, invisibly.

\code{NULL}, invisibly.
}
\description{
Summary method for class \code{"prev_fit_bglm"}

Summary method for class \code{"prev_fit_bglm"}
}
