% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs_BayesGLM.R
\name{surfaces_Param_BayesGLM}
\alias{surfaces_Param_BayesGLM}
\title{surfaces}
\arguments{
\item{surfL, surfR}{For cortex spatial model. Left and right cortex surface
geometry in GIFTI format ("*.surf.gii"). These can be a file path to
a GIFTI file or a \code{"surf"} object from \code{ciftiTools}.

Surfaces can alternatively be provided through the \code{$surf} metadata in
\code{BOLD} if it is \code{"xifti"} data. If neither are provided, by default the
HCP group-average fs_LR inflated surfaces included in \code{ciftiTools} will be
used for the cortex spatial model.}
}
\description{
surfaces
}
