% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateICA.R
\name{templateICA}
\alias{templateICA}
\title{}
\usage{
templateICA(
  template_mean,
  template_var,
  BOLD,
  scale = TRUE,
  mesh = NULL,
  maxQ = NULL,
  common_smoothness = TRUE,
  maxiter = 100,
  epsilon = 0.001,
  verbose = TRUE,
  kappa_init = NULL,
  dim_reduce_flag = TRUE
)
}
\arguments{
\item{template_mean}{}

\item{template_var}{}

\item{BOLD}{}

\item{scale}{}

\item{mesh}{}

\item{maxQ}{}

\item{common_smoothness}{}

\item{maxiter}{}

\item{epsilon}{}

\item{verbose}{}

\item{kappa_init}{}

\item{dim_reduce_flag}{}
}
\value{

}
\description{

}
