% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tile.plot}
\alias{tile.plot}
\title{}
\usage{
tile.plot(
  tile_df,
  col = NULL,
  ncols = NULL,
  main = "",
  zlim = NULL,
  na.color = "grey80"
)
}
\arguments{
\item{tile_df}{}

\item{col}{}

\item{ncols}{}

\item{main}{}

\item{zlim}{}

\item{na.color}{}
}
\value{

}
\description{

}
