% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{trim_INLA_Param}
\alias{trim_INLA_Param}
\title{trim_INLA}
\arguments{
\item{trim_INLA}{(logical) should the \code{INLA_model_obj} within the
result be trimmed to only what is necessary to use \code{id_activations}?
Default: \code{TRUE}.}
}
\description{
trim_INLA
}
