% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_utils.R
\name{TrQbb}
\alias{TrQbb}
\title{Trace of Q beta' beta}
\usage{
TrQbb(kappa2, beta_hat, spde)
}
\arguments{
\item{kappa2}{scalar}

\item{beta_hat}{a vector}

\item{spde}{an spde object}
}
\value{
a scalar
}
\description{
Trace of Q beta' beta
}
\keyword{internal}
