% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_utils.R
\name{TrQEww}
\alias{TrQEww}
\title{Trace approximation function}
\usage{
TrQEww(kappa2, spde, P, mu, Vh)
}
\arguments{
\item{kappa2}{a scalar}

\item{spde}{spde object}

\item{P}{Matrix of dimension nk by N_s found by \code{solve(Sig_inv,Vh)}}

\item{mu}{posterior mean}

\item{Vh}{matrix of random variables with \code{nrow(Sig_inv)} rows and Ns
columns}
}
\value{
a scalar
}
\description{
Trace approximation function
}
\keyword{internal}
