% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_utils.R
\name{TrSigB}
\alias{TrSigB}
\title{Hutchinson estimator of the trace}
\usage{
TrSigB(P, B, vh)
}
\arguments{
\item{P}{Matrix of dimension nk by N_s found by \code{solve(Sig_inv,Vh)}}

\item{B}{Matrix of dimension nk by nk inside the desired trace product}

\item{vh}{Matrix of dimension nk by N_s in which elements are -1 or 1 with
equal probability.}
}
\value{
a scalar estimate of the trace of \code{Sigma \%*\% B}
}
\description{
Hutchinson estimator of the trace
}
\keyword{internal}
