% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{unmask_Mdat2In}
\alias{unmask_Mdat2In}
\title{Unmask data}
\usage{
unmask_Mdat2In(x, maskIn, maskMdat, fill_val = NA)
}
\arguments{
\item{x}{The matrix to unmask}

\item{maskIn, maskMdat}{The input mask, and the modeled locations mask, as
logical vectors. The latter should be a subset of the former. They should
have the same length.}

\item{fill_val}{The fill value. Default: \code{NA}.}
}
\description{
\code{maskMdat} is a subset of \code{maskIn}. \code{x} is masked by
\code{maskMdat}; this function makes it masked by \code{maskIn} instead
and uses \code{fill_val} for the previously masked-out locations.
}
\keyword{internal}
