% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateKappa2}
\alias{updateKappa2}
\title{Update the value of kappa2}
\usage{
updateKappa2(phi, in_list, n_sess, a_star, b_star, tol)
}
\arguments{
\item{phi}{a scalar}

\item{in_list}{a list with elements Cmat, Gmat, and GtCinvG}

\item{n_sess}{the integer number of sessions}

\item{a_star}{pre-computed coefficient}

\item{b_star}{pre-computed coefficient}

\item{tol}{the numeric tolerance for finding the optimal value of kappa2}
}
\description{
Update the value of kappa2
}
