% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_diagnosticICA.R
\name{UpdateTheta.diagnosticICA}
\alias{UpdateTheta.diagnosticICA}
\title{}
\usage{
UpdateTheta.diagnosticICA(
  template_mean,
  template_var,
  BOLD,
  theta,
  C_diag,
  return_MAP = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{template_mean}{}

\item{template_var}{}

\item{BOLD}{}

\item{theta}{}

\item{C_diag}{}

\item{return_MAP}{}

\item{verbose}{}
}
\value{

}
\description{

}
