% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_templateICA.R
\name{UpdateTheta}
\alias{UpdateTheta}
\alias{UpdateTheta.spatial}
\alias{UpdateTheta.independent}
\title{}
\usage{
UpdateTheta.spatial(
  template_mean,
  template_var,
  mesh,
  BOLD,
  theta,
  C_diag,
  s0_vec,
  D,
  Dinv_s0,
  common_smoothness = TRUE,
  verbose = FALSE,
  return_MAP = FALSE,
  dim_reduce_flag,
  update = c("all", "kappa", "A")
)

UpdateTheta.independent(
  template_mean,
  template_var,
  BOLD,
  theta,
  C_diag,
  verbose
)
}
\arguments{
\item{template_mean}{}

\item{template_var}{}

\item{mesh}{\verb{}}

\item{BOLD}{}

\item{theta}{}

\item{C_diag}{}

\item{s0_vec}{}

\item{D}{}

\item{Dinv_s0}{}

\item{common_smoothness}{}

\item{verbose}{}

\item{return_MAP}{}

\item{dim_reduce_flag}{}

\item{update}{}
}
\value{

}
\description{

}
