% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{validate_spatial}
\alias{validate_spatial}
\title{Validate \code{spatial}}
\usage{
validate_spatial(spatial)
}
\arguments{
\item{spatial}{\code{spatial}}
}
\value{
\code{NULL}, invisibly
}
\description{
Validate \code{spatial}
}
\keyword{internal}
