% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{verbose_Param}
\alias{verbose_Param}
\title{verbose}
\arguments{
\item{verbose}{\code{1} (default) to print occasional updates during model
computation; \code{2} for occasional updates as well as running INLA in
verbose mode (if \code{Bayes}), or \code{0} for no printed updates.}
}
\description{
verbose
}
