% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{verbose_Param_direct_TRUE}
\alias{verbose_Param_direct_TRUE}
\title{verbose: direct only, TRUE}
\arguments{
\item{verbose}{Should occasional updates be printed? Default: \code{TRUE}.}
}
\description{
verbose: direct only, TRUE
}
