% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex_areas.R
\name{vertex_areas}
\alias{vertex_areas}
\title{Surface area of each vertex}
\usage{
vertex_areas(mesh)
}
\arguments{
\item{mesh}{An \code{"inla.mesh"} object (see \code{\link{make_mesh}} for
surface data).}
}
\value{
Vector of areas
}
\description{
Compute surface areas of each vertex in a triangular mesh.
}
\section{INLA Requirement}{

This function requires the \code{INLA} package, which is not a CRAN package.
See \url{https://www.r-inla.org/download-install} for easy installation instructions.
}

