% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol2mesh.R
\name{vol2mesh}
\alias{vol2mesh}
\title{Construct a triangular mesh from a 2D volumetric mask (need to generalize this to a 3D volume)}
\usage{
vol2mesh(mask)
}
\arguments{
\item{mask}{A matrix of 0s and 1s representing a volumetric mask}
}
\value{
An inla.mesh object.  See \code{help(inla.mesh.2d)} for details.
}
\description{
Construct a triangular mesh from a 2D volumetric mask (need to generalize this to a 3D volume)
}
\examples{
\dontrun{}
}
