% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol2spde.R
\name{vol2spde}
\alias{vol2spde}
\title{Construct a triangular mesh from a 3D volumetric mask}
\usage{
vol2spde(mask, res = c(2, 2, 2), buffer = c(1, 1, 3, 4, 4), neighbor_order = 1)
}
\arguments{
\item{mask}{An array of 0s and 1s representing a volumetric mask}

\item{res}{The spatial resolution in each direction, in mm. For example, c(2,2,2) indicates 2mm isotropic voxels.}

\item{buffer}{Size of extra voxels layers around the bounding box, in terms of voxels. Set to NULL for no buffer.}

\item{neighbor_order}{What order of neighbors of data locations to keep? (1 = first-order neighbors, 2 = first- and second-order neighbors, etc.)}
}
\value{
An inla.spde2 object.
}
\description{
Construct a triangular mesh from a 3D volumetric mask
}
\section{INLA Requirement}{

This function requires the \code{INLA} package, which is not a CRAN package.
See \url{https://www.r-inla.org/download-install} for easy installation instructions.
}

