% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{activations.classical}
\alias{activations.classical}
\title{Identification of areas of activation in a General Linear Model using classical methods}
\usage{
activations.classical(
  x,
  gamma = 0,
  alpha = 0.05,
  correction = c("FWER", "FDR", "none"),
  fields,
  session,
  mesh = NULL
)
}
\arguments{
\item{x}{A \code{BayesGLM} object}

\item{gamma, alpha, correction, fields, session}{See \code{\link{activations}}.}

\item{mesh}{(Optional) An \code{"inla.mesh"} object (see \code{\link{make_mesh}} for
surface data). Only necessary for computing surface areas of identified activations.}
}
\value{
A matrix corresponding to the
0-1 activation status for the model coefficients.
}
\description{
Identification of areas of activation in a General Linear Model using classical methods
}
\keyword{internal}
