% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateICA.R
\name{activations}
\alias{activations}
\title{}
\usage{
activations(
  result,
  u = 0,
  alpha = 0.01,
  type = ">",
  method_p = "BH",
  verbose = FALSE,
  which.ICs = NULL,
  deviation = FALSE
)
}
\arguments{
\item{result}{}

\item{u}{}

\item{alpha}{}

\item{type}{}

\item{method_p}{}

\item{verbose}{}

\item{which.ICs}{}

\item{deviation}{}
}
\value{

}
\description{

}
