% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{ar_order_Param}
\alias{ar_order_Param}
\title{ar_order}
\arguments{
\item{ar_order}{(For prewhitening) The order of the autoregressive (AR) model
to use for prewhitening. If \code{0}, do not prewhiten. Default: \code{6}.

For multi-session modeling, note that a single AR model is used; its
coefficients will be the average estimate from each session.}
}
\description{
ar_order
}
