% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{ar_smooth_Param}
\alias{ar_smooth_Param}
\title{ar_smooth}
\arguments{
\item{ar_smooth}{(numeric) FWHM parameter for smoothing the AR model
coefficient estimates for prewhitening. Remember that
\eqn{\sigma = \frac{FWHM}{2*sqrt(2*log(2)}}. Set to \code{0} or \code{NULL}
to not do any smoothing. Default: \code{5}.}
}
\description{
ar_smooth
}
