% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{Bayes_Param}
\alias{Bayes_Param}
\title{Bayes}
\arguments{
\item{Bayes}{Fir the spatial Bayesian GLM? Default: \code{TRUE}. If
\code{FALSE}, only fit the classical GLM. (The classical GLM is always
returned, whether \code{Bayes} is \code{TRUE} or \code{FALSE}.)}
}
\description{
Bayes
}
