% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{Bayes_Param}
\alias{Bayes_Param}
\title{Bayes}
\arguments{
\item{Bayes}{Perform spatial Bayesian modeling? Default: \code{TRUE}. If
\code{FALSE}, only perform classical (massive univariate) modeling. (The classical GLM
result is always returned, whether \code{Bayes} is \code{TRUE} or \code{FALSE}.)}
}
\description{
Bayes
}
