% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{Bayes_Param}
\alias{Bayes_Param}
\title{Bayes}
\arguments{
\item{Bayes}{If \code{TRUE} (default), will fit a spatial Bayesian GLM in
addition to the classical GLM. (The classical GLM is always returned.)}
}
\description{
Bayes
}
