% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesfMRI-package.R
\docType{package}
\name{BayesfMRI-package}
\alias{BayesfMRI}
\alias{BayesfMRI-package}
\title{BayesfMRI: Spatial Bayesian Methods for Task Functional MRI Studies}
\description{
Performs a spatial Bayesian general linear model (GLM) for task functional magnetic resonance imaging (fMRI) data on the cortical surface. Additional models include group analysis and inference to detect thresholded areas of activation. Includes direct support for the 'CIFTI' neuroimaging file format. For more information see A. F. Mejia, Y. R. Yue, D. Bolin, F. Lindgren, M. A. Lindquist (2020) \doi{10.1080/01621459.2019.1611582} and D. Spencer, Y. R. Yue, D. Bolin, S. Ryan, A. F. Mejia (2022) \doi{10.1016/j.neuroimage.2022.118908}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mandymejia/BayesfMRI}
  \item Report bugs at \url{https://github.com/mandymejia/BayesfMRI/issues}
}

}
\author{
\strong{Maintainer}: Amanda Mejia \email{mandy.mejia@gmail.com}

Authors:
\itemize{
  \item Daniel Spencer \email{danieladamspencer@gmail.com} (\href{https://orcid.org/0000-0002-9705-3605}{ORCID})
}

Other contributors:
\itemize{
  \item Damon Pham \email{damondpham@gmail.com} (\href{https://orcid.org/0000-0001-7563-4727}{ORCID}) [contributor]
  \item David Bolin \email{david.bolin@kaust.edu.sa} [contributor]
  \item Sarah Ryan \email{sarahryan320@gmail.com} [contributor]
  \item Yu (Ryan) Yue \email{yybaruch@gmail.com} [contributor]
}

}
