% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM.R
\name{BayesGLM}
\alias{BayesGLM}
\title{Applies spatial Bayesian GLM to task fMRI data}
\usage{
BayesGLM(data, vertices = NULL, faces = NULL, mesh = NULL,
  mask = NULL, scale = TRUE, return_INLA_result = TRUE,
  outfile = NULL)
}
\arguments{
\item{data}{A list of sessions, where each session is a list with elements
BOLD, design and nuisance.  See \code{?create.session} and \code{?is.session} for more details.
List element names represent session names.}

\item{vertices}{A Vx3 matrix of vertex locations of the triangular mesh in Euclidean space.}

\item{faces}{A Wx3 matrix, where each row contains the vertex indices for a given face or triangle in the triangular mesh. W is the number of faces in the mesh.}

\item{mesh}{A \code{inla.mesh} object.  Must be provided if and only if \code{vertices} and \code{faces} are not.}

\item{mask}{Sarah knows what this does!}

\item{scale}{If TRUE, scale timeseries data so estimates represent percent signal change.  Else, do not scale.}

\item{return_INLA_result}{If TRUE, object returned will include the INLA model object (can be large).  Default is TRUE. Required for running \code{id_activations} on \code{BayesGLM} model object.}

\item{outfile}{File name where results will be written (for use by \code{BayesGLM_group}).}
}
\value{
A list containing...
}
\description{
Applies spatial Bayesian GLM to task fMRI data
}
\examples{
\dontrun{}
}
