% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM.R
\name{BayesGLM}
\alias{BayesGLM}
\title{BayesGLM}
\usage{
BayesGLM(
  data,
  beta_names = NULL,
  vertices = NULL,
  faces = NULL,
  mesh = NULL,
  mask = NULL,
  scale_BOLD = TRUE,
  scale_design = TRUE,
  Bayes = TRUE,
  ar_order = 6,
  ar_smooth = 5,
  num.threads = 4,
  return_INLA_result = TRUE,
  outfile = NULL,
  verbose = FALSE,
  contrasts = NULL,
  avg_sessions = FALSE,
  trim_INLA = TRUE
)
}
\arguments{
\item{data}{A list of sessions, where each session is a list with elements
BOLD, design and nuisance. See \code{?create.session} and \code{?is.session}
for more details.
List element names represent session names.}

\item{beta_names}{(Optional) Names of tasks represented in design matrix}

\item{vertices}{A \eqn{V x 3} matrix, where each row contains the Euclidean
coordinates at which a given vertex in the mesh is located. \eqn{V} is the
number of vertices in the mesh}

\item{faces}{An \eqn{F x 3} matrix, where each row contains the vertex
indices for a given triangular face in the mesh. \eqn{F} is the number of
faces in the mesh.}

\item{mesh}{An \code{"inla.mesh"} object (see \code{\link{make_mesh}} for
surface data).}

\item{mask}{(Optional) A length \eqn{V} logical vector indicating if each
vertex is to be included.}

\item{scale_BOLD}{Scale timeseries data so estimates represent percent signal
change? Default: \code{TRUE}. If \code{FALSE}, the data will just be
centered.}

\item{scale_design}{Scale the design matrix by dividing each column by its
maximum and then subtracting the mean? Default: \code{TRUE}.}

\item{Bayes}{If TRUE, will fit a spatial Bayesian GLM in addition to the
classical GLM. Classical GLM results are always returned.}

\item{ar_order}{(numeric) Controls prewhitening. If greater than zero, this
should be a number indicating the order of the autoregressive model to use
for prewhitening. If zero, do not prewhiten. Default: \code{6}.}

\item{ar_smooth}{FWHM parameter for smoothing. Remember that
\eqn{\sigma = \frac{FWHM}{2*sqrt(2*log(2)}}. Set to \code{0} or \code{NULL}
to not do any smoothing. Default: \code{5}.}

\item{num.threads}{The maximum number of threads to use in the inla-program
for model estimation. Default: \code{4}.}

\item{return_INLA_result}{Return the INLA model object? (It can be large.)
Default: \code{TRUE}. Required for running \code{id_activations}
after, but not for running BayesGLM_joint after to get
posterior quantities of group means or contrasts.}

\item{outfile}{File name where results will be written (for use by
\code{BayesGLM2}).}

\item{verbose}{Should INLA be run in verbose mode? Default: \code{FALSE}.}

\item{contrasts}{List of contrast vectors to be passed to \code{inla::inla}.}

\item{avg_sessions}{Average estimates for betas over multiple
sessions? Default: \code{FALSE}.}

\item{trim_INLA}{(logical) should the \code{INLA_result} objects within the
result be trimmed to only what is necessary to use \code{id_activations()}? Default: \code{TRUE}.}
}
\value{
A list containing...
}
\description{
Applies spatial Bayesian GLM to task fMRI data
}
\section{INLA Requirement}{

This function requires the \code{INLA} package, which is not a CRAN package.
See \url{https://www.r-inla.org/download-install} for easy installation instructions.
}

