% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM.R
\name{BayesGLM}
\alias{BayesGLM}
\title{BayesGLM}
\usage{
BayesGLM(
  data,
  beta_names = NULL,
  vertices = NULL,
  faces = NULL,
  mesh = NULL,
  mask = NULL,
  contrasts = NULL,
  scale_BOLD = c("auto", "mean", "sd", "none"),
  scale_design = TRUE,
  Bayes = TRUE,
  EM = FALSE,
  ar_order = 6,
  ar_smooth = 5,
  aic = FALSE,
  num.threads = 4,
  return_INLA_result = TRUE,
  outfile = NULL,
  verbose = FALSE,
  avg_sessions = FALSE,
  meanTol = 1e-06,
  varTol = 1e-06,
  emTol = 0.001,
  trim_INLA = TRUE
)
}
\arguments{
\item{data}{A list of sessions in the \code{"BfMRI.sess"} object format. Each
session is a list with elements "BOLD", "design" and "nuisance" (optional).
The name of each element in \code{data} is the name of that session. See
\code{?is.BfMRI.sess} for details.}

\item{beta_names}{(Optional) Names of tasks represented in design matrix}

\item{vertices}{A \eqn{V x 3} matrix, where each row contains the Euclidean
coordinates at which a given vertex in the mesh is located. \eqn{V} is the
number of vertices in the mesh}

\item{faces}{An \eqn{F x 3} matrix, where each row contains the vertex
indices for a given triangular face in the mesh. \eqn{F} is the number of
faces in the mesh.}

\item{mesh}{An \code{"inla.mesh"} object (see \code{\link{make_mesh}} for
surface data).}

\item{mask}{(Optional) A length \eqn{V} logical vector indicating if each
vertex is to be included.}

\item{contrasts}{List of contrast vectors to be passed to \code{inla::inla}.}

\item{scale_BOLD}{Option for scaling the BOLD response.

\if{html}{\out{<div class="sourceCode">}}\preformatted{If \code{"auto"} (default), will use mean scaling except if demeaned data
is detected, in which case sd scaling will be used instead.

\code{"mean"} scaling will scale the data to percent local signal change.

\code{"sd"} scaling will scale the data by local standard deviation.

\code{"none"} will only center the data, not scale it.
}\if{html}{\out{</div>}}}

\item{scale_design}{Scale the design matrix by dividing each column by its
maximum and then subtracting the mean? Default: \code{TRUE}.}

\item{Bayes}{If \code{TRUE} (default), will fit a spatial Bayesian GLM in
addition to the classical GLM. Classical GLM results are always returned.}

\item{EM}{(logical) Should the EM implementation of the Bayesian GLM be used?
Default: \code{FALSE}.}

\item{ar_order}{(numeric) Controls prewhitening. If greater than zero, this
should be a number indicating the order of the autoregressive model to use
for prewhitening. If zero, do not prewhiten. Default: \code{6}.}

\item{ar_smooth}{(numeric) FWHM parameter for smoothing. Remember that
\eqn{\sigma = \frac{FWHM}{2*sqrt(2*log(2)}}. Set to \code{0} or \code{NULL}
to not do any smoothing. Default: \code{5}.}

\item{aic}{Use the AIC to select AR model order between \code{0} and \code{ar_order}?}

\item{num.threads}{The maximum number of threads to use in the inla-program
for model estimation. Default: \code{4}.}

\item{return_INLA_result}{Return the INLA model object? (It can be large.)
Default: \code{TRUE}. Required for running \code{id_activations}
after, but not for running BayesGLM_joint after to get
posterior quantities of group means or contrasts.}

\item{outfile}{File name where results will be written (for use by
\code{BayesGLM2}).}

\item{verbose}{Should INLA be run in verbose mode? Default: \code{FALSE}.}

\item{avg_sessions}{Average estimates for betas over multiple
sessions? Default: \code{FALSE}.}

\item{meanTol, varTol}{Tolerance for mean and variance of each data location. Locations which
do not meet these thresholds are masked out of the analysis. Default: \code{1e-6}.}

\item{emTol}{The stopping tolerance for the EM algorithm. Default: \code{1e-3}.}

\item{trim_INLA}{(logical) should the \code{INLA_result} objects within the
result be trimmed to only what is necessary to use \code{id_activations()}? Default: \code{TRUE}.}
}
\value{
A list containing...
}
\description{
Applies spatial Bayesian GLM to task fMRI data
}
\section{INLA Requirement}{

This function requires the \code{INLA} package, which is not a CRAN package.
See \url{https://www.r-inla.org/download-install} for easy installation instructions.
}

