% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM.R
\name{BayesGLM}
\alias{BayesGLM}
\title{}
\usage{
BayesGLM(
  cifti_fname,
  surfL_fname = NULL,
  surfR_fname = NULL,
  sphereL_fname = NULL,
  sphereR_fname = NULL,
  brainstructures = c("left", "right", "subcortical"),
  wb_path = NULL,
  design = NULL,
  onsets = NULL,
  TR = NULL,
  nuisance = NULL,
  nuisance_include = c("drift", "dHRF"),
  scale_BOLD = TRUE,
  scale_design = TRUE,
  GLM_method = "both",
  session_names = NULL,
  resamp_res = 10000,
  num.threads = 4,
  verbose = FALSE,
  write_dir = NULL,
  outfile = NULL,
  return_INLA_result = FALSE
)
}
\arguments{
\item{cifti_fname}{}

\item{surfL_fname}{}

\item{surfR_fname}{}

\item{sphereL_fname}{}

\item{sphereR_fname}{}

\item{brainstructures}{}

\item{wb_path}{}

\item{design}{}

\item{onsets}{}

\item{TR}{}

\item{nuisance}{}

\item{nuisance_include}{}

\item{scale_BOLD}{}

\item{scale_design}{}

\item{GLM_method}{}

\item{session_names}{}

\item{resamp_res}{}

\item{num.threads}{}

\item{verbose}{}

\item{write_dir}{}

\item{outfile}{}

\item{return_INLA_result}{}
}
\value{

}
\description{

}
\details{

}
