% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM2.R
\name{BayesGLM2}
\alias{BayesGLM2}
\alias{BayesGLM_group}
\title{}
\usage{
BayesGLM2(
  results,
  contrasts = NULL,
  quantiles = NULL,
  excursion_type = NULL,
  gamma = 0,
  alpha = 0.05,
  nsamp_theta = 50,
  nsamp_beta = 100,
  no_cores = NULL,
  verbose = TRUE
)

BayesGLM_group(
  results,
  contrasts = NULL,
  quantiles = NULL,
  excursion_type = NULL,
  gamma = 0,
  alpha = 0.05,
  nsamp_theta = 50,
  nsamp_beta = 100,
  no_cores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{results}{}

\item{contrasts}{}

\item{quantiles}{}

\item{excursion_type}{}

\item{gamma}{}

\item{alpha}{}

\item{nsamp_theta}{}

\item{nsamp_beta}{}

\item{no_cores}{}

\item{verbose}{}
}
\value{

}
\description{

}
\details{
\verb{}
\verb{}

\verb{}
}
\section{}{
NA
}

