% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM2.R
\name{BayesGLM2}
\alias{BayesGLM2}
\alias{BayesGLM_group}
\title{Group-level Bayesian GLM}
\usage{
BayesGLM2(
  results,
  contrasts = NULL,
  quantiles = NULL,
  excursion_type = NULL,
  contrast_names = NULL,
  gamma = 0,
  alpha = 0.05,
  nsamp_theta = 50,
  nsamp_beta = 100,
  no_cores = NULL,
  verbose = TRUE
)

BayesGLM_group(
  results,
  contrasts = NULL,
  quantiles = NULL,
  excursion_type = NULL,
  gamma = 0,
  alpha = 0.05,
  nsamp_theta = 50,
  nsamp_beta = 100,
  no_cores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{results}{Either (1) a length \eqn{M} list of \code{"BayesGLM"} objects,
or (2) a length \eqn{M} character vector of files storing \code{"BayesGLM"}
objects saved with \code{\link{saveRDS}}.}

\item{contrasts}{(Optional) A list of vectors, each length
\eqn{M \times L \times S}, specifying the contrast(s) of interest
across subjects, where \eqn{M} is the number of subjects, \code{L} is the
number of tasks, and S is the number of sessions. See Details for more
information. If \code{NULL} (default), the average for each task across
subjects will be computed.}

\item{quantiles}{(Optional) Vector of posterior quantiles to return in
addition to the posterior mean.}

\item{excursion_type}{(For inference only) The type of excursion function for
the contrast (">", "<", "!="), or a vector thereof (each element
corresponding to one contrast).  If \code{NULL}, no inference performed.}

\item{contrast_names}{(Optional) Names of contrasts.}

\item{gamma}{(For inference only) Activation threshold for the excursion set,
or a vector thereof (each element corresponding to one contrast). Default:
\code{0}.}

\item{alpha}{(For inference only) Significance level for activation for the
excursion set, or a vector thereof (each element corresponding to one
contrast). Default: \code{.05}.}

\item{nsamp_theta}{Number of theta values to sample from posterior. Default:
\code{50}.}

\item{nsamp_beta}{Number of beta vectors to sample conditional on each theta
value sampled. Default: \code{100}.}

\item{no_cores}{The number of cores to use for sampling betas in parallel. If
\code{NULL} (default), do not run in parallel.}

\item{verbose}{Should occasional updates be printed? Default: \code{TRUE}.}
}
\value{
A list containing the estimates, PPMs and areas of activation for each contrast.
}
\description{
Performs group-level Bayesian GLM estimation and inference using the joint
approach described in Mejia et al. (2020).
}
\details{
Each contrast vector specifies a group-level summary of interest. Let \eqn{M}
be the number of subjects and \eqn{L} be the number of tasks. For example,
the contrast vector
\code{rep(rep(c(1/(M*nS),rep(0, L-1)),nS),M)}
represents the group average for the first task for \eqn{M} subjects;
\code{c(rep(rep(c(1/(M1*nS),rep(0, L-1)),nS),M1), rep(rep(c(-1/(M2*nS),rep(0, L-1)),nS),M2))}
represents the difference between the first \eqn{M_1} subjects and the remaining \eqn{M_2}
subjects (M1+M2=M) for the first task;
\code{rep(rep(c(1/(M*nS),-1/(M*nS),rep(0, L-2)), nS),M)}
represents the difference between the first two tasks, averaged over all
subjects.
}
\section{INLA Requirement}{

This function requires the \code{INLA} package, which is not a CRAN package.
See \url{https://www.r-inla.org/download-install} for easy installation instructions.
}

