% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM2.R
\name{BayesGLM2}
\alias{BayesGLM2}
\alias{BayesGLM_group}
\title{Performs group-level Bayesian GLM estimation and inference using the joint approach described in Mejia et al. (2019)}
\usage{
BayesGLM2(
  results,
  contrasts = NULL,
  quantiles = NULL,
  excursion_type = NULL,
  gamma = 0,
  alpha = 0.05,
  nsamp_theta = 50,
  nsamp_beta = 100,
  no_cores = NULL,
  verbose = TRUE
)

BayesGLM_group(
  results,
  contrasts = NULL,
  quantiles = NULL,
  excursion_type = NULL,
  gamma = 0,
  alpha = 0.05,
  nsamp_theta = 50,
  nsamp_beta = 100,
  no_cores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{results}{Either (1) a list of length M of objects of class BayesGLM,
or (2) a character vector of length M of file names output from the BayesGLM function.
M is the number of subjects.}

\item{contrasts}{(Optional) A list of vectors, each length M*K, specifying the contrast(s)
of interest across subjects, where M is the number of subjects and K is the number of tasks.
See Details for more information. Default is to compute the average for each task across subjects.}

\item{quantiles}{(Optional) Vector of posterior quantiles to return in addition to the posterior mean}

\item{excursion_type}{(For inference only) The type of excursion function for the contrast (">", "<", "!="),
or a vector thereof (each element corresponding to one contrast).  If NULL, no inference performed.}

\item{gamma}{(For inference only) Activation threshold for the excursion set, or a vector thereof (each element corresponding to one contrast).}

\item{alpha}{(For inference only) Significance level for activation for the excursion set, or a vector thereof (each element corresponding to one contrast).}

\item{nsamp_theta}{Number of theta values to sample from posterior. Default is 50.}

\item{nsamp_beta}{Number of beta vectors to sample conditional on each theta value sampled. Default is 100.}

\item{no_cores}{The number of cores to use for sampling betas in parallel. If NULL, do not run in parallel.}

\item{verbose}{Logical indicating if progress updates should be displayed.}
}
\value{
A list containing the estimates, PPMs and areas of activation for each contrast.
}
\description{
Performs group-level Bayesian GLM estimation and inference using the joint approach described in Mejia et al. (2019)
}
\details{
Each contrast vector specifies a group-level summary of interest.  Let M be the number of subjects and K be the number of tasks.
For example, the contrast vector \code{rep(c(1/M, rep(0, M-1)), K)} represents the group average for the first task;
\code{c( rep(c(1/M1, rep(0, M1-1)), K), rep(c(-1/M2, rep(0, M2-1)), K))} represents the difference between the first M1 subjects and the remaining M2 subjects (M1+M2=M) for the first task;
\code{rep( c(1/M,-1/M,rep(0, K-2)) ,M)} represents the difference between the first two tasks, averaged over all subjects.
}
\note{
This function requires the \code{INLA} package, which is not a CRAN package. See \url{http://www.r-inla.org/download} for easy installation instructions.
}
