% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{BayesGLM_argChecks}
\alias{BayesGLM_argChecks}
\title{Bayes GLM arg checks}
\usage{
BayesGLM_argChecks(
  scale_BOLD,
  Bayes = TRUE,
  EM = FALSE,
  ar_order = 6,
  ar_smooth = 5,
  aic = FALSE,
  n_threads = 4,
  return_INLA = c("trimmed", "full", "minimal"),
  verbose = 1,
  meanTol = 1e-06,
  varTol = 1e-06,
  emTol = 0.001
)
}
\arguments{
\item{scale_BOLD}{See \code{\link{BayesGLM}}.}

\item{Bayes, EM}{See \code{\link{BayesGLM}}.}

\item{ar_order, ar_smooth, aic}{See \code{\link{BayesGLM}}.}

\item{n_threads}{See \code{\link{BayesGLM}}.}

\item{return_INLA}{See \code{\link{BayesGLM}}.}

\item{verbose}{See \code{\link{BayesGLM}}.}

\item{meanTol, varTol, emTol}{See \code{\link{BayesGLM}}.}
}
\value{
The arguments that may have changed, in a list: \code{scale_BOLD},
\code{do_Bayesian}, \code{do_EM}, and \code{do_pw}.
}
\description{
Checks arguments for \code{BayesGLM} and \code{BayesGLM_fun}
}
\details{
Avoids duplicated code between \code{BayesGLM} and \code{BayesGLM_fun}
}
\keyword{internal}
