% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM.R
\name{BayesGLM_cifti}
\alias{BayesGLM_cifti}
\title{}
\usage{
BayesGLM_cifti(
  cifti_fname,
  surfL_fname = NULL,
  surfR_fname = NULL,
  brainstructures = c("left", "right"),
  design = NULL,
  onsets = NULL,
  TR = NULL,
  nuisance = NULL,
  nuisance_include = c("drift", "dHRF"),
  scale_BOLD = TRUE,
  scale_design = TRUE,
  GLM_method = "both",
  ar_order = 6,
  ar_smooth = 5,
  session_names = NULL,
  resamp_res = 10000,
  num.threads = 4,
  verbose = FALSE,
  outfile = NULL,
  return_INLA_result = FALSE,
  avg_sessions = TRUE,
  trim_INLA = TRUE
)
}
\arguments{
\item{cifti_fname}{}

\item{surfL_fname}{}

\item{surfR_fname}{}

\item{brainstructures}{}

\item{design, onsets, TR}{}

\item{nuisance}{}

\item{nuisance_include}{}

\item{scale_BOLD}{}

\item{scale_design}{}

\item{GLM_method}{}

\item{ar_order}{}

\item{ar_smooth}{}

\item{session_names}{}

\item{resamp_res}{}

\item{num.threads}{}

\item{verbose}{}

\item{outfile}{}

\item{return_INLA_result}{}

\item{avg_sessions}{}

\item{trim_INLA}{\verb{}\verb{}}
}
\value{

}
\description{

}
\section{}{
NA
}

