% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_cifti.R
\name{BayesGLM_cifti}
\alias{BayesGLM_cifti}
\title{BayesGLM for CIFTI}
\usage{
BayesGLM_cifti(
  cifti_fname,
  surfL_fname = NULL,
  surfR_fname = NULL,
  brainstructures = c("left", "right"),
  design = NULL,
  onsets = NULL,
  TR = NULL,
  nuisance = NULL,
  dHRF = c(0, 1, 2),
  hpf = NULL,
  DCT = if (is.null(hpf)) {     4 } else {     NULL },
  scale_BOLD = c("auto", "mean", "sd", "none"),
  scale_design = TRUE,
  Bayes = TRUE,
  EM = TRUE,
  ar_order = 6,
  ar_smooth = 5,
  aic = FALSE,
  resamp_res = 10000,
  num.threads = 4,
  verbose = FALSE,
  outfile = NULL,
  return_INLA_result = FALSE,
  avg_sessions = TRUE,
  session_names = NULL,
  meanTol = 1e-06,
  varTol = 1e-06,
  emTol = 0.001,
  trim_INLA = TRUE
)
}
\arguments{
\item{cifti_fname}{File path (or vector thereof, for multiple sessions) of CIFTI-format fMRI timeseries data (*.dtseries.nii).}

\item{surfL_fname}{File path of GIFTI-format left cortical surface (*.surf.gii). Must be provided if brainstructures includes "left" and Bayes=TRUE.}

\item{surfR_fname}{File path of GIFTI-format right cortical surface (*.surf.gii). Must be provided if brainstructures includes "right" and Bayes=TRUE.}

\item{brainstructures}{Character vector indicating which brain structure(s)
to obtain: \code{"left"} (left cortical surface) and/or \code{"right"} (right
cortical surface). Default: \code{c("left","right")} (entire cortical surface).
Note that the subcortical models have not yet been implemented.}

\item{design, onsets, TR}{Either provide \code{design}, or provide both \code{onsets} and \code{TR}.

\code{design} is a \eqn{T x K} task design matrix (or list of such
matrices, for multiple-session modeling) with column names representing
tasks. Each column represents the expected BOLD response due to each task,
a convolution of the hemodynamic response function (HRF) and the task
stimulus. Note that the scale of the regressors will affect the scale and
interpretation of the beta coefficients, so imposing a proper scale (e.g.,
set maximum to 1) is recommended.

\code{onsets} is a matrix of onsets (first column) and durations (second column)
for each task in seconds, organized as a list where each element of the
list corresponds to one task. Names of list should be task names. (Or for
multi-session modeling, a list of such lists.)

\code{TR} is the temporal resolution of the data in seconds.}

\item{nuisance}{(Optional) A TxJ matrix of nuisance signals
(or list of such matrices, for multiple-session modeling).}

\item{dHRF}{Logical indicating whether the temporal derivative of each column
in the design matrix should be added to \code{nuisance}. Default: \code{TRUE}.}

\item{hpf, DCT}{Add DCT bases to \code{nuisance} to apply a temporal
high-pass filter to the data? Only one of these arguments should be provided.
\code{hpf} should be the filter frequency; if it is provided, \code{TR}
must be provided too. The number of DCT bases to include will be computed
to yield a filter with as close a frequency to \code{hpf} as possible.
Alternatively, \code{DCT} can be provided to directly specify the number
of DCT bases to include.

Default: \code{DCT=4} (use four DCT bases for high-pass filtering; for
typical \code{TR} this amounts to lower filter frequency than the
approximately .01 Hz used in most studies.)}

\item{scale_BOLD}{Option for scaling the BOLD response.\preformatted{If \code{"auto"} (default), will use mean scaling except if demeaned data
is detected, in which case sd scaling will be used instead.

\code{"mean"} scaling will scale the data to percent local signal change.

\code{"sd"} scaling will scale the data by local standard deviation.

\code{"none"} will only center the data, not scale it.
}}

\item{scale_design}{Scale the design matrix by dividing each column by its
maximum and then subtracting the mean? Default: \code{TRUE}.}

\item{Bayes}{If TRUE, will fit a spatial Bayesian GLM in addition to the
classical GLM. Classical GLM results are always returned.}

\item{EM}{(logical) Should the EM implementation of the Bayesian GLM be used?
Default: \code{TRUE}.}

\item{ar_order}{(numeric) Controls prewhitening. If greater than zero, this
should be a number indicating the order of the autoregressive model to use
for prewhitening. If zero, do not prewhiten. Default: \code{6}.}

\item{ar_smooth}{FWHM parameter for smoothing. Remember that
\eqn{\sigma = \frac{FWHM}{2*sqrt(2*log(2)}}. Set to \code{0} or \code{NULL}
to not do any smoothing. Default: \code{5}.}

\item{aic}{Use the AIC to select AR model order between \code{0} and \code{ar_order}? Default: \code{FALSE}.}

\item{resamp_res}{The number of vertices to which each cortical surface should be resampled, or NULL if no resampling is to be performed. For computational feasibility, a value of 10000 or lower is recommended.}

\item{num.threads}{The maximum number of threads to use in the inla-program
for model estimation. Default: \code{4}.}

\item{verbose}{Should INLA be run in verbose mode? Default: \code{FALSE}.}

\item{outfile}{(Optional) File name (without extension) of output file for
\code{"BayesGLM"} result to use in Bayesian group modeling.
\code{"_left.rds"} or \code{"_right.rds"} will be appended for the left
cortex and right cortex results, respectively. Default: \code{NULL}
(do not save the results to any file).}

\item{return_INLA_result}{Return the INLA model object? (It can be large.)
Default: \code{FALSE}. Required for running \code{id_activations}
after, but not for running BayesGLM_joint after to get
posterior quantities of group means or contrasts.}

\item{avg_sessions}{Average estimates for betas over multiple
sessions? Default: \code{FALSE}.}

\item{session_names}{(Optional) A vector of names corresponding to each
session. Ignored if \code{avg_sessions == TRUE}.}

\item{meanTol, varTol}{Tolerance for mean and variance of each data location. Locations which
do not meet these thresholds are masked out of the analysis. Default: \code{1e-6}.}

\item{emTol}{The stopping tolerance for the EM algorithm. Default: \code{1e-3}.}

\item{trim_INLA}{(logical) should the \code{INLA_result} objects within the
result be trimmed to only what is necessary to use \code{id_activations()}? Default: \code{TRUE}.}
}
\value{
An object of class \code{"BayesGLM"}, a list containing...
}
\description{
Performs spatial Bayesian GLM on the cortical surface for fMRI task activation
}
\section{Connectome Workbench Requirement}{

This function uses a system wrapper for the 'wb_command' executable. The
user must first download and install the Connectome Workbench, available
from https://www.humanconnectome.org/software/get-connectome-workbench .
}

