% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM.R
\name{BayesGLM_cifti}
\alias{BayesGLM_cifti}
\title{Performs whole-brain spatial Bayesian GLM for fMRI task activation}
\usage{
BayesGLM_cifti(
  cifti_fname,
  surfL_fname = NULL,
  surfR_fname = NULL,
  brainstructures = c("left", "right", "subcortical"),
  wb_path = NULL,
  design = NULL,
  onsets = NULL,
  TR = NULL,
  nuisance = NULL,
  nuisance_include = c("drift", "dHRF"),
  scale_BOLD = TRUE,
  scale_design = TRUE,
  GLM_method = "both",
  session_names = NULL,
  resamp_res = 10000,
  num.threads = 4,
  verbose = FALSE,
  write_dir = NULL,
  outfile = NULL,
  return_INLA_result = FALSE,
  avg_betas_over_sessions = FALSE
)
}
\arguments{
\item{cifti_fname}{File path (or vector thereof, for multiple-session modeling) of CIFTI-format fMRI timeseries data (*.dtseries.nii).}

\item{surfL_fname}{File path of GIFTI-format left cortical surface (*.surf.gii). Must be provided if brainstructures includes "left" and GLM_method is "Bayesian" or "both".}

\item{surfR_fname}{File path of GIFTI-format right cortical surface (*.surf.gii). Must be provided if brainstructures includes "right" and GLM_method is "Bayesian" or "both".}

\item{brainstructures}{Character vector indicating which brain structure(s)
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right
cortical surface) and/or \code{"subcortical"} (subcortical and cerebellar
gray matter). Can also be \code{"all"} (obtain all three brain structures).
Default: \code{c("left","right")} (cortical surface only).}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable.
If not provided, should be set with
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}

\item{design, onsets, TR}{All or none must be provided.

\code{design} is a \eqn{T x K} task design matrix (or list of such
matrices, for multiple-session modeling) with column names representing
tasks. Each column represents the expected BOLD response due to each task,
a convolution of the hemodynamic response function (HRF) and the task
stimulus. Note that the scale of the regressors will affect the scale and
interpretation of the beta coefficients, so imposing a proper scale (e.g.,
set maximum to 1) is recommended.

\code{onsets} is a matrix of onsets (first column) and durations (second column)
for each task in seconds, organized as a list where each element of the
list corresponds to one task. Names of list should be task names. (Or for
multi-session modeling, a list of such lists.)

\code{TR} is the temporal resolution of the data in seconds.}

\item{nuisance}{(Optional) A TxJ matrix of nuisance signals (or list of such matrices, for multiple-session modeling).}

\item{nuisance_include}{(Optional) Additional nuisance covariates to include.  Default is 'drift' (linear and quadratic drift terms) and 'dHRF' (temporal derivative of each column of design matrix).}

\item{scale_BOLD}{Scale timeseries data so estimates represent percent signal
change? Default: \code{TRUE}. If \code{FALSE}, the data will just be
centered.}

\item{scale_design}{Scale the design matrix by dividing each column by its
maximum and then subtracting the mean? Default: \code{TRUE}.}

\item{GLM_method}{Either 'Bayesian' for spatial Bayesian GLM only, 'classical' for the classical GLM only, or 'both' to return both classical and Bayesian estimates of task activation.}

\item{session_names}{(Optional) A vector of names corresponding to each
session.}

\item{resamp_res}{The number of vertices to which each cortical surface should be resampled, or NULL if no resampling is to be performed. For computational feasibility, a value of 10000 or lower is recommended.}

\item{num.threads}{The maximum number of threads to use in the inla-program
for model estimation. Default: \code{4}.}

\item{verbose}{Should INLA be run in verbose mode? Default: \code{FALSE}.}

\item{write_dir}{(Optional) Location where to write resampled data (if resample != NULL) and output files. If NULL, use the current working directory.}

\item{outfile}{(Optional) File name (without extension) of output file for BayesGLM result to use in Bayesian group modeling.}

\item{return_INLA_result}{Return the INLA model object? (It can be large.)
Default: \code{FALSE}. Required for running \code{\link{id_activations}}
after, but not for running \code{\link{BayesGLM_joint}} after to get
posterior quantities of group means or contrasts.}

\item{avg_betas_over_sessions}{(logical) Should estimates for betas be averaged together over multiple sessions?}
}
\value{
An object of class BayesGLM, a list containing ...
}
\description{
Performs whole-brain spatial Bayesian GLM for fMRI task activation
}
\details{
This function uses a system wrapper for the 'wb_command' executable. The user must first download and install the Connectome Workbench,
available from https://www.humanconnectome.org/software/get-connectome-workbench. The 'wb_path' argument is the full file path to the 'wb_command' executable file.

The subcortical brain structure labels range from 3-21 and represent:
3 Accumbens-L
4 Accumbens-R
5 Amygdala-L
6 Amygdala-R
7 Brain Stem
8 Caudate-L
9 Caudate-R
10 Cerebellum-L
11 Cerebellum-R
12 Diencephalon-L
13 Diencephalon-R
14 Hippocampus-L
15 Hippocampus-R
16 Pallidum-L
17 Pallidum-R
18 Putamen-L
19 Putamen-R
20 Thalamus-L
21 Thalamus-R
}
