% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_cifti.R
\name{BayesGLM_cifti}
\alias{BayesGLM_cifti}
\title{BayesGLM for CIFTI}
\usage{
BayesGLM_cifti(
  cifti_fname,
  surfL_fname = NULL,
  surfR_fname = NULL,
  brainstructures = c("left", "right"),
  design = NULL,
  onsets = NULL,
  TR = NULL,
  nuisance = NULL,
  dHRF = c(0, 1, 2),
  hpf = NULL,
  DCT = if (is.null(hpf)) {
     4
 } else {
     NULL
 },
  resamp_res = 10000,
  task_names = NULL,
  session_names = NULL,
  contrasts = NULL,
  scale_BOLD = c("auto", "mean", "sd", "none"),
  scale_design = TRUE,
  Bayes = TRUE,
  ar_order = 6,
  ar_smooth = 5,
  aic = FALSE,
  num.threads = 4,
  return_INLA_result = TRUE,
  outfile = NULL,
  verbose = FALSE,
  avg_sessions = TRUE,
  meanTol = 1e-06,
  varTol = 1e-06,
  emTol = 0.001,
  trim_INLA = TRUE
)
}
\arguments{
\item{cifti_fname}{fMRI timeseries data in CIFTI format ("*.dtseries.nii").
For single-session analysis this can be a file path to a CIFTI file or a
\code{"xifti"} object from the \code{ciftiTools} package. For multi-session
analysis this can be a vector of file paths or a list of \code{"xifti"}
objects.}

\item{surfL_fname}{Left cortex surface geometry in GIFTI format
("*.surf.gii"). This can be a file path to a GIFTI file or a \code{"surf"}
object from the \code{ciftiTools} package. This argument is only used if
\code{brainstructures} includes \code{"left"} and \code{Bayes==TRUE}. If
it's not provided, the HCP group-average inflated surface included in the
\code{ciftiTools} package will be used.}

\item{surfR_fname}{Right cortex surface geometry in GIFTI format
("*.surf.gii"). This can be a file path to a GIFTI file or a \code{"surf"}
object from the \code{ciftiTools} package. This argument is only used if
\code{brainstructures} includes \code{"right"} and \code{Bayes==TRUE}. If
it's not provided, the HCP group-average inflated surface included in the
\code{ciftiTools} package will be used.}

\item{brainstructures}{Character vector indicating which brain structure(s)
to analyze: \code{"left"} (left cortical surface) and/or \code{"right"}
(right cortical surface). Default: \code{c("left","right")} (both
hemispheres). Note that the subcortical models have not yet been implemented.}

\item{design, onsets, TR}{Either provide \code{design}, or provide both
\code{onsets} and \code{TR}.

\code{design} is a \eqn{T \times K} task design matrix (or list of such
matrices, for multiple-session modeling) with column names representing
tasks. Each column represents the expected BOLD response due to each task,
a convolution of the hemodynamic response function (HRF) and the task
stimulus. Note that the scale of the regressors will affect the scale and
interpretation of the beta coefficients, so imposing a proper scale (e.g.,
set maximum to 1) is recommended.

\code{onsets} is a \eqn{K}-length list in which the name of each element is
the name of the corresponding task, and the value of each element is a
matrix of onsets (first column) and durations (second column) for each
stimuli (each row) of the corresponding task. For multi-session modeling,
this argument should be a list of such lists.

\code{TR} is the temporal resolution of the data, in seconds.}

\item{nuisance}{(Optional) A \eqn{T \times J} matrix of nuisance signals
(or list of such matrices, for multiple-session modeling).}

\item{dHRF}{Set to \code{1} to add the temporal derivative of each column
in the design matrix, \code{2} to add the second derivatives too, or
\code{0} to not add any columns. Default: \code{1}.}

\item{hpf, DCT}{Add DCT bases to \code{nuisance} to apply a temporal
high-pass filter to the data? Only one of these arguments should be provided.
\code{hpf} should be the filter frequency; if it is provided, \code{TR}
must be provided too. The number of DCT bases to include will be computed
to yield a filter with as close a frequency to \code{hpf} as possible.
Alternatively, \code{DCT} can be provided to directly specify the number
of DCT bases to include.

Default: \code{DCT=4}. For typical \code{TR}, four DCT bases amounts to a
lower frequency cutoff than the approximately .01 Hz used in most studies.
We selected this default to err on the side of retaining more low-frequency
information, but we recommend setting these arguments to values most
appropriate for the data analysis at hand.

Using at least two DCT bases is as sufficient as using linear and quadratic
drift terms in the design matrix. So if DCT detrending is being used, there
is no need to add linear and quadratic drift terms to \code{nuisance}.}

\item{resamp_res}{The number of vertices to which each cortical surface
should be resampled, or \code{NULL} to not resample. For computational
feasibility, a value of \code{10000} or lower is recommended.}

\item{task_names}{(Optional) Names of tasks represented in design matrix.}

\item{session_names}{(Optional, and only relevant for multi-session modeling)
Names of each session. Default: \code{NULL}. In \code{\link{BayesGLM}} this
argument will overwrite the names of the list entries in \code{data}, if
both exist.}

\item{contrasts}{List of contrast vectors to be passed to \code{inla::inla}.}

\item{scale_BOLD}{Option for scaling the BOLD response.

\if{html}{\out{<div class="sourceCode">}}\preformatted{If \code{"auto"} (default), will use mean scaling except if demeaned data
is detected, in which case sd scaling will be used instead.

\code{"mean"} scaling will scale the data to percent local signal change.

\code{"sd"} scaling will scale the data by local standard deviation.

\code{"none"} will only center the data, not scale it.
}\if{html}{\out{</div>}}}

\item{scale_design}{Scale the design matrix by dividing each column by its
maximum and then subtracting the mean? Default: \code{TRUE}.}

\item{Bayes}{If \code{TRUE} (default), will fit a spatial Bayesian GLM in
addition to the classical GLM. Classical GLM results are always returned.}

\item{ar_order}{(numeric) Controls prewhitening. If greater than zero, this
should be a number indicating the order of the autoregressive model to use
for prewhitening. If zero, do not prewhiten. Default: \code{6}.}

\item{ar_smooth}{(numeric) FWHM parameter for smoothing. Remember that
\eqn{\sigma = \frac{FWHM}{2*sqrt(2*log(2)}}. Set to \code{0} or \code{NULL}
to not do any smoothing. Default: \code{5}.}

\item{aic}{Use the AIC to select AR model order between \code{0} and
\code{ar_order}? Default: \code{FALSE}.}

\item{num.threads}{The maximum number of threads to use in the inla-program
for model estimation. Default: \code{4}.}

\item{return_INLA_result}{Return the INLA model object? (It can be large.)
Default: \code{TRUE}. Required for running \code{id_activations}
after, but not for running BayesGLM_joint after to get
posterior quantities of group means or contrasts.}

\item{outfile}{File name where results will be written (for use by
\code{BayesGLM2}).}

\item{verbose}{Should INLA be run in verbose mode? Default: \code{FALSE}.}

\item{avg_sessions}{Average estimates for betas over multiple
sessions? Default: \code{TRUE}.}

\item{meanTol, varTol}{Tolerance for mean and variance of each data location.
Locations which do not meet these thresholds are masked out of the analysis.
Default: \code{1e-6} for both.}

\item{emTol}{The stopping tolerance for the EM algorithm. Default:
\code{1e-3}.}

\item{trim_INLA}{(logical) should the \code{INLA_result} objects within the
result be trimmed to only what is necessary to use \code{id_activations}?
Default: \code{TRUE}.}
}
\value{
An object of class \code{"BayesGLM_cifti"}: a list with elements
\describe{
\item{betas_Bayesian}{The task coefficients for the Bayesian model.}
\item{betas_classical}{The task coefficients for the classical model.}
\item{GLMs_Bayesian}{The entire list of GLM results, except for parameters estimated for the classical model.}
\item{GLMs_classical}{Parameters estimated for the classical model from the GLM.}
\item{session_names}{The names of the sessions.}
\item{n_sess_orig}{The number of sessions (before averaging, if applicable).}
\item{task_names}{The task part of the design matrix, after centering and scaling, but before any nuisance regression or prewhitening.}
}
}
\description{
Performs spatial Bayesian GLM on the cortical surface for fMRI task activation
}
\section{Connectome Workbench Requirement}{

This function uses a system wrapper for the 'wb_command' executable. The
user must first download and install the Connectome Workbench, available
from https://www.humanconnectome.org/software/get-connectome-workbench .
}

\section{INLA Requirement}{

This function requires the \code{INLA} package, which is not a CRAN package.
See \url{https://www.r-inla.org/download-install} for easy installation instructions.
}

