% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_group.R
\name{BayesGLM_group}
\alias{BayesGLM_group}
\title{Applies joint approach to group-level analysis to task fMRI data}
\usage{
BayesGLM_group(result, A, contrasts = NULL, thresholds = 0,
  alpha = 0.05, no_cores = NULL)
}
\arguments{
\item{result}{Either (1) a list of length M of objects of class BayesGLM, or (2) a character vector of length M of file names output from the BayesGLM function. M is the number of subjects.}

\item{A}{A matrix to translate between original and mesh locations.}

\item{contrasts}{A vector specifying the contrast of interest.  See Details for more information.}

\item{thresholds}{The vector of activation thresholds}

\item{alpha}{The significance level for activation}

\item{no_cores}{The number of cores to use for sampling in parallel}
}
\value{
A list containing...
}
\description{
Applies joint approach to group-level analysis to task fMRI data
}
\details{
The contrast vector specifies the group-level quantity of interest.  For example, the vector rep(1,M*K) would return the group average for each of K tasks;
the vector \code{c(rep(1,M1*K)}, \code{rep(-1,M2*K))} would return the difference between the average within two groups of size M1 and M2, respectively, for each of K tasks;
the vector \code{rep(rep(1,-1,0,0),each=V),M)} would return the difference between the first two tasks (of 4), averaged over all subjects.
}
\examples{
\dontrun{}
}
