% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_vol3D.R
\name{BayesGLM_vol3D}
\alias{BayesGLM_vol3D}
\title{BayesGLM for 3D volume}
\usage{
BayesGLM_vol3D(
  data,
  locations,
  labels,
  groups_df = NULL,
  scale_BOLD = TRUE,
  scale_design = TRUE,
  return_INLA_result = FALSE,
  outfile = NULL,
  num.threads = 6,
  verbose = FALSE
)
}
\arguments{
\item{data}{A list of sessions, where each session is a list with elements
BOLD, design and nuisance.  See \code{?create.session} and \code{?is.session} for more details.
List element names represent session names.}

\item{locations}{Vx3 matrix of x,y,z coordinates of each voxel}

\item{labels}{Vector of length V of region labels}

\item{groups_df}{Data frame indicating the name and model group of each region.  See Details.}

\item{scale_BOLD}{Scale timeseries data so estimates represent percent signal
change? Default: \code{TRUE}. If \code{FALSE}, the data will just be
centered.}

\item{scale_design}{Scale the design matrix by dividing each column by its
maximum and then subtracting the mean? Default: \code{TRUE}.}

\item{return_INLA_result}{Return the INLA model object? (It can be large.)
Default: \code{FALSE}. Required for running \code{id_activations}
after, but not for running BayesGLM_joint after to get
posterior quantities of group means or contrasts.}

\item{outfile}{File name where results will be written (for use by \code{BayesGLM_grp}).}

\item{num.threads}{The maximum number of threads to use in the inla-program
for model estimation. Default: \code{4}.}

\item{verbose}{Should INLA be run in verbose mode? Default: \code{FALSE}.}
}
\value{
A list containing...
}
\description{
Applies spatial Bayesian GLM to task fMRI data for 3D subcortical volumes
}
\details{
The subcortical data is separated into regions, whose sizes range
from approximately 100 voxels to approximately 9000 voxels.  Smaller regions
are grouped together to improve model fit.
The \code{groups_df} argument specifies which regions are grouped together.
This argument should be a data frame with R rows (the number of regions) and
three columns: label, region, and group.
The label column is the numerical identifier of each region; the region
column contains the region names, and the group column contains the model
group assignments (e.g. 1,2,3). Regions to be excluded
from analysis are indicated by NA in the group assignment.
}
\section{INLA Requirement}{

This function requires the \code{INLA} package, which is not a CRAN package.
See \url{https://www.r-inla.org/download-install} for easy installation instructions.
}

