% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_vol3D.R
\name{BayesGLM_vol3D}
\alias{BayesGLM_vol3D}
\title{Applies spatial Bayesian GLM to task fMRI data for 3D subcortical volumes}
\usage{
BayesGLM_vol3D(data, spde_obj = NULL, scale = TRUE,
  return_INLA_result = TRUE, outfile = NULL)
}
\arguments{
\item{data}{A list of sessions, where each session is a list with elements
BOLD, design and nuisance.  See \code{?create.session} and \code{?is.session} for more details.
List element names represent session names.}

\item{scale}{If TRUE, scale timeseries data so estimates represent percent signal change.  Else, do not scale.}

\item{return_INLA_result}{If TRUE, object returned will include the INLA model object (can be large).  Default is TRUE. Required for running \code{id_activations} on \code{BayesGLM} model object.}

\item{outfile}{File name where results will be written (for use by \code{BayesGLM_grp}).}
}
\value{
A list containing...
}
\description{
Applies spatial Bayesian GLM to task fMRI data for 3D subcortical volumes
}
\examples{
\dontrun{}
}
