% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_functions.R
\name{BayesGLMEM}
\alias{BayesGLMEM}
\title{Expectation Maximization for the Bayesian General Linear Model}
\usage{
BayesGLMEM(
  data,
  beta_names = NULL,
  vertices = NULL,
  faces = NULL,
  mesh = NULL,
  mask = NULL,
  scale_BOLD = TRUE,
  scale_design = TRUE,
  EM_method = "separate",
  use_SQUAREM = TRUE,
  tol = NULL,
  num.threads = 1,
  outfile = NULL,
  verbose = FALSE,
  avg_sessions = TRUE
)
}
\arguments{
\item{data}{A list of sessions, where each session is a list with elements
BOLD, design and nuisance. See \code{?create.session} and \code{?is.session}
for more details.
List element names represent session names.}

\item{beta_names}{(Optional) Names of tasks represented in design matrix}

\item{vertices}{A \eqn{V x 3} matrix, where each row contains the Euclidean
coordinates at which a given vertex in the mesh is located. \eqn{V} is the
number of vertices in the mesh}

\item{faces}{An \eqn{F x 3} matrix, where each row contains the vertex
indices for a given triangular face in the mesh. \eqn{F} is the number of
faces in the mesh.}

\item{mesh}{An \code{"inla.mesh"} object (see \code{\link{make_mesh}} for
surface data).}

\item{mask}{(Optional) A length \eqn{V} logical vector indicating if each
vertex is to be included.}

\item{scale_BOLD}{Scale timeseries data so estimates represent percent signal
change? Default: \code{TRUE}. If \code{FALSE}, the data will just be
centered.}

\item{scale_design}{Scale the design matrix by dividing each column by its
maximum and then subtracting the mean? Default: \code{TRUE}.}

\item{EM_method}{Either "joint" or "separate" for choosing whether covariates
should share hyperparameter values.}

\item{use_SQUAREM}{(logical) Should the SQUAREM package be used to speed up
convergence?}

\item{tol}{If use_SQUAREM == TRUE, an absolute change limit for
when the EM algorithm should be stopped (Default = 1e-3). If use_SQUAREM ==
FALSE, a percent change limit for when the EM algorithm should be stopped
(Default = 1). A value of NULL will result in the default value being used.}

\item{num.threads}{(optional) allows users to specify the number of threads used
to work in parallel across the different tasks}

\item{outfile}{(Optional) File name (without extension) of output file for
BayesGLMEM result to use in Bayesian group modeling.}

\item{verbose}{Should occasional updates be printed? Default: \code{TRUE}.}

\item{avg_sessions}{Average estimates for betas over multiple
sessions? Default: \code{TRUE}.}
}
\value{
A list containing...
}
\description{
Applies spatial Bayesian GLM to task fMRI data
}
\section{INLA Requirement}{

This function requires the \code{INLA} package, which is not a CRAN package.
See \url{https://www.r-inla.org/download-install} for easy installation instructions.
}

