% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.posterior.thetasamp.R
\name{beta.posterior.thetasamp}
\alias{beta.posterior.thetasamp}
\title{Internal function used in joint approach to group-analysis}
\usage{
beta.posterior.thetasamp(
  theta,
  spde,
  Xcros,
  Xycros,
  contrasts,
  quantiles,
  excursion_type,
  gamma,
  alpha,
  nsamp_beta = 100
)
}
\arguments{
\item{theta}{A sample of theta (hyperparameters).}

\item{spde}{A SPDE object from inla.spde2.matern() function.}

\item{Xcros}{A crossproduct of design matrix.}

\item{Xycros}{A crossproduct of design matrix and BOLD y.}

\item{contrasts}{A list of vectors of length M*K specifying the contrasts of interest.  See Details for more information.}

\item{quantiles}{Vector of posterior quantiles to return in addition to the posterior mean}

\item{excursion_type}{Vector of excursion function type (">", "<", "!=") for each contrast}

\item{gamma}{Vector of activation thresholds for each contrast}

\item{alpha}{Significance level for activation for the excursion sets}

\item{nsamp_beta}{The number of samples to draw from p(beta|theta,y)}
}
\value{
A list containing...
}
\description{
Internal function used in joint approach to group-analysis
}
\details{
The contrast vector specifies the group-level quantity of interest.  For example, the vector \code{rep(1/M,M*K)} would return the group average for each of K tasks;
the vector \verb{c(rep(1/M1,M1*K)}, \verb{rep(-1/M2,M2*K))} would return the difference between the average within two groups of size M1 and M2, respectively, for each of K tasks;
the vector \verb{rep(rep(1/M,-1/M,0,...,0),each=V),M)} would return the difference between the first two tasks, averaged over all subjects.
}
\note{
This function requires the \code{INLA} package, which is not a CRAN package. See \url{http://www.r-inla.org/download} for easy installation instructions.
}
