% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.posterior.thetasamp.R
\name{beta.posterior.thetasamp}
\alias{beta.posterior.thetasamp}
\title{Internal function used in joint approach to group-analysis}
\usage{
beta.posterior.thetasamp(
  theta,
  spde,
  Xcros,
  Xycros,
  contrasts,
  thresholds,
  thresholds.contr,
  type,
  type.contr,
  alpha,
  alpha.contr,
  ind_beta
)
}
\arguments{
\item{theta}{A sample of theta (hyperparameters).}

\item{spde}{A SPDE object from inla.spde2.matern() function.}

\item{Xcros}{A crossproduct of design matrix.}

\item{Xycros}{A crossproduct of design matrix and BOLD y.}

\item{contrasts}{A list of vectors of length M*K specifying the contrasts of interest.  See Details for more information.}

\item{thresholds}{A vector of thresholds for activation maps.}

\item{thresholds.contr}{The list of activation thresholds for contrasts.}

\item{type}{The type of excursion function for mean beta (">", "<", "!=")}

\item{type.contr}{The list of types of excursion function for contrasts (">", "<", "!=")}

\item{alpha}{The significance level for activation}

\item{alpha.contr}{The list of activation significance levels for contrasts}

\item{ind_beta}{A vector of indices of beta.}
}
\value{
A list containing...
}
\description{
Internal function used in joint approach to group-analysis
}
\details{
The contrast vector specifies the group-level quantity of interest.  For example, the vector \code{rep(1/M,M*K)} would return the group average for each of K tasks;
the vector \verb{c(rep(1/M1,M1*K)}, \verb{rep(-1/M2,M2*K))} would return the difference between the average within two groups of size M1 and M2, respectively, for each of K tasks;
the vector \verb{rep(rep(1/M,-1/M,0,...,0),each=V),M)} would return the difference between the first two tasks, averaged over all subjects.
}
\note{
This function requires the \code{INLA} package, which is not a CRAN package. See \url{http://www.r-inla.org/download} for easy installation instructions.
}
\examples{
\dontrun{}
}
