% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.posterior.thetasamp.R
\name{beta.posterior.thetasamp}
\alias{beta.posterior.thetasamp}
\title{Internal function used in joint approach to group-analysis}
\usage{
beta.posterior.thetasamp(theta, spde, Xcros, Xycros, thresholds,
  alpha = 0.01, ind_beta)
}
\arguments{
\item{theta}{}

\item{spde}{A SPDE object from inla.spde2.matern() function.}

\item{Xcros}{A crossproduct of design matrix.}

\item{Xycros}{A crossproduct of design matrix and BOLD y.}

\item{thresholds}{A vector of thresholds for activation maps.}

\item{ind_beta}{A vector of indices of beta.}
}
\value{
A list containing...
}
\description{
Internal function used in joint approach to group-analysis
}
\examples{
\dontrun{}
}
