% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM2.R
\name{beta.posterior.thetasamp}
\alias{beta.posterior.thetasamp}
\title{}
\usage{
beta.posterior.thetasamp(
  theta,
  spde,
  Xcros,
  Xycros,
  contrasts,
  quantiles,
  excursion_type,
  gamma,
  alpha,
  nsamp_beta = 100
)
}
\arguments{
\item{theta}{}

\item{spde}{}

\item{Xcros}{}

\item{Xycros}{}

\item{contrasts}{}

\item{quantiles}{}

\item{excursion_type}{}

\item{gamma}{}

\item{alpha}{}

\item{nsamp_beta}{}
}
\value{

}
\description{

}
\section{}{
NA
}

\keyword{internal}
