% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{check_BayesGLM}
\alias{check_BayesGLM}
\title{Check arguments and packages for BayesGLM-related functions}
\usage{
check_BayesGLM(require_PARDISO = TRUE)
}
\description{
Check arguments and packages for \code{\link{BayesGLM}},
\code{\link{BayesGLM_slice}}, and \code{\link{BayesGLM_cifti}}.
}
\keyword{internal}
