% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicalGLM.R
\name{classicalGLM}
\alias{classicalGLM}
\title{Fit classical massive univariate GLM to task fMRI data}
\usage{
classicalGLM(
  data,
  mask = NULL,
  scale_BOLD = TRUE,
  scale_design = TRUE,
  avg_sessions = TRUE
)
}
\arguments{
\item{data}{A list of sessions, where each session is a list with elements
BOLD, design and nuisance.  See \code{?create.session} and \code{?is.session} for more details.
List element names represent session names.}

\item{mask}{(Optional) A length \eqn{V} logical vector indicating if each
vertex is to be included.}

\item{scale_BOLD}{Scale timeseries data so estimates represent percent signal
change? Default: \code{TRUE}. If \code{FALSE}, the data will just be
centered.}

\item{scale_design}{Scale the design matrix by dividing each column by its
maximum and then subtracting the mean? Default: \code{TRUE}.}

\item{avg_sessions}{Average estimates for betas over multiple
sessions? Default: \code{TRUE}.}
}
\value{
A list of lists containing classical GLM task activation estimates, standard error estimates, and degrees of freedom. Each list element represents a session.
}
\description{
Fit classical massive univariate GLM to task fMRI data
}
