% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicalGLM.R
\name{classicalGLM}
\alias{classicalGLM}
\title{}
\usage{
classicalGLM(
  data,
  mask = NULL,
  scale_BOLD = TRUE,
  scale_design = TRUE,
  avg_sessions = TRUE
)
}
\arguments{
\item{data}{}

\item{mask}{}

\item{scale_BOLD}{}

\item{scale_design}{}

\item{avg_sessions}{}
}
\value{

}
\description{

}
