% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{combine_sessions_Param}
\alias{combine_sessions_Param}
\title{combine_sessions}
\arguments{
\item{combine_sessions}{If multiple sessions are provided, should their data
be combined and analyzed as a single session?

If \code{TRUE} (default), the multiple sessions will be concatenated along
time after scaling and nuisance regression, but before prewhitening. If
\code{FALSE}, each session will be analyzed separately, except that a single
estimate of the AR model coefficients for prewhitening is used, estimated
across all sessions.}
}
\description{
combine_sessions
}
