% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_templateICA.R
\name{compute_mu_s}
\alias{compute_mu_s}
\title{Computes posterior mean and precision matrix of s}
\usage{
compute_mu_s(y_vec, D, Dinv_s0, R_inv, theta, P, C_diag)
}
\arguments{
\item{y_vec}{Vectorized, dimension-reduced fMRI data, grouped by locations. A vector of length QV.}

\item{D}{Sparse diagonal matrix of template standard deviations}

\item{Dinv_s0}{The inverse of D times s0_vec}

\item{R_inv}{Estimate of inverse spatial correlation matrix (sparse)}

\item{theta}{List of current parameter estimates}

\item{P}{Permutation matrix for regrouping by locations (instead of by ICs.)}

\item{C_diag}{Diagonals of residual covariance of the first level model. A vector of length Q.}
}
\value{
A list containing the posterior mean \eqn{\mu} (mu) and precision \eqn{\Omega} (Omega) of s=(s1,...,sQ), along with the supporting vector m, where \eqn{\mu = \Omega^{-1}m}.
}
\description{
Computes posterior mean and precision matrix of s
}
