% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_vertex_areas.R
\name{compute_vertex_areas}
\alias{compute_vertex_areas}
\title{Compute surface areas of each vertex in a triangular mesh.}
\usage{
compute_vertex_areas(mesh)
}
\arguments{
\item{mesh}{An \code{"inla.mesh"} object (see \code{\link{make_mesh}} for
surface data).}
}
\value{
Vector of areas
}
\description{
Compute surface areas of each vertex in a triangular mesh.
}
\note{
This function requires the \code{INLA} package, which is not a CRAN package. See \url{http://www.r-inla.org/download} for easy installation instructions.
}
