% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_utils.R
\name{create_listRcpp}
\alias{create_listRcpp}
\title{Function to prepare objects for use in Rcpp functions}
\usage{
create_listRcpp(spde)
}
\arguments{
\item{spde}{an spde object}
}
\value{
The SPDE matrices with the correct data formats
}
\description{
Function to prepare objects for use in Rcpp functions
}
\keyword{internal}
