% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spde_vol3D.R
\name{create_spde_vol3D}
\alias{create_spde_vol3D}
\title{Create SPDE for 3D volumetric data}
\usage{
create_spde_vol3D(locs, lab, value = NULL, max_dist = NULL)
}
\arguments{
\item{lab}{Labels of data points}

\item{value}{Label value to analyze}

\item{max_dist}{Maximum distance of new vertex locations from the original vertex locations}

\item{loc}{Locations of data points}
}
\value{
SPDE
}
\description{
Create SPDE for 3D volumetric data
}
