% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spde_vol3D.R
\name{create_spde_vol3D}
\alias{create_spde_vol3D}
\title{}
\usage{
create_spde_vol3D(locs, labs, lab_set = NULL)
}
\arguments{
\item{locs}{}

\item{labs}{}

\item{lab_set}{}
}
\value{

}
\description{

}
\section{}{
NA
}

