% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs_BayesGLM_cifti.R
\name{design_param_BayesGLM_cifti}
\alias{design_param_BayesGLM_cifti}
\title{design}
\arguments{
\item{design}{A numeric matrix or \code{data.frame}, or a
\code{"BayesfMRI_design"} object from \code{\link{make_design}}.}
}
\description{
design
}
