% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticICA.R
\name{diagnosticICA}
\alias{diagnosticICA}
\title{Perform diagnostic independent component analysis (ICA) using expectation-maximization (EM)}
\usage{
diagnosticICA(
  template_mean,
  template_var,
  BOLD,
  scale = TRUE,
  maxQ = NULL,
  maxiter = 100,
  epsilon = 0.001,
  verbose = TRUE
)
}
\arguments{
\item{template_mean}{(A list of G matrices, each VxL) template mean estimates for each group 1 to G}

\item{template_var}{(A list of G matrices, each VxL) template variance estimates for each group 1 to G}

\item{BOLD}{(VxT matrix) BOLD fMRI data matrix, where T is the number of volumes (time points) and V is the number of brain locations}

\item{scale}{Logical indicating whether BOLD data should be scaled by the spatial standard deviation before model fitting. If done when estimating templates, should be done here too.}

\item{maxQ}{Maximum number of ICs (template+nuisance) to identify (L <= maxQ <= T)}

\item{maxiter}{Maximum number of EM iterations}

\item{epsilon}{Smallest proportion change between iterations (e.g. .01)}

\item{verbose}{If TRUE, display progress of algorithm}
}
\value{
A list containing the posterior probabilities of group membership, the estimated independent components S (a VxL matrix), their mixing matrix A (a TxL matrix), the number of nuisance ICs estimated (Q_nuis)
}
\description{
Perform diagnostic independent component analysis (ICA) using expectation-maximization (EM)
}
