% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_diagnosticICA.R
\name{EM_diagnosticICA}
\alias{EM_diagnosticICA}
\title{EM Algorithm for Diagnostic ICA Models}
\usage{
EM_diagnosticICA(
  template_mean,
  template_var,
  BOLD,
  theta0,
  C_diag,
  maxiter = 100,
  epsilon = 0.001,
  verbose = TRUE
)
}
\arguments{
\item{template_mean}{(A list of G matrices, each VxL) template mean estimates for each group 1 to G}

\item{template_var}{(A list of G matrices, each VxL) template variance estimates for each group 1 to G}

\item{BOLD}{(VxL matrix) dimension-reduced fMRI data}

\item{theta0}{(list) initial guess at mixing matrix A (LxL)}

\item{C_diag}{(Lx1) diagonal elements of residual covariance after dimension reduction}

\item{maxiter}{maximum number of EM iterations}

\item{epsilon}{smallest proportion change between iterations (e.g. .001)}

\item{verbose}{If TRUE, display progress of algorithm}
}
\value{
A list with 5 elements: theta (list of final parameter estimates), group_probs (posterior probabilities of group membership), subICmean (estimates of subject-level ICs), subICvar (variance of subject-level ICs), and success (flag indicating convergence (\code{TRUE}) or not (\code{FALSE}))
}
\description{
EM Algorithm for Diagnostic ICA Models
}
