% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{estimate_model}
\alias{estimate_model}
\title{}
\usage{
estimate_model(
  formula,
  data,
  A,
  spde,
  prec_initial,
  num.threads = 4,
  int.strategy = "eb",
  verbose = FALSE,
  contrasts = NULL,
  lincomb = NULL
)
}
\arguments{
\item{formula}{}

\item{data}{}

\item{A}{}

\item{spde}{}

\item{prec_initial}{}

\item{num.threads}{}

\item{int.strategy}{}

\item{verbose}{}

\item{contrasts}{}

\item{lincomb}{}
}
\value{

}
\description{

}
\section{}{
NA
}

