% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM.R
\name{estimate_model}
\alias{estimate_model}
\title{Estimate INLA model}
\usage{
estimate_model(
  formula,
  data,
  A,
  spde,
  prec_initial,
  num.threads = 4,
  int.strategy = "eb",
  verbose = FALSE,
  contrasts = NULL,
  lincomb = NULL
)
}
\arguments{
\item{formula}{Formula to put into inla}

\item{data}{Dataset}

\item{A}{Large, sparse observation matrix}

\item{spde}{The spatial model, an object of class inla.spde}

\item{prec_initial}{Initial precision}

\item{num.threads}{Number of threads}

\item{int.strategy}{INLA strategy for numerical integration.  "eb" (empirical Bayes) is recommended for computational efficiency, or "ccd" for greater accuracy}

\item{verbose}{Logical indicating if should run in a verbose mode (default FALSE).}

\item{contrasts}{A list of contrast vectors to be passed to
\code{\link{inla}}.}

\item{lincomb}{A linear combinations object created with \code{inla.make.lincomb}}
}
\value{
Results from INLA
}
\description{
Estimate INLA model
}
\note{
This function requires the \code{INLA} package, which is not a CRAN package. See \url{http://www.r-inla.org/download} for easy installation instructions.
}
