% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{estimate_model}
\alias{estimate_model}
\title{Estimate INLA model}
\usage{
estimate_model(
  formula,
  data,
  A,
  spde,
  prec_initial,
  num.threads = 4,
  int.strategy = "eb",
  verbose = FALSE,
  contrasts = NULL,
  lincomb = NULL,
  inla.mode = "experimental"
)
}
\arguments{
\item{formula}{Formula to put into inla}

\item{data}{Dataset}

\item{A}{Large, sparse observation matrix}

\item{spde}{The spatial model, an object of class inla.spde}

\item{prec_initial}{Initial precision}

\item{num.threads}{The maximum number of threads to use in the inla-program
for model estimation. Default: \code{4}.}

\item{int.strategy}{INLA strategy for numerical integration.  "eb" (empirical Bayes) is recommended for computational efficiency, or "ccd" for greater accuracy}

\item{verbose}{Should INLA be run in verbose mode? Default: \code{FALSE}.}

\item{contrasts}{List of contrast vectors to be passed to \code{\link{inla}}.}

\item{lincomb}{A linear combinations object created with \code{inla.make.lincomb}}

\item{inla.mode}{specifies the INLA mode to use (defaults to experimental)}
}
\value{
Results from INLA
}
\description{
Estimate INLA model
}
\section{INLA Requirement}{

This function requires the \code{INLA} package, which is not a CRAN package.
See \url{https://www.r-inla.org/download-install} for easy installation instructions.
}

