% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.R
\name{estimate_template.cifti}
\alias{estimate_template.cifti}
\title{Estimate Template for Template or Diagnostic ICA}
\usage{
estimate_template.cifti(
  cifti_fnames,
  cifti_fnames2 = NULL,
  GICA_fname,
  inds = NULL,
  scale = TRUE,
  brainstructures = c("left", "right"),
  verbose = TRUE
)
}
\arguments{
\item{cifti_fnames}{Vector of file paths of CIFTI-format fMRI timeseries (*.dtseries.nii) for template estimation}

\item{cifti_fnames2}{Vector of file paths of "retest" CIFTI-format fMRI timeseries (*.dtseries.nii) for template estimation.  Must be from the same subjects and in the same order as cifti_fnames.  If none specified, will create pseudo test-retest data from single session.}

\item{GICA_fname}{File path of CIFTI-format group ICA maps (ending in .d*.nii)}

\item{inds}{Indicators of which group ICs to include in template. If NULL, use all group ICs.}

\item{scale}{Logical indicating whether BOLD data should be scaled by the spatial standard deviation before template estimation.}

\item{brainstructures}{Vector of brainstructures to include ('left','right','surface').  Default is c('left','right').}

\item{verbose}{If TRUE, display progress updates}
}
\value{

}
\description{
Estimate Template for Template or Diagnostic ICA
}
