% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursions.R
\name{excursions.inla.no.spurious}
\alias{excursions.inla.no.spurious}
\title{Excursion sets with INLA and removal}
\usage{
excursions.inla.no.spurious(
  result.inla,
  stack,
  name = NULL,
  tag = NULL,
  u,
  alpha = 1,
  type,
  n.iter = 10000,
  F.limit,
  u.link = FALSE,
  method,
  ind = NULL,
  max.size,
  verbose = FALSE,
  max.threads = 0,
  seed = NULL,
  mesh,
  area.limit,
  use.continuous = FALSE,
  plot.progress = FALSE
)
}
\arguments{
\item{result.inla}{Result object from INLA call.}

\item{stack}{The stack object used in the INLA call.}

\item{name}{The name of the component for which to do the calculation. This argument should
only be used if a stack object is not provided, use the tag argument otherwise.}

\item{tag}{The tag of the component in the stack for which to do the calculation. This argument
should only be used if a stack object is provided, use the name argument otherwise.}

\item{u}{Excursion or contour level.}

\item{alpha}{Error probability for the excursion set.}

\item{type}{Type of region:
\itemize{
\item{'>' }{positive excursion region}
\item{'<' }{negative excursion region}
\item{'!=' }{contour avoiding region}
\item{'=' }{contour credibility region}}}

\item{n.iter}{Number or iterations in the MC sampler that is used for approximating probabilities. The default value is 10000.}

\item{F.limit}{The limit value for the computation of the F function. F is set to NA for all nodes where F<1-F.limit. Default is F.limit = \code{alpha}.}

\item{u.link}{If u.link is TRUE, u is assumed to be in the scale of the data and is then transformed to the scale of the linear predictor (default FALSE).}

\item{method}{Method for handeling the latent Gaussian structure:
\itemize{
\item{'EB' }{Empirical Bayes (default)}
\item{'QC' }{Quantile correction, rho must be provided if QC is used.}}}

\item{ind}{Indices of the nodes that should be analysed (optional).}

\item{max.size}{Maximum number of nodes to include in the set of interest (optional).}

\item{verbose}{Should occasional updates be printed? Default: \code{FALSE}.}

\item{max.threads}{The maximum number of threads to use in the inla-program
for model estimation. \code{0} (default) will use the maximum number of
threads allowed by the system.}

\item{seed}{Random seed (optional). Default: \code{NULL}.}

\item{mesh}{The mesh on which the model is defined.}

\item{area.limit}{Positive number. All connected excursion sets with an area smaller than this
number are removed.}

\item{use.continuous}{Logical parameter indicating whether the areas of the excursion sets
should be calculated using the \code{continuous} function in \code{excursions}. If FALSE, the
function uses the approximation that the area for each node is the integral of the FEM basis function \eqn{\phi_i}.}

\item{plot.progress}{Logical parameter that indicates whether the results should be plotted.}
}
\value{
If \code{use.continuous = FALSE}, an item of class \code{excurobj}.
Otherwise a list with the same elements as the output of \code{continuous}.
}
\description{
Calculation of excursion sets with the removal of small areas
}
\details{
Function similar to the excursions.inla function, but which also takes a mesh as input
and computes modified excursion sets where all regions with areas smaller than area.limit are
removed in the internal calculations.
}
