% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursions.R
\name{excursions.no.spurious}
\alias{excursions.no.spurious}
\title{}
\usage{
excursions.no.spurious(
  alpha,
  u,
  mu,
  Q,
  type,
  n.iter = 10000,
  Q.chol,
  F.limit,
  vars,
  rho = NULL,
  method = "EB",
  ind,
  max.size,
  verbose = 0,
  max.threads = 0,
  seed,
  area.limit,
  use.continuous = FALSE,
  factor = 0,
  plot.progress = FALSE,
  mesh
)
}
\arguments{
\item{alpha}{}

\item{u}{}

\item{mu}{}

\item{Q}{}

\item{type}{}

\item{n.iter}{}

\item{Q.chol}{}

\item{F.limit}{}

\item{vars}{}

\item{rho}{}

\item{method}{}

\item{ind}{}

\item{max.size}{}

\item{verbose}{}

\item{max.threads}{}

\item{seed}{}

\item{area.limit}{}

\item{use.continuous}{}

\item{factor}{}

\item{plot.progress}{}

\item{mesh}{}
}
\value{

}
\description{

}
