% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_estimates.R
\name{extract_estimates}
\alias{extract_estimates}
\title{Extract Estimates}
\usage{
extract_estimates(object, mask, session_names)
}
\arguments{
\item{object}{An object of class ‘"inla"’, a result of a call to inla}

\item{mask}{Logical vector used to map beta estimates back to whole-brain field}

\item{session_names}{Name of session}
}
\value{
Estimates from inla model
}
\description{
Extract Estimates
}
\examples{
\dontrun{}
}
