% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{extract_estimates}
\alias{extract_estimates}
\title{}
\usage{
extract_estimates(object, session_names, mask = NULL, stat = "mean")
}
\arguments{
\item{object}{}

\item{session_names}{}

\item{mask}{}

\item{stat}{}
}
\value{

}
\description{

}
\keyword{internal}
