% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F.logwt.R
\name{F.logwt}
\alias{F.logwt}
\title{Internal function used in joint approach to group-analysis}
\usage{
F.logwt(theta, spde, mu.theta, Q.theta, M)
}
\arguments{
\item{theta}{A vector of hyperparameter values at which to compute the posterior log density}

\item{spde}{A SPDE object from inla.spde2.matern() function.}

\item{mu.theta}{Posterior mean from combined subject-level models.}

\item{Q.theta}{Posterior precision matrix from combined subject-level models.}

\item{M}{Number of subjects}
}
\value{
A list containing...
}
\description{
Internal function used in joint approach to group-analysis
}
\examples{
\dontrun{}
}
