% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM2_utils.R
\name{F.logwt}
\alias{F.logwt}
\title{F logwt}
\usage{
F.logwt(theta, spde, mu_theta, Q_theta, nN)
}
\arguments{
\item{theta}{A vector of hyperparameter values at which to compute the posterior log density}

\item{spde}{A SPDE object from inla.spde2.matern() function, determines prior precision matrix}

\item{mu_theta}{Posterior mean from combined subject-level models.}

\item{Q_theta}{Posterior precision matrix from combined subject-level models.}

\item{nN}{Number of subjects}
}
\value{
The prior density
}
\description{
Internal function used in joint approach to group-analysis for combining across models
}
\section{INLA Requirement}{

This function requires the \code{INLA} package, which is not a CRAN package.
See \url{https://www.r-inla.org/download-install} for easy installation instructions.
}

\keyword{internal}
