% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursions.R
\name{find.smallest.activation}
\alias{find.smallest.activation}
\title{}
\usage{
find.smallest.activation(res, mesh, area.limit, factor, area.el)
}
\arguments{
\item{res}{}

\item{mesh}{}

\item{area.limit}{}

\item{factor}{}

\item{area.el}{}
}
\description{

}
\details{

}
\keyword{internal}
