% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_densities.R
\name{get_posterior_densities}
\alias{get_posterior_densities}
\title{Extracts posterior density estimates for hyperparameters}
\usage{
get_posterior_densities(result, spde, beta_names)
}
\arguments{
\item{result}{Result}

\item{spde}{SPDE}

\item{beta_names}{Names of covariates}
}
\value{
Long-form data frame
}
\description{
Extracts posterior density estimates for hyperparameters
}
\examples{
\dontrun{}
}
