% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM.R
\name{get_posterior_densities}
\alias{get_posterior_densities}
\title{}
\usage{
get_posterior_densities(object, spde, beta_names = NULL)
}
\arguments{
\item{object}{}

\item{spde}{}

\item{beta_names}{}
}
\value{

}
\description{

}
\note{

}
