% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{get_posterior_densities}
\alias{get_posterior_densities}
\title{Extracts posterior density estimates for hyperparameters}
\usage{
get_posterior_densities(INLA_model_obj, spde, task_names)
}
\arguments{
\item{INLA_model_obj}{An object of class \code{"inla"}, a result of a call to
\code{inla()}}

\item{spde}{The model used for the latent fields in the \code{inla()} call,
an object of class \code{"inla.spde"}}

\item{task_names}{Descriptive names of model regressors (tasks).}
}
\value{
Long-form data frame containing posterior densities for the
hyperparameters associated with each latent field
}
\description{
Extracts posterior density estimates for hyperparameters
}
\section{INLA Requirement}{

This function requires the \code{INLA} package, which is not a CRAN package.
See \url{https://www.r-inla.org/download-install} for easy installation instructions.
}

\keyword{internal}
