% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_densities_vol3D.R
\name{get_posterior_densities_vol3D}
\alias{get_posterior_densities_vol3D}
\title{Extracts posterior density estimates for hyperparameters}
\usage{
get_posterior_densities_vol3D(object, spde)
}
\arguments{
\item{object}{An object of class ‘"inla"’, a result of a call to \code{inla()}}

\item{spde}{The model used for the latent fields in the \code{inla()} call, an object of class ‘"inla.spde"’}
}
\value{
Long-form data frame containing posterior densities for the hyperparameters associated with each latent field
}
\description{
Extracts posterior density estimates for hyperparameters
}
\examples{
\dontrun{}
}
