% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_vol3D.R
\name{get_posterior_densities_vol3D}
\alias{get_posterior_densities_vol3D}
\title{Extracts posterior density estimates for hyperparameters}
\usage{
get_posterior_densities_vol3D(object, spde)
}
\arguments{
\item{object}{An object of class \code{"inla"}, a result of a call to
\code{inla()}}

\item{spde}{The model used for the latent fields in the \code{inla()} call,
an object of class \code{"inla.spde"}}
}
\value{
Long-form data frame containing posterior densities for the
hyperparameters associated with each latent field
}
\description{
Extracts posterior density estimates for hyperparameters
}
\section{INLA Requirement}{

This function requires the \code{INLA} package, which is not a CRAN package.
See \url{https://www.r-inla.org/download-install} for easy installation instructions.
}

