% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_classical.R
\name{GLM_classical}
\alias{GLM_classical}
\title{Classical GLM}
\usage{
GLM_classical(
  BOLD,
  design,
  nK2,
  nV_D,
  field_names,
  design_type,
  valid_cols,
  nT,
  do_pw,
  compute_SE = TRUE
)
}
\arguments{
\item{BOLD}{BOLD timeseries in vector form (TVx1), result of \code{sparse_and_PW}}

\item{design}{List of large sparse design matrices (each TVxV), one per regressor, result of \code{sparse_and_PW}}

\item{nK2, nV_D, field_names, design_type}{See \code{BayesGLM_fun}.}

\item{valid_cols, nT}{See \code{BayesGLM_fun}.}

\item{do_pw}{Has prewhitening been performed on the data and design?}

\item{compute_SE}{Compute SE of model coefficients?}
}
\value{
A list of results
}
\description{
Classical GLM for BayesGLM_fun (internal function)
}
\keyword{internal}
