% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_multi.R
\name{GLM_multi}
\alias{GLM_multi}
\title{GLM multi}
\usage{
GLM_multi(y, X, X2, Xc = NULL, verbose = TRUE)
}
\arguments{
\item{y, X, X2}{BOLD, design, nuisance}

\item{Xc}{(Optional) canonical design matrix}

\item{verbose}{verbose?}
}
\value{
Results for GLM multi
}
\description{
GLM multi
}
\keyword{internal}
