% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{Calculate gradient}
\usage{
gradient(x)
}
\arguments{
\item{x}{If x is a matrix, compute gradient of each column of x. Otherwise, temporarily transform to a matrix with 1 column}
}
\value{
Gradient of x
}
\description{
Calculate gradient
}
