% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_HRFs.R
\name{HRF}
\alias{HRF}
\title{Calculate the canonical (double-gamma) HRF}
\usage{
HRF(t, a1 = 6, b1 = 0.9, a2 = 12, b2 = 0.9, c = 0.35)
}
\arguments{
\item{t}{time vector}

\item{a1}{delay of response, default is 6}

\item{b1}{response dispersion, default is 0.9}

\item{a2}{delay of undershoot, default is 12}

\item{b2}{dispersion of undershoot, default is 0.9}

\item{c}{scale of undershoot, default is 0.35}
}
\value{
HRF vector corresponding to time
}
\description{
Calculate the canonical (double-gamma) HRF
}
\examples{
downsample <- 100
HRF(seq(0, 30, by=1/downsample))

}
