% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_activations.R
\name{id_activations.classical}
\alias{id_activations.classical}
\title{Identification of areas of activation in a General Linear Model using classical methods}
\usage{
id_activations.classical(model_obj, alpha = 0.05)
}
\arguments{
\item{model_obj}{A \code{BayesGLM_cifti} or \code{BayesGLM_slice} object}

\item{alpha}{A significance level to be used in both the family-wise error
rate (FWER) or false discovery rate (FDR) multiple correction methods of
determining significance.}
}
\value{
A list of length equal to the number of hemispheres of model fit
data. In each list element, there will be two matrices corresponding to the
0-1 activation status for the model coefficients.
}
\description{
Identification of areas of activation in a General Linear Model using classical methods
}
