% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_activations.R
\name{id_activations.em}
\alias{id_activations.em}
\title{Identify activations using joint posterior probabilities with EM results}
\usage{
id_activations.em(
  model_obj,
  field_names = NULL,
  session_name = NULL,
  alpha = 0.05,
  threshold,
  area.limit = NULL
)
}
\arguments{
\item{model_obj}{An object of class ‘"BayesGLM"’, a result of a call to BayesGLMEM}

\item{field_names}{Name of latent field or vector of names on which to identify activations.  By default, analyze all tasks.}

\item{session_name}{(character) The name of the session that should be examined.
If \code{NULL} (default), the average across all sessions is used.}

\item{alpha}{Significance level (e.g. 0.05)}

\item{threshold}{Activation threshold (e.g. 1 for 1 percent signal change if scale=TRUE in model estimation)}

\item{area.limit}{Below this value, activations will be considered spurious.  If NULL, no limit.}
}
\value{
A list with two elements: \code{active}, which gives a matrix of zeros
and ones of the same dimension as \code{model_obj$beta_estimates${session_name}},
and \code{excur_result}, an object of class \code{"excurobj"} (see \code{\link{excursions.inla}} for
more information).
}
\description{
Identifies areas of activation given an activation threshold and significance
level using joint posterior probabilities
}
\details{
For a given latent field, identifies locations that exceed a certain activation
threshold (e.g. 1 percent signal change) at a given significance level, based on the joint
posterior distribution of the latent field.
}
