% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_activations.R
\name{id_activations.posterior}
\alias{id_activations.posterior}
\title{}
\usage{
id_activations.posterior(
  model_obj,
  field_names = NULL,
  session_name = NULL,
  alpha = 0.05,
  threshold,
  area.limit = NULL,
  excur_method = c("EB", "QC")
)
}
\arguments{
\item{model_obj}{}

\item{field_names}{}

\item{session_name}{}

\item{alpha}{}

\item{threshold}{}

\item{area.limit}{}

\item{excur_method}{}
}
\value{

}
\description{

}
\details{

}
