% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_activations.R
\name{id_activations}
\alias{id_activations}
\title{}
\usage{
id_activations(
  model_obj,
  method = c("posterior", "2means"),
  field_name = NULL,
  threshold = NULL,
  alpha = NULL,
  area.limit = NULL,
  type = c("point", "sequential"),
  n_sample = NULL
)
}
\arguments{
\item{model_obj}{}

\item{method}{}

\item{field_name}{}

\item{threshold}{}

\item{alpha}{}

\item{area.limit}{}

\item{type}{}

\item{n_sample}{}
}
\value{
\verb{}\verb{}\verb{}\verb{}
}
\description{

}
\details{

}
