% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_activations.R
\name{id_activations_cifti}
\alias{id_activations_cifti}
\title{Identify Task Activations}
\usage{
id_activations_cifti(
  model_obj,
  field_names = NULL,
  session_name = NULL,
  alpha = 0.05,
  method = c("Bayesian", "classical"),
  threshold = NULL,
  correction = c("FWER", "FDR", "none"),
  verbose = TRUE
)
}
\arguments{
\item{model_obj}{Result of \code{BayesGLM_cifti} model call, of class
\code{"BayesGLM_cifti"}}

\item{field_names}{The name(s) of the latent field(s) on which to
identify activations. If \code{NULL} (default), analyze all tasks.}

\item{session_name}{The name of the session to examine.
If \code{NULL} (default), the average across all sessions is used.}

\item{alpha}{Significance level. Default: \code{0.05}.}

\item{method}{\code{"Bayesian"} (default) or \code{"classical"}.}

\item{threshold}{Activation threshold, for example \code{1} for 1\\% signal
change if \code{scale_BOLD=="mean"} during model estimation. Setting a
threshold is required for the Bayesian method; \code{NULL} (default) will
use a threshold of zero for the classical method.}

\item{correction}{For the classical method only: Type of multiple comparisons
correction: \code{"FWER"} (Bonferroni correction, the default), \code{"FDR"}
(Benjamini Hochberg), or \code{"none"}.
depending on the method that should be used to find the excursions set. Note that to ID
activations for averages across sessions, the method chosen must be \code{EB}. The difference
in the methods is that the \code{EB} method assumes Gaussian posterior distributions for the parameters.}

\item{verbose}{Print progress updates? Default: \code{TRUE}.}
}
\description{
Identify areas of activation for each task
}
