% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_activations.R
\name{id_activations_cifti}
\alias{id_activations_cifti}
\title{Identify Task Activations}
\usage{
id_activations_cifti(
  model_obj,
  field_name = NULL,
  threshold = NULL,
  alpha = 0.05,
  area.limit = NULL
)
}
\arguments{
\item{field_name}{Name of latent field, or vector of names, on which to identify activations}

\item{threshold}{activation threshold (e.g. 0.01 for 1 percent signal change)}

\item{alpha}{Significance level (e.g. 0.05)}

\item{area.limit}{Below this value, activations will be considered spurious.  If NULL (default), no limit.}

\item{result}{Result of BayesGLM_cifti model call (of class BayesGLM_cifti)}
}
\description{
Identify areas of activation for each task
}
