% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_templateICA.R
\name{loglik_kappa_est}
\alias{loglik_kappa_est}
\title{}
\usage{
loglik_kappa_est(par, delta, D_diag, mesh, C1 = 1/(4 * pi), Q = NULL)
}
\arguments{
\item{par}{}

\item{delta}{}

\item{D_diag}{}

\item{mesh}{\verb{}}

\item{C1}{}

\item{Q}{}
}
\value{

}
\description{

}
