% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_A_mat.R
\name{make_A_mat}
\alias{make_A_mat}
\title{Make A matrix}
\usage{
make_A_mat(surf, surf_rs)
}
\arguments{
\item{surf}{The full-resolution \code{"surf"} object.}

\item{surf_rs}{The downsampled \code{"surf"} object.}
}
\value{
The A matrix
}
\description{
Make the A matrix for downsampling surface mesh data to a lower resolution.
}
\keyword{internal}
