% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{make_data_list}
\alias{make_data_list}
\title{}
\usage{
make_data_list(y, X, betas, repls)
}
\arguments{
\item{y}{}

\item{X}{\verb{}}

\item{betas}{}

\item{repls}{}
}
\value{

}
\description{

}
\keyword{internal}
