% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_Bayes_utils.R
\name{make_data_list}
\alias{make_data_list}
\title{Make data list for \code{estimate_model}}
\usage{
make_data_list(y, X, betas, repls)
}
\arguments{
\item{y}{Vectorized BOLD data (all voxels, sessions, etc.)}

\item{X}{List (length = number of sessions) of sparse design matrices size TVxVK from each session, each created using \code{sparse_and_PW()}}

\item{betas}{List (length = number of fields) of bbeta objects from organize_replicates}

\item{repls}{List (length = number of fields) of repl objects from organize_replicates}
}
\value{
List
}
\description{
Make data list to be passed to \code{estimate_model}
}
\keyword{internal}
