% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{make_formula}
\alias{make_formula}
\title{}
\usage{
make_formula(beta_names, repl_names, hyper_initial = NULL)
}
\arguments{
\item{beta_names}{}

\item{repl_names}{}

\item{hyper_initial}{}
}
\value{
\verb{}
}
\description{

}
\keyword{internal}
