% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_HRFs.R
\name{make_HRFs}
\alias{make_HRFs}
\title{Make HRFs}
\usage{
make_HRFs(onsets, TR, duration, downsample = 100, deriv = 0)
}
\arguments{
\item{onsets}{A matrix of onsets (first column) and durations (second column)
for each task in SECONDS (set duration to zero for event related design),
organized as a list where each element of the list corresponds to one task.
Names of list indicate task names.}

\item{TR}{Temporal resolution of fMRI data, in SECONDS.}

\item{duration}{Length of fMRI timeseries, in SCANS.}

\item{downsample}{Downsample factor for convolving stimulus boxcar or stick
function with canonical HRF. Default: \code{100}.}

\item{deriv}{This can take the value of 0, 1, or 2, and will use the HRF
function, the first derivative of the HRF, or the second derivative of the
HRF, respectively.}
}
\value{
Design matrix containing one HRF column for each task
}
\description{
Create HRF design matrix columns from onsets and durations
}
\examples{
onsets <- list(taskA=cbind(c(2,17,23),4)) # one task, 3 four sec-long stimuli
TR <- .72 # .72 seconds per volume, or (1/.72) Hz
duration <- 300 # session is 300 volumes long (300*.72 seconds long)
make_HRFs(onsets, TR, duration)

}
