% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_HRFs.R
\name{make_HRFs}
\alias{make_HRFs}
\title{Make HRFs}
\usage{
make_HRFs(
  onsets,
  nTime,
  TR,
  dHRF = c(1, 0, 2),
  upsample = 100,
  a1 = 6,
  b1 = 1,
  a2 = 16/6 * a1 * sqrt(b1),
  b2 = b1,
  c = 1/6
)
}
\arguments{
\item{onsets, nTime, TR, dHRF}{See \code{\link{make_design}}.}

\item{upsample}{Upsample factor for convolving stimulus boxcar or stick
function with canonical HRF. Default: \code{100}.}

\item{a1}{(Optional) delay of response. Default: \code{6}}

\item{b1}{Response dispersion. Default: \code{1}}

\item{a2}{Delay of undershoot. Default: \code{16/6 * a1 * sqrt(b1) == 16}}

\item{b2}{Dispersion of undershoot. Default: \code{b1 == 1}}

\item{c}{Scale of undershoot. Default: \code{1/6}}
}
\value{
A session-length list. Each entry is a sublist with components:
\describe{
\item{HRF}{The upsampled HRFs.}
\item{stimulus}{The task stimulus.}
\item{design}{The corresponding design matrix.}
}
}
\description{
Make HRFs from onsets. This is a helper function to \code{\link{make_design}}.
}
\examples{
# onsets: 1 session, 1 task, 3 stimuli, each 4 sec. long
onsets <- list(session1=list(taskA=cbind(c(2,17,23),4))) 
TR <- .72 # .72 seconds per volume, or (1/.72) Hz
nTime <- 300 # session has 300 volumes. Session duration is 300*.72 sec.
make_HRFs(onsets, nTime, TR)

}
