% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mesh.R
\name{make_mesh}
\alias{make_mesh}
\title{Make Mesh}
\usage{
make_mesh(vertices, faces, mask = NULL)
}
\arguments{
\item{vertices}{Matrix of vertices}

\item{faces}{Matrix of faces}

\item{mask}{Mask to be applied to mesh. If none given, full mesh is used.}
}
\value{
Triangular mesh from matrices and vertices
}
\description{
Make triangular mesh from faces and vertices. If mask is given, then mesh is masked.
}
