% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_vertices_faces.R
\name{mask_vertices_faces}
\alias{mask_vertices_faces}
\title{}
\usage{
mask_vertices_faces(vertices, faces, mask)
}
\arguments{
\item{vertices}{}

\item{faces}{}

\item{mask}{}
}
\value{

}
\description{

}
