% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary.R
\name{mask_with_boundary}
\alias{mask_with_boundary}
\title{Apply Mask With Boundary To Mesh}
\usage{
mask_with_boundary(
  vertices,
  faces,
  mask,
  width1 = 4,
  k1 = 2,
  width2 = 6,
  k2 = 3
)
}
\arguments{
\item{faces}{a V x 3 matrix of integers. Each row defines a face by the index
of three vertices.}

\item{mask}{a length-V logical vector. Each entry corresponds to the vertex
with the same index. TRUE indicates vertices within the input mask.}

\item{width1, width2}{the width of the middle/outer region. All vertices in the middle/outer region
are between 1 and \code{width1} edges away from the closest vertex in \code{mask}/middle region.}

\item{k1, k2}{roughly, the triangle size multiplier. Every \code{k1}/\code{k2} vertex within
every \code{k1}/\code{k2} layer (beginning with the innermost layer) will be retained;
the rest will be discarded. If the mesh originally has triangles of regular
size, the sides of the triangles in the middle/outer region will be about
\code{k1}/\code{k2} as long.}
}
\value{
A new mesh (list with components "vertices" and "faces")
}
\description{
Make a boundary around a mask with two levels of decimation, and apply to a mask.
}
\details{

}
