% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs_BayesGLM.R
\name{nuisance_Param_BayesGLM}
\alias{nuisance_Param_BayesGLM}
\title{nuisance}
\arguments{
\item{nuisance}{(Optional) A \eqn{T \times N_{nuis}} matrix of nuisance signals,
where \eqn{T} is the number of timepoints and \eqn{N} is the number of
nuisance signals, or a list of these for multi-session analysis. Nuisance
signals are regressed from the fMRI data and design matrix prior to GLM
computation. Nuisance signals can include motion regressors, HRF derivatives
not being modeled as tasks, and other sources of noise.

Detrending/high-pass filtering is accomplished by adding DCT bases to the
nuisance matrix; see the parameters \code{hpf} and \code{DCT}.

Do not add spike regressors for scrubbing to the \code{nuisance} matrix.
Rather, provide these in \code{scrub} so that their corresponding timepoints
are also removed from the BOLD data after nuisance regression.}
}
\description{
nuisance
}
