% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{num.threads_Param}
\alias{num.threads_Param}
\title{num.threads}
\arguments{
\item{num.threads}{The maximum number of threads to use for parallel
computations: prewhitening parameter estimation, and the inla-program model
estimation. Default: \code{4}. Note that parallel prewhitening requires the
\code{parallel} package.}
}
\description{
num.threads
}
