% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_replicates.R
\name{organize_replicates}
\alias{organize_replicates}
\title{Organize replicates}
\usage{
organize_replicates(n_sess, n_task, mesh = NULL)
}
\arguments{
\item{n_sess}{The number of sessions sharing hyperparameters (can be different tasks)}

\item{n_task}{Number of regressors or tasks}

\item{mesh}{The mesh for the data}
}
\value{
replicates vector and betas for sessions
}
\description{
Organize replicates
}
\examples{
\dontrun{}
}
