% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_replicates.R
\name{organize_replicates}
\alias{organize_replicates}
\title{}
\usage{
organize_replicates(n_sess, n_task, mesh)
}
\arguments{
\item{n_sess}{}

\item{n_task}{}

\item{mesh}{}
}
\value{

}
\description{

}
\details{

}
