% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_replicates.R
\name{organize_replicates}
\alias{organize_replicates}
\title{Organize replicates}
\usage{
organize_replicates(n_sess, field_names, n_mesh)
}
\arguments{
\item{n_sess}{The number of sessions sharing hyperparameters (can be different fields)}

\item{field_names}{Vector of names for each field}

\item{n_mesh}{Number of mesh locations}
}
\value{
replicates vector and betas for sessions
}
\description{
beta and repl vectors are of length \eqn{n_mesh \times n_sess \times n_field}.
The ith repl vector is an indicator vector for the cells corresponding to the ith column of x.
The ith beta vector contains data indices (e.g. 1,...,V) in the cells corresponding to the ith column of x.
}
\keyword{internal}
