% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_replicates.R
\name{organize_replicates}
\alias{organize_replicates}
\title{Organize replicates}
\usage{
organize_replicates(n_sess, task_names, mesh)
}
\arguments{
\item{n_sess}{The number of sessions sharing hyperparameters (can be different tasks)}

\item{task_names}{Vector of names for each task}

\item{mesh}{An \code{"inla.mesh"} object (see \code{\link{make_mesh}} for
surface data).}
}
\value{
replicates vector and betas for sessions
}
\description{
beta and repl vectors are of length \eqn{nvox \times n_sess \times n_task}.
The ith repl vector is an indicator vector for the cells corresponding to the ith column of x.
The ith beta vector contains data indices (e.g. 1,...,V) in the cells corresponding to the ith column of x.
}
\keyword{internal}
