% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R, R/util_RED.R
\name{plot.BayesfMRI.spde}
\alias{plot.BayesfMRI.spde}
\title{Plot BayesfMRI.spde objects}
\usage{
plot.BayesfMRI.spde(object, colors = NULL, values = NULL, alpha = 0.5)

plot.BayesfMRI.spde(object, colors = NULL, values = NULL, alpha = 0.5)
}
\arguments{
\item{object}{Object of class BayesfMRI.spde (see \code{help(create_spde_vol3D)})}

\item{colors}{(Optional) Vector of colors to represent each region.}

\item{alpha}{Transparency level.}
}
\value{



}
\description{
Plot BayesfMRI.spde objects

Plot BayesfMRI.spde objects
}
