% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R, R/util_RED.R
\name{plot.BayesGLM}
\alias{plot.BayesGLM}
\title{Plot BayesGLM objects}
\usage{
\method{plot}{BayesGLM}(object, session_name = NULL, pal = NULL, ...)

\method{plot}{BayesGLM}(object, session_name = NULL, pal = NULL, ...)
}
\arguments{
\item{object}{an object of class "BayesGLM"}

\item{session_name}{NULL if BayesGLM object contains a single session; otherwise, the name of the session whose estimates to plot}

\item{pal}{If NULL, viridis palette with 64 colors will be used.  Otherwise, specify a vector of color names.}
}
\description{
Summary method for class "BayesGLM"

Summary method for class "BayesGLM"
}
