% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BayesGLM.R
\name{plot.BayesGLM}
\alias{plot.BayesGLM}
\title{Plot results from a BayesfMRI object}
\usage{
\method{plot}{BayesGLM}(BayesGLM_object, mask, session_name = NULL)
}
\arguments{
\item{BayesGLM_object}{An object of class "BayesfMRI"}

\item{mask}{The mask used to convert from cortical surface to volumetric data}

\item{session_name}{The name of the session to plot the results from (defaults to the first session)}
}
\value{
Currently a list of final coefficient array images. This will change in future versions.
}
\description{
Plot results from a BayesfMRI object
}
