% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.BayesGLM_cifti}
\alias{plot.BayesGLM_cifti}
\title{}
\usage{
\method{plot}{BayesGLM_cifti}(x, session = NULL, method = NULL, idx = NULL, zlim = c(-1, 1), ...)
}
\arguments{
\item{x}{}

\item{session}{}

\item{method}{}

\item{idx}{}

\item{zlim}{}

\item{...}{}
}
\description{

}
