% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.BGLM2}
\alias{plot.BGLM2}
\title{S3 method: use \code{\link[ciftiTools]{view_xifti}} to plot a \code{"BGLM2"} object}
\usage{
\method{plot}{BGLM2}(x, idx = NULL, stat = c("contrasts", "activations"), zlim = c(-1, 1), ...)
}
\arguments{
\item{x}{An object of class "BGLM2"}

\item{idx}{Which contrast should be plotted? Give the numeric indices or the
names. \code{NULL} (default) will show all contrasts. This argument
overrides the \code{idx} argument to \code{\link[ciftiTools]{view_xifti}}.}

\item{stat}{Estimates of the \code{"contrasts"} (default), or their
thresholded \code{"activations"}.}

\item{zlim}{Overrides the \code{zlim} argument for
\code{\link[ciftiTools]{view_xifti}}. Default: \code{c(-1, 1)}.}

\item{...}{Additional arguments to \code{\link[ciftiTools]{view_xifti}}}
}
\value{
Result of the call to \code{ciftiTools::view_cifti}.
}
\description{
S3 method: use \code{\link[ciftiTools]{view_xifti}} to plot a \code{"BGLM2"} object
}
