% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prewhitening.R
\name{prewhiten.v}
\alias{prewhiten.v}
\title{Prewhitening for a single time series}
\usage{
prewhiten.v(AR_coeffs, ntime, AR_var = 1)
}
\arguments{
\item{AR_coeffs}{A vector of AR(p) coefficients for an order p autoregressive
prewhitening}

\item{ntime}{(scalar) The length of the time series to be prewhitened}

\item{AR_var}{(scalar) The variance of the time series to include a
normalization term to make the prewhitened series have variance equal to 1.
If none is provided, then the default is 1, resulting in no normalization.}
}
\value{
A sparse matrix with the class \code{dsCMatrix} with \code{ntime}
rows and columns that can be postmultiplied by the original time series
in order to perform prewhitening.
}
\description{
Prewhitening for a single time series
}
\examples{
\dontrun{
x <- arima.sim(list(ar = c(.3,.2,.1)), n = 100)
acf(x)
ar_model <- ar.yw(x,order.max = 6, aic = F)
pw_mat <- prewhiten.v(AR_coeffs = ar_model$ar,
                      ntime = length(x),
                      AR_var = ar_model$var.pred)
y <- as.vector(pw_mat \%*\% as.matrix(x))
acf(y)
plot(x, type = 'l')
lines(y, col = 'red')
}
}
