% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prewhitening.R
\name{prewhiten_cifti}
\alias{prewhiten_cifti}
\title{}
\usage{
prewhiten_cifti(
  data,
  mask = NULL,
  scale_BOLD = TRUE,
  scale_design = TRUE,
  ar_order = 6,
  ar_smooth = 5,
  cifti_data,
  hemisphere,
  num.threads = NULL
)
}
\arguments{
\item{data}{}

\item{mask}{}

\item{scale_BOLD}{}

\item{scale_design}{}

\item{ar_order}{}

\item{ar_smooth}{}

\item{cifti_data}{}

\item{hemisphere}{}

\item{num.threads}{}
}
\value{

}
\description{

}
