% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prewhitening.R
\name{prewhiten_cifti}
\alias{prewhiten_cifti}
\title{Prewhiten cifti session data}
\usage{
prewhiten_cifti(
  data,
  scale_BOLD = TRUE,
  scale_design = TRUE,
  ar_order = 6,
  surface_sigma = NULL,
  cifti_data,
  wb_path = NULL,
  hemisphere = NULL
)
}
\arguments{
\item{data}{List of sessions (see \code{is.session} and \code{is.session_pw})}

\item{scale_BOLD}{(logical) Should the BOLD response be scaled? (Default is TRUE)}

\item{scale_design}{(logical) Should the design matrix be scaled? (Default is TRUE)}

\item{ar_order}{Order of the AR used to prewhiten the data at each location}

\item{surface_sigma}{range parameter for smoothing. Remember that
sigma = \code{FWHM / (2*sqrt(2*log(2))})}

\item{cifti_data}{A \code{xifti} object used to map the AR coefficient
estimates onto the surface mesh for smoothing.}

\item{wb_path}{Relative path to the Human Connectome Project workbench}

\item{hemisphere}{'left' or 'right'}
}
\value{
The prewhitened data (in a list), the smoothed, averaged AR
coefficient estimates used in the prewhitening, the smoothed, average
residual variance after prewhitening, and the value given for \code{ar_order}.
}
\description{
Prewhiten cifti session data
}
