% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prewhitening.R
\name{prewhiten_cifti}
\alias{prewhiten_cifti}
\title{Prewhiten cifti session data}
\usage{
prewhiten_cifti(
  data,
  mask = NULL,
  scale_BOLD = c("auto", "mean", "sd", "none"),
  scale_design = TRUE,
  ar_order = 6,
  ar_smooth = 5,
  cifti_data,
  hemisphere,
  num.threads = NULL
)
}
\arguments{
\item{data}{List of sessions (see \code{is.session} and \code{is.session_pw})}

\item{mask}{(Optional) A length \eqn{V} logical vector indicating if each
vertex is to be included.}

\item{scale_BOLD}{Option for scaling the BOLD response.

\if{html}{\out{<div class="sourceCode">}}\preformatted{If \code{"auto"} (default), will use mean scaling except if demeaned data
is detected, in which case sd scaling will be used instead.

\code{"mean"} scaling will scale the data to percent local signal change.

\code{"sd"} scaling will scale the data by local standard deviation.

\code{"none"} will only center the data, not scale it.
}\if{html}{\out{</div>}}}

\item{scale_design}{(logical) Should the design matrix be scaled? (Default is TRUE)}

\item{ar_order}{Order of the AR used to prewhiten the data at each location}

\item{ar_smooth}{FWHM parameter for smoothing. Remember that
\eqn{\sigma = \frac{FWHM}{2*sqrt(2*log(2)}}. Set to \code{0} or \code{NULL}
to not do any smoothing. Default: \code{5}.}

\item{cifti_data}{A \code{xifti} object used to map the AR coefficient
estimates onto the surface mesh for smoothing.}

\item{hemisphere}{'left' or 'right'}

\item{num.threads}{(scalar) The number of threads to use in parallelizing the
prewhitening}
}
\value{
The prewhitened data (in a list), the smoothed, averaged AR
coefficient estimates used in the prewhitening, the smoothed, average
residual variance after prewhitening, and the value given for \code{ar_order}.
}
\description{
Prewhiten cifti session data
}
