% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prewhitening.R
\name{prewhiten_prep}
\alias{prewhiten_prep}
\title{}
\usage{
prewhiten_prep(
  data,
  mask = NULL,
  scale_BOLD = TRUE,
  scale_design = TRUE,
  ar_order = 6
)
}
\arguments{
\item{data}{}

\item{mask}{}

\item{scale_BOLD}{}

\item{scale_design}{}

\item{ar_order}{}
}
\value{

}
\description{

}
