% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prewhitening.R
\name{prewhiten_prep}
\alias{prewhiten_prep}
\title{Estimate residual autocorrelation for prewhitening}
\usage{
prewhiten_prep(
  data,
  mask = NULL,
  scale_BOLD = c("auto", "mean", "sd", "none"),
  scale_design = TRUE,
  ar_order = 6
)
}
\arguments{
\item{data}{List of sessions (see \code{is.session} and \code{is.session_pw})}

\item{mask}{(Optional) A length \eqn{V} logical vector indicating if each
vertex is to be included.}

\item{scale_BOLD}{Option for scaling the BOLD response.

\if{html}{\out{<div class="sourceCode">}}\preformatted{If \code{"auto"} (default), will use mean scaling except if demeaned data
is detected, in which case sd scaling will be used instead.

\code{"mean"} scaling will scale the data to percent local signal change.

\code{"sd"} scaling will scale the data by local standard deviation.

\code{"none"} will only center the data, not scale it.
}\if{html}{\out{</div>}}}

\item{scale_design}{(logical) Should the design matrix be scaled? (Default is TRUE)}

\item{ar_order}{Order of the AR used to prewhiten the data at each location}
}
\value{
The prewhitened data (in a list), the smoothed, averaged AR
coefficient estimates used in the prewhitening, the smoothed, average
residual variance after prewhitening, and the value given for \code{ar_order}.
}
\description{
Estimate residual autocorrelation for prewhitening
}
