% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_utils.R
\name{Q_prime}
\alias{Q_prime}
\title{Gives the portion of the Q matrix independent of phi}
\usage{
Q_prime(kappa2, spde)
}
\arguments{
\item{kappa2}{scalar}

\item{spdean}{spde object}
}
\value{
a dgCMatrix
}
\description{
Gives the portion of the Q matrix independent of phi
}
\keyword{internal}
