% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_mask.R
\name{retro_mask_act}
\alias{retro_mask_act}
\title{Retroactively mask activations}
\usage{
retro_mask_act(x, Masks)
}
\arguments{
\item{x}{The activations object}

\item{Masks}{The masks to be applied to each brain structure of \code{x}.}
}
\value{
The masked result
}
\description{
Retroactively mask activations
}
\keyword{internal}
