% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_mask.R
\name{retro_mask_act}
\alias{retro_mask_act}
\title{Retroactively mask activations}
\usage{
retro_mask_act(x)
}
\arguments{
\item{x}{The activation list}
}
\value{
The masked result
}
\description{
Retroactively mask activations
}
\keyword{internal}
