% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{return_INLA_Param}
\alias{return_INLA_Param}
\title{return_INLA}
\arguments{
\item{return_INLA}{Return the INLA model object? (It can be large.)
Default: \code{TRUE}. Required for running \code{id_activations}
after, but not for running BayesGLM2 after to get
posterior quantities of group means or contrasts.}
}
\description{
return_INLA
}
