% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{return_INLA_Param}
\alias{return_INLA_Param}
\title{return_INLA}
\arguments{
\item{return_INLA}{Return the INLA model object? (It can be large.) Use
\code{"trimmed"} (default) returns the results sufficient for
\code{\link{id_activations}} and \code{\link{BayesGLM2}}; \code{"minimal"}
returns enough for \code{\link{BayesGLM2}} but not
\code{\link{id_activations}}; \code{"full"} returns the full \code{inla}
output.}
}
\description{
return_INLA
}
