% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{scale_BOLD_Param}
\alias{scale_BOLD_Param}
\title{scale_BOLD}
\arguments{
\item{scale_BOLD}{Controls scaling the BOLD response at each location.
\describe{
\item{"mean":}{   Scale the data to percent local signal change.}
\item{"sd":}{   Scale the data by local standard deviation.}
\item{"none":}{   Center the data but do not scale it.}
}}
}
\description{
scale_BOLD
}
