% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{scale_BOLD_Param}
\alias{scale_BOLD_Param}
\title{scale_BOLD}
\arguments{
\item{scale_BOLD}{Option for scaling the BOLD response.

\code{"auto"} (default) will use \code{"mean"} scaling except if demeaned
data is detected (if any mean is less than one), in which case \code{"sd"}
scaling will be used instead.

\code{"mean"} scaling will scale the data to percent local signal change.

\code{"sd"} scaling will scale the data by local standard deviation.

\code{"none"} will only center the data, not scale it.}
}
\description{
scale_BOLD
}
