% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_design_mat.R
\name{scale_design_mat}
\alias{scale_design_mat}
\title{}
\usage{
scale_design_mat(design_mat)
}
\arguments{
\item{design_mat}{}
}
\value{

}
\description{

}
\examples{
library(neuRosim)
# Task 1
t1 <-
  specifydesign(
    onsets = seq(0, 200, by = 40),
    durations = 1,
    totaltime = 200,
    TR = 1,
    effectsize = 1.3,
    conv = "double-gamma",
    param = list(list(a1 = 6, a2 = 12, b1 = 0.9, b2 = 0.9, c = 0.15))
  )
# Task 2
t2 <-
  specifydesign(
    onsets = seq(20, 200, by = 40),
    durations = 1,
    totaltime = 200,
    TR = 1,
    effectsize = 1.3,
    conv = "double-gamma",
    param = list(list(a1 = 6, a2 = 12, b1 = 0.9, b2 = 0.9, c = 0.15))
  )
A <- cbind(t1,t2) # This is the design matrix
B <- scale_design_mat(A)
}
