% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{scale_design_mat}
\alias{scale_design_mat}
\title{Scale the design matrix}
\usage{
scale_design_mat(design_mat)
}
\arguments{
\item{design_mat}{The original (unscaled) design matrix that is T x K, where
T is the number of time points, and k is the number of field covariates}
}
\value{
A scaled design matrix
}
\description{
Scale the design matrix
}
\keyword{internal}
