% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_timeseries.R
\name{scale_timeseries}
\alias{scale_timeseries}
\title{Scale the BOLD timeseries}
\usage{
scale_timeseries(
  BOLD,
  scale = c("auto", "mean", "sd", "none"),
  transpose = TRUE
)
}
\arguments{
\item{BOLD}{\eqn{V \times T} matrix of fMRI data. Rows are
time points; columns are data locations (vertices/voxels).}

\item{scale}{Option for scaling units.

\if{html}{\out{<div class="sourceCode">}}\preformatted{If \code{"auto"} (default), will use mean scaling except if demeaned data
is detected, in which case sd scaling will be used instead.

\code{"mean"} scaling will scale the data to percent local signal change.

\code{"sd"} scaling will scale the data by local standard deviation.

\code{"none"} will only center the data, not scale it.
}\if{html}{\out{</div>}}}

\item{transpose}{Check orientation of data? If \code{TRUE}, will transpose
the data when the number of time points is greater than the number of voxels.}
}
\value{
Scale to units of percent local signal change and centers
}
\description{
Scale the fMRI data, with several options for the scaling units.
}
\examples{
nT <- 30
nV <- 400
X <- matrix(rnorm(nV*nT), nrow=nV)
scale_timeseries(X)

}
