% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_timeseries.R
\name{scale_timeseries}
\alias{scale_timeseries}
\title{Scale the BOLD timeseries}
\usage{
scale_timeseries(BOLD, scale = TRUE, transpose = TRUE)
}
\arguments{
\item{BOLD}{Input fMRI data (V x T)}

\item{scale}{Option for scaling units to percent local signal change AND
centering. If false, only centers.}

\item{transpose}{Check orientation of data, which, if TRUE, will transpose
the data when the number of time points is greater than the number of voxels.
Note: this is not always true for subcortical regions.}
}
\value{
Scale to units of percent local signal change and centers
}
\description{
Scale the BOLD timeseries
}
