% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_timeseries.R
\name{scale_timeseries}
\alias{scale_timeseries}
\title{Scale the BOLD timeseries}
\usage{
scale_timeseries(BOLD)
}
\arguments{
\item{BOLD}{Input fMRI data}
}
\value{
Scale to units of percent local signal change and centers
}
\description{
Scale the BOLD timeseries
}
\examples{
\dontrun{}
}
