% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{session_names_Param}
\alias{session_names_Param}
\title{session_names}
\arguments{
\item{session_names}{The names of the task-fMRI \code{BOLD} sessions, for
multi-session analysis. If not provided here, will be inferred from
\code{names(BOLD)}, inferred from \code{names(design)}, or generated
automatically, in that order.}
}
\description{
session_names
}
