% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_slice_data.R
\name{simulate_slice_data}
\alias{simulate_slice_data}
\title{Create simulated slice data for the BayesGLM_surface function}
\usage{
simulate_slice_data(
  num_sessions = 1,
  num_tasks = 2,
  active_centers = NULL,
  active_size = NULL,
  beta_weights = NULL,
  vary_active = T,
  num_time = 200,
  binary_template = NULL
)
}
\arguments{
\item{num_sessions}{a number}

\item{num_tasks}{a number}

\item{active_centers}{a matrix, with the number of rows corresponding to the
number of activation regions. There should be two columns, corresponding
to the x- and y-coordinates for the centers.}

\item{active_size}{a vector of length \code{nrow(active_centers)}}

\item{beta_weights}{a matrix, with the number of rows equal to
\code{num_tasks} and the number of columns equal to the number of rows in
\code{active_centers}}

\item{vary_active}{logical, if there are multiple sessions, should the
activation centers vary between sessions?}

\item{num_time}{length of the time series that is generated}

\item{binary_template}{(optional) a binary brain slice image}
}
\value{
A list in which the first element, \code{data}, is a list of
sessions, each with components \code{BOLD} (a matrix) and \code{design}
(a matrix). The second list element is named \code{betas}, which is a list
of sessions, each containing a list of the true values of the coefficients
used to generate the data to be analyzed.
}
\description{
Create simulated slice data for the BayesGLM_surface function
}
\examples{
data <- simulate_slice_data()
}
