% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPDE_from_voxel.R
\name{SPDE_from_voxel}
\alias{SPDE_from_voxel}
\title{SPDE from voxel model}
\usage{
SPDE_from_voxel(spatial)
}
\arguments{
\item{spatial}{See \code{BayesGLM}.}
}
\value{
List
}
\description{
SPDE from voxel model
}
\keyword{internal}
