% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mesh.R
\name{submesh}
\alias{submesh}
\title{Remove part of a mesh without using INLA functions}
\usage{
submesh(mask, mesh)
}
\arguments{
\item{mask}{a 0-1 vector}

\item{mesh}{a mesh resulting from a call to \code{make_mesh} with
\code{use_INLA = FALSE}}
}
\value{
a mesh object with fewer vertices than the original input mesh
}
\description{
Remove part of a mesh without using INLA functions
}
\keyword{internal}
