% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGLM_utils.R
\name{summary.BayesGLM}
\alias{summary.BayesGLM}
\alias{print.summary.BayesGLM}
\alias{print.BayesGLM}
\title{Summarize a \code{"BayesGLM"} object}
\usage{
\method{summary}{BayesGLM}(object, ...)

\method{print}{summary.BayesGLM}(x, ...)

\method{print}{BayesGLM}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"BayesGLM"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"summary.BayesGLM"}.}
}
\value{
A \code{"summary.BayesGLM"} object, a list summarizing the properties
of \code{object}.

\code{NULL}, invisibly.

\code{NULL}, invisibly.
}
\description{
Summary method for class \code{"BayesGLM"}
}
