% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tile.plot}
\alias{tile.plot}
\title{Lightweight tile plot function}
\usage{
tile.plot(tile_df, col = NULL, ncols = NULL, main = "", na.color = "grey80")
}
\arguments{
\item{tile_df}{A data frame with three columns: \code{row}, \code{col}, and
\code{value} describing locations within a matrix. See
\code{\link{melt_mat}} for an example.}

\item{col}{A color palette}

\item{ncols}{The number of colors for a color palette, if \code{col} is not
provided.}

\item{main}{Plot title (character)}

\item{na.color}{The color that should be used to represent \code{NA} values
in the tile plot}
}
\value{
A tile plot done in base R graphics
}
\description{
This function has no dependencies outside of base R, but provides a
reasonable approximation to the functionality of
\code{ggplot2::geom_raster}.
}
\examples{
x <- matrix(rnorm(50*50),50,50)
x_df <- melt_mat(x)
tile.plot(x_df)
}
