% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_diagnosticICA.R
\name{UpdateTheta.diagnosticICA}
\alias{UpdateTheta.diagnosticICA}
\title{Parameter Estimates in EM Algorithm for Diagnostic ICA Model}
\usage{
UpdateTheta.diagnosticICA(template_mean, template_var, BOLD, theta, C_diag)
}
\arguments{
\item{template_mean}{(A list of G matrices, each LxV) template mean estimates for each group 1 to G}

\item{template_var}{(A list of G matrices, each LxV) template variance estimates for each group 1 to G}

\item{BOLD}{(LxV matrix) dimension-reduced fMRI data}

\item{C_diag}{(Lx1) diagonal elements of residual covariance after dimension reduction}

\item{theta0}{(list) current parameter estimates}

\item{maxiter}{maximum number of EM iterations}

\item{epsilon}{smallest proportion change between iterations (e.g. .001)}

\item{verbose}{If TRUE, display progress of algorithm}

\item{return_MAP}{If TRUE, returns the posterior mean and variance of the latent fields and group membership instead of the parameter estimates}
}
\value{
An updated list of parameter estimates, theta, OR if return_MAP=TRUE, the posterior mean and precision of the latent fields
}
\description{
Parameter Estimates in EM Algorithm for Diagnostic ICA Model
}
