% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_diagnosticICA.R
\name{UpdateTheta.diagnosticICA}
\alias{UpdateTheta.diagnosticICA}
\title{Parameter Estimates in EM Algorithm for Diagnostic ICA Model}
\usage{
UpdateTheta.diagnosticICA(
  template_mean,
  template_var,
  BOLD,
  theta,
  C_diag,
  return_MAP = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{template_mean}{(A list of G matrices, each VxL) template mean estimates for each group 1 to G}

\item{template_var}{(A list of G matrices, each VxL) template variance estimates for each group 1 to G}

\item{BOLD}{(VxL matrix) dimension-reduced fMRI data}

\item{theta}{(list) current parameter estimates}

\item{C_diag}{(Lx1) diagonal elements of residual covariance after dimension reduction}

\item{return_MAP}{If TRUE, returns the posterior mean and variance of the latent fields and group membership instead of the parameter estimates}

\item{verbose}{If TRUE, display progress of algorithm}
}
\value{
An updated list of parameter estimates, theta, OR if return_MAP=TRUE, the posterior mean and precision of the latent fields
}
\description{
Parameter Estimates in EM Algorithm for Diagnostic ICA Model
}
