% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{verbose_Param}
\alias{verbose_Param}
\title{verbose}
\arguments{
\item{verbose}{Should updates be printed? Use \code{1} (default) for
occasional updates, \code{2} for occasional updates as well as running INLA
in verbose mode (if applicable), or \code{0} for no updates.}
}
\description{
verbose
}
